ClimaLand

LSM Model Types and methods

Missing docstring.

Missing docstring for ClimaLand.SoilPlantHydrologyModel. Check Documenter's build log for details.

ClimaLand.LandSoilBiogeochemistry — Type

struct LandSoilBiogeochemistry{
    FT,
    SEH <: Soil.EnergyHydrology{FT},
    SB <: Soil.Biogeochemistry.SoilCO2Model{FT},
} <: AbstractLandModel{FT}

A concrete type of land model used for simulating systems with a soil energy, hydrology, and biogeochemistry component.

soil: The soil model

soilco2: The biochemistry model

ClimaLand.LandHydrology — Type

struct LandHydrology{
    FT,
    SM <: Soil.AbstractSoilModel{FT},
    SW <: Pond.AbstractSurfaceWaterModel{FT},
} <: AbstractLandModel{FT}

A concrete type of land model used for simulating systems with a soil and surface water component.

soil: The soil model

surface_water: The surface water model

Missing docstring.

Missing docstring for ClimaLand.make_interactions_update_aux. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaLand.initialize_interactions. Check Documenter's build log for details.

ClimaLand.land_components — Function

land_components(land::AbstractLandModel)

Returns the component names of the land model, by calling propertynames(land).

ClimaLand.lsm_aux_vars — Function

lsmauxvars(m::AbstractLandModel)

Returns the additional aux variable symbols for the model in the form of a tuple.

lsm_aux_vars(m::SoilCanopyModel)

The names of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

These include the broadband albedo of the land surface α_sfc, defined as the ratio of SWu/SWd, and T_sfc, defined as the temperature a blackbody with emissivity ϵ_sfc would have in order to emit the same LWu as the land surface does. This is called the [effective temperature](https://en.wikipedia.org/wiki/Effectivetemperature) in some fields, and is not the same as the skin temperature (defined e.g. Equation 7.13 of Bonan, 2019, Climate Change and Terrestrial Ecosystem Modeling. DOI: 10.1017/9781107339217).

lsm_aux_vars(m::LandHydrologyModel)

The names of the additional auxiliary variables that are included in the integrated Soil-Snow model.

ClimaLand.lsm_aux_types — Function

lsmauxtypes(m::AbstractLandModel)

Returns the shared additional aux variable types for the model in the form of a tuple.

lsm_aux_types(m::SoilCanopyModel)

The types of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

lsm_aux_types(m::LandHydrologyModel)

The types of the additional auxiliary variables that are included in the integrated Soil-Snow model.

ClimaLand.lsm_aux_domain_names — Function

lsmauxdomain_names(m::AbstractLandModel)

Returns the additional domain symbols in the form of a tuple e.g. :surface or :subsurface.

This is only required for variables shared between land submodels, and only needed for multi-component models, not standalone components. Component-specific variables should be listed as prognostic or auxiliary variables which do not require this to initialize.

lsm_aux_domain_names(m::SoilCanopyModel)

The domain names of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

lsm_aux_domain_names(m::LandHydrologyModel)

The domain names of the additional auxiliary variables that are included in the integrated Soil-Snow model.

Missing docstring.

Missing docstring for ClimaLand.domain_name. Check Documenter's build log for details.

Land Hydrology

ClimaLand.infiltration_capacity — Function

function infiltration_capacity(
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
)

Function which computes the infiltration capacity of the soil based on soil characteristics, moisture levels, and pond height.

Defined such that positive means into soil.

ClimaLand.infiltration_at_point — Function

infiltration_at_point(η::FT, i_c::FT, P::FT)

Returns the infiltration given pond height η, infiltration capacity, and precipitation.

This is defined such that positive means into soil.

ClimaLand.PrognosticRunoff — Type

PrognosticRunoff <: Pond.AbstractSurfaceRunoff

Concrete type of Pond.AbstractSurfaceRunoff for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically.

This is paired with Soil.RunoffBC: both are used at the same time, ensuring the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

ClimaLand.RunoffBC — Type

RunoffBC <: Soil.AbstractSoilBC

Concrete type of Soil.AbstractSoilBC for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically. This infiltration is then used to set an upper boundary condition for the soil.

This is paired with Pond.PrognosticRunoff: both are used at the same time, ensuring that the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

SoilCanopyModel

Missing docstring.

Missing docstring for ClimaLand.PrognosticSoil. Check Documenter's build log for details.

ClimaLand.RootExtraction — Type

RootExtraction{FT} <: Soil.AbstractSoilSource{FT}

Concrete type of Soil.AbstractSoilSource, used for dispatch in an LSM with both soil and plant hydraulic components.

This is paired with the source term Canopy.PrognosticSoil:both are used at the same time, ensuring that the water flux into the roots is extracted correctly from the soil.

LandSoilBiogeochemistry

Missing docstring.

Missing docstring for ClimaLand.PrognosticMet. Check Documenter's build log for details.