Snow Model

Snow Model and Parameters

ClimaLand.Snow.SnowModel — Type

struct SnowModel{
    FT,
    PS <: SnowParameters{FT},
    BC,
    D,
} <: AbstractSnowModel{FT}

A container/type for the bulk snow model, based on the UEB snow model of Tarboton et al. (1995) and Tarboton and Luce (1996).

ClimaLand.Snow.SnowModel — Method

SnowModel(;
    parameters::SnowParameters{FT, DM, PSE},
    domain::ClimaLand.Domains.AbstractDomain,
    boundary_conditions::BC
) where {FT, DM, PSE, BC}

Construct a SnowModel with parameters, domain, and boundary_conditions.

ClimaLand.Snow.SnowModel — Method

SnowModel(
    FT,
    domain,
    forcing,
    toml_dict::CP.ParamDict,
    Δt;
    prognostic_land_components = (:snow,),
    z_0m = toml_dict["snow_momentum_roughness_length"],
    z_0b = toml_dict["snow_scalar_roughness_length"],
    ϵ_snow = toml_dict["snow_emissivity"],
    α_snow = ConstantAlbedoModel(toml_dict["snow_albedo"]),
    density = MinimumDensityModel(toml_dict["snow_density"]),
    scf = WuWuSnowCoverFractionModel(toml_dict, FT(1)),
    θ_r = toml_dict["holding_capacity_of_water_in_snow"],
    Ksat = toml_dict["wet_snow_hydraulic_conductivity"],
    ΔS = toml_dict["delta_S"],
)

Creates a SnowModel model with the given float type FT, domain, toml_dict, forcing, and prognostic land components.

When running the snow model in standalone mode, provide prognostic_land_components = (:snow,), while for running integrated land models, this should be a list of the component models. This value of this argument must be the same across all components in the integrated land model.

Default parameterizations and parameters can be overwritten using keyword arguments.

ClimaLand.Snow.SnowParameters — Type

SnowParameters{FT <: AbstractFloat, PSE}

A struct for storing parameters of the SnowModel.

Note that in our current implementation of runoff, a physical timescale is required and computed using Ksat and the depth of the snow. For shallow snowpacks, this will fall below the timestep of the model. For that reason, we pass the timestep of the model as a parameter, and take the larger of the timestep and the physical timescale as the value used in the model. Future implementations will revisit this.

density: Choice of parameterization for snow density

z_0m: Roughness length over snow for momentum (m)

z_0b: Roughness length over snow for scalars (m)

α_snow: Albedo parameterization for snow

ϵ_snow: Emissivity of snow (unitless)

θ_r: Volumetric holding capacity of water in snow (unitless)

Ksat: Hydraulic conductivity of wet snow (m/s)

κ_ice: Thermal conductivity of ice (W/m/K)

Δt: Timestep of the model (s)

ΔS: Parameter to prevent dividing by zero when computing snow temperature (m)

scf: Snow cover fraction parameterization

earth_param_set: Clima-wide parameters

ClimaLand.Snow.SnowParameters — Method

function SnowParameters(
    toml_dict::CP.ParamDict,
    Δt;
    ρ_snow = toml_dict["snow_density"],
    α_snow_param = toml_dict["snow_albedo"],
    density::DM = MinimumDensityModel(ρ_snow),
    α_snow::AM = ConstantAlbedoModel(α_snow_param),
    scf ::SCFM = WuWuSnowCoverFractionModel(
        toml_dict,
        CP.float_type(toml_dict)(1.0),
    ),
    z_0m = toml_dict["snow_momentum_roughness_length"],
    z_0b = toml_dict["snow_scalar_roughness_length"],
    κ_ice = toml_dict["thermal_conductivity_of_water_ice"],
    ϵ_snow = toml_dict["snow_emissivity"],
    θ_r = toml_dict["holding_capacity_of_water_in_snow"],
    Ksat = toml_dict["wet_snow_hydraulic_conductivity"],
    ΔS = toml_dict["delta_S"],
) where {DM, AM, SCFM}

TOML dictionary constructor for the SnowParameters` struct.

Δt = 450.0
# TOML Dictionary:
import ClimaParams as CP
toml_dict = CP.create_toml_dict(Float32);
ClimaLand.Canopy.SnowParameters(toml_dict, Δt; ϵ_snow = Float32(0.99), Ksat = Float32(1e-4))

Snow Functions of State

ClimaLand.Snow.specific_heat_capacity — Function

specific_heat_capacity(q_l::FT,
                       parameters::SnowParameters{FT}
                       ) where {FT}

Computes the specific heat capacity of the snow, neglecting any contribution from air in the pore spaces, given the liquid water mass fraction q_l and other parameters.

ClimaLand.Snow.snow_surface_temperature — Function

snow_surface_temperature(T::FT) where {FT}

Returns the snow surface temperature assuming it is the same as the bulk temperature T.

ClimaLand.Snow.snow_thermal_conductivity — Function

snow_thermal_conductivity(ρ_snow::FT,
                     parameters::SnowParameters{FT},
                     ) where {FT}

Computes the thermal conductivity, given the density of the snow, according to Equation 5.33 from Bonan's textbook, which in turn is taken from Jordan (1991).

We have adjusted the original equation to make the coefficients non-dimensional by multiplying by the first by x = ρ_ice/ρ_ice and the second by x², with ρ_ice in kg/m³.

When ρ_snow = ρ_ice, we recover κ_snow = κ_ice.

ClimaLand.Snow.snow_bulk_temperature — Function

snow_bulk_temperature(U::FT,
                      S::FT,
                      q_l::FT,
                      parameters::SnowParameters{FT}) where {FT}

Computes the bulk snow temperature from the snow water equivalent S, energy per unit area U, liquid water fraction ql, and specific heat capacity cs, along with other needed parameters.

If there is no snow (U = S = q_l = 0), the bulk temperature is the reference temperature, which is 273.16K.

ClimaLand.Snow.liquid_mass_fraction — Function

liquid_mass_fraction(S::FT, S_l::FT)::FT where {FT}

Diagnoses the liquid mass fraction from the prognostic variables Sl and S, while preventing from division by zero and enforcing that ql = S_l/S must lie between 0 and 1.

ClimaLand.Snow.maximum_liquid_mass_fraction — Function

maximum_liquid_mass_fraction(ρ_snow::FT, T::FT, parameters::SnowParameters{FT}) where {FT}

Computes the maximum liquid water mass fraction, given the density of the snow ρ_snow and other parameters.

ClimaLand.Snow.runoff_timescale — Function

runoff_timescale(z::FT, Ksat::FT, Δt::FT) where {FT}

Computes the timescale for liquid water to percolate and leave the snowpack, given the depth of the snowpack z and the hydraulic conductivity Ksat.

ClimaLand.Snow.compute_water_runoff — Function

compute_energy_runoff(S::FT, S_l::FT, T::FT, parameters) where {FT}

Computes the rate of change in the snow water equivalent S due to loss of liquid water (runoff) from the snowpack.

Runoff occurs as the snow melts and exceeds the water holding capacity.

ClimaLand.Snow.energy_from_q_l_and_swe — Function

energy_from_q_l_and_swe(S::FT, q_l::FT, parameters) where {FT}

A helper function for compute the snow energy per unit area, given snow water equivalent S, liquid fraction q_l, and snow model parameters.

This assumes that the snow is at the freezing point.

ClimaLand.Snow.energy_from_T_and_swe — Function

energy_from_T_and_swe(S::FT, T::FT, parameters) where {FT}

A helper function for compute the snow energy per unit area, given snow water equivalent S, bulk temperature T, and snow model parameters.

The liquid mass fraction is assumed to be zero if T<=T_freeze, and 1 otherwise.

ClimaLand.Snow.energy_flux_falling_rain — Function

energy_flux_falling_rain(atmos, p, parameters)

Returns the energy flux of falling rain for a PrescribedAtmosphere, approximated as ρl el(Tatmos) * Pliq. The energy is per unit volume of liquid water, and P_liq is expressed as the volume flux of liquid water resulting from the rain.

This method can be extended to coupled simulations, where atmos is of type CoupledAtmosphere, and the energy flux of the falling rain is passed in the cache p. In that case, this should specify atmos::PrescribedAtmosphere.

ClimaLand.Snow.energy_flux_falling_snow — Function

energy_flux_falling_snow(atmos, p, parameters)

Returns the energy flux of falling snow for a PrescribedAtmosphere, approximated as ρesnow * Psnow, where ρesnow = -LHf0 * ρliq. This is a negative internal energy, due the to negative contribution of the latent heat of melting to the energy of the snow, and it neglects the sensible heat portion of the snow. The energy is per unit volume of liquid water, and P_snow is expressed as the volume flux of liquid water resulting from the snow.

This method can be extended to coupled simulations, where atmos is of type CoupledAtmosphere, and the energy flux of the falling snow is passed in the cache p. In that case, this should specify atmos::PrescribedAtmosphere.

ClimaLand.Snow.update_snow_cover_fraction! — Function

update_snow_cover_fraction!(x::FT; z0 = FT(1e-1), β_scf = FT(2))::FT where {FT}

Returns the snow cover fraction from snow depth z, from Wu, Tongwen, and Guoxiong Wu. "An empirical formula to compute snow cover fraction in GCMs." Advances in Atmospheric Sciences 21 (2004): 529-535.

ClimaLand.Snow.update_snow_albedo! — Function

update_snow_albedo!(α, m::ConstantAlbedoModel, Y, p, t, earth_param_set)

Updates the snow albedo α in place with the current albedo, according to the ConstantAlbedoModel.

update_snow_albedo!(α, m::ZenithAngleAlbedoModel, Y, p, t, earth_param_set)

Updates the snow albedo α in place with the current albedo, according to the ZenithAngleAlbedoModel.

Computing fluxes for snow

ClimaLand.Snow.snow_boundary_fluxes! — Function

snow_boundary_fluxes!(bc::AtmosDrivenSnowBC, model::SnowModel, Y, p, t)

Updates in place various volumetric water flux (m/s) and energy flux (W/m^2) terms for the snow model:

p.snow.turbulent fluxes (latent, sensible, and evaporative fluxes)

p.snow.R_n (radiative fluxes)

p.snow.total_water_flux

p.snow.total_energy_flux

The two latter fluxes also include contributions from fluxes due to melt and precipitation, but note that precipitation and melt flux are not computed or updated in snow_boundary_fluxes currently. Instead, they are updated in update_aux!, which happens prior to the snow_boundary_fluxes! call, and used in the snow_boundary_fluxes! call.

This function calls the turbulent_fluxes! and net_radiation! functions, which use the snow surface conditions as well as the atmos and radiation conditions in order to compute the surface fluxes using Monin Obukhov Surface Theory. It also accounts for the presence of other components, if run as part of an integrated land model, and their effect on boundary conditions.

snow_boundary_fluxes!(
    bc::AtmosDrivenSnowBC,
    prognostic_land_components::Val{(:snow,)},
    model::SnowModel{FT},
    Y,
    p,
    t,
) where {FT}

Computes the boundary fluxes for the snow model in standalone mode.

The ground heat flux is assumed to be zero, and the snow surface is assumed to be bare (no vegetation).

snow_boundary_fluxes!(
    bc::AtmosDrivenSnowBC,
    prognostic_land_components::Val{(:snow, :soil)},
    model::SnowModel{FT},
    Y,
    p,
    t,
) where {FT}

A method of snow_boundary_fluxes! which computes the boundary fluxes for the snow model accounting for a heat flux between the soil and snow.

The snow surface is assumed to be bare (no vegetation).

Currently this is almost identical to the method for snow alone, except for the inclusion of the ground heat flux (precomputed by the integrated land model). However, this will change more if e.g. we allow for transmission of radiation through the snowpack.

ClimaLand.Snow.phase_change_flux — Function

 phase_change_flux(U::FT, S::FT, q_l::FT, energy_flux::FT, parameters) where {FT}

Computes the volume flux of liquid water undergoing phase change, given the applied energy flux and current state of U,S,q_l.

ClimaLand.Snow.AtmosDrivenSnowBC — Type

AtmosDrivenSnowBC{
    A <: AbstractAtmosphericDrivers,
    B <: AbstractRadiativeDrivers,
    C::Tuple
} <: AbstractSnowBC

A struct used to specify the snow fluxes, referred to as "boundary conditions", at the surface and bottom of the snowpack, for water and energy.

These fluxes include turbulent surface fluxes computed with Monin-Obukhov theory, and radiative fluxes.

atmos: The atmospheric conditions driving the model

radiation: The radiative fluxes driving the model

prognostic_land_components: Prognostic land components present

Snow parameterizations

ClimaLand.Snow.WuWuSnowCoverFractionModel — Type

WuWuSnowCoverFractionModel{FT <: AbstractFloat} <: AbstractSnowCoverFractionModel{FT}

Establishes the snow cover parameterization of Wu, Tongwen, and Guoxiong Wu. "An empirical formula to compute snow cover fraction in GCMs." Advances in Atmospheric Sciences 21 (2004): 529-535, scf = min(β_scf * z̃ / (z̃ + 1), 1),

where z̃ = snow depth per ground area / 0.106 m, and β_scf is computed using a resolution dependent formula: β_scf = max(β0 - γ(horz_degree_res - 1.5), β_min), where horz_degree_res is the horizontal resolution of the simulation, in degrees, and β0, β_min and γ are unitless. It is correct to think of β0, β_min, γ, and z0 as the free parameters, while horz_degree_res is provided and β_scf is determined.

β0, β_min, γ, and β_scf must be > 0.

From Wu and Wu et al, β0 ∼ 1.77 and γ ∼ 0.08, over a range of 1.5-4.5∘

ClimaLand.Snow.ZenithAngleAlbedoModel — Type

ZenithAngleAlbedoModel{FT <: AbstractFloat} <: AbstractAlbedoModel{FT}

Establishes the albedo parameterization where albedo depends on the cosine of the zenith angle of the sun, as

$\alpha = f(x) [\alpha_0 + \Delta\alpha \cdot \text{exp}(-k\cos(\theta s))]$

where cos θs is the cosine of the zenith angle, α_0, Δα, and k are free parameters. The factor out front is a function of x = ρ_snow/ρ_liq, of the form f(x) = min(1 - β(x-x0), 1). The parameters x0 ∈ [0,1] and β ∈ [0,1] are free. Choose β = 0 to remove this dependence on snow density.

Note: If this choice is used, the field cosθs must appear in the cache p.drivers. This is available through the PrescribedRadiativeFluxes object.

ClimaLand.Snow.ConstantAlbedoModel — Type

ConstantAlbedoModel{FT <: AbstractFloat} <: AbstractAlbedoModel{FT}

Establishes the albedo parameterization where albedo is treated as a constant spatially and temporally.