Vegetation

Models

ClimaLSM.Roots.AbstractVegetationModel — Type

AbstractVegetationModel{FT} <: AbstractModel{FT}

An abstract type for vegetation models.

Concrete types include a plant hydraulics model, but future types will include multi-layer canopy models and possibly a big leaf model.

ClimaLSM.Roots.RootsModel — Type

RootsModel{FT, PS, D, RE, T, B} <: AbstractVegetationModel{FT}

Defines, and constructs instances of, the RootsModel type, which is used for simulation flow of water to/from soil, along roots of different depths, along a stem, to a leaf, and ultimately being lost from the system by transpiration.

This model can be used in standalone mode by prescribing the transpiration rate and soil pressure at the root tips, or with a dynamic soil model using ClimaLSM.

param_set

Parameters required by the root model

domain

The root model domain, of type AbstractVegetationDomain

root_extraction

The root extraction model, of type AbstractRootExtraction

transpiration

The transpiration model, of type AbstractTranspiration

Roots Diagnostic Variables

ClimaLSM.Roots.p_to_theta — Function

p_to_theta(p::FT) where {FT}

Computes the pressure (p) given the volumetric water content (theta). Currently this is using appropriate vG parameters for loamy type soil. The pressure (MPa) must be converted to meters (head) for use in the van Genuchten formula.

ClimaLSM.Roots.theta_to_p — Function

theta_to_p(theta::FT) where {FT}

Computes the volumetric water content (moles/moles) given pressure (p). Currently this is using appropriate vG parameters for loamy type soil. First the head (m) is computed, and then converted to a pressure in MPa.

ClimaLSM.Roots.flow — Function

function flow(
    z1::FT,
    z2::FT,
    p1::FT,
    p2::FT,
    a::FT,
    b::FT,
    Kmax::FT,
) where {FT}

Computes the flow of water (moles/sec) given the height and pressures at two points. Here, a, b, andKmax` are parameters which parameterize the hydraulic conductance of the pathway along which the flow occurs.

ClimaLSM.Roots.flow_out_roots — Function

Roots.flow_out_roots(
    re::PrognosticSoilPressure{FT},
    model::Roots.RootsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

An extension of the Roots.flow_out_roots function, which returns the net flow of water between the roots and the soil, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flow of water between roots and soil at each soil layer.

flow_out_roots(
    re::PrescribedSoilPressure{FT},
    model::RootsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

A method which computes the flow between the soil and the stem, via the roots, in the case of a standalone root model with prescribed soil pressure (in MPa) at the root tips.

This assumes that the stem compartment is the first element of Y.roots.rwc.

Roots Parameters

ClimaLSM.Roots.RootsParameters — Type

RootsParameters{FT <: AbstractFloat}

A struct for holding parameters of the Root Model. Eventually to be used with ClimaParameters.

a_root

controls the shape and steepness of conductance vs. pressure curve, for roots: unitless

b_root

controls the steepness of the relative conductance vs. pressure curve, for roots: inverse MPa

a_stem

controls the shape and steepness of relative conductance vs. pressure curve, for stems: unitless

b_stem

controls the steepness of the conductance vs. pressure curve, for stems: inverse MPa

size_reservoir_stem_moles

the physical size of the stem, in moles

size_reservoir_leaf_moles

the physical size of the leaves, in moles

K_max_root_moles

water conductance in roots (moles/s/MPa) when pressure is zero, a maximum

K_max_stem_moles

water conductance in stems (moles/s/MPa) when pressure is zero, a maximum

Roots Methods and Types

Missing docstring.

Missing docstring for ClimaLSM.Roots.flow_out_roots. Check Documenter's build log for details.

ClimaLSM.Roots.PrescribedSoilPressure — Type

PrescribedSoilPressure{FT} <: AbstractRootExtraction{FT}

A concrete type used for dispatch when computing the flow_out_roots, in the case where the soil pressure at each root layer is prescribed.

ClimaLSM.Roots.PrescribedTranspiration — Type

PrescribedTranspiration{FT} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is prescribed.

ClimaLSM.Roots.AbstractRootExtraction — Type

AbstractRootExtraction{FT <: AbstractFloat}

An abstract type for types representing different models of water exchange between soil and plants.

Currently, only a prescribed soil pressure is supported for standalone plant hydraulics. Use within an LSM requires types defined within ClimaLSM, and include a prognostic soil pressure for models with both soil and roots.