Autotrophic Respiration

Models and Parameters

ClimaLand.Canopy.AutotrophicRespirationModel — Type

AutotrophicRespirationModel{FT, ARP <: AutotrophicRespirationParameters{FT},} <: AbstractAutotrophicRespirationModel{FT}

The JULES autotrophic respiration model.

Clark, D. B., et al. "The Joint UK Land Environment Simulator (JULES), model description–Part 2: carbon fluxes and vegetation dynamics." Geoscientific Model Development 4.3 (2011): 701-722.

ClimaLand.Canopy.AutotrophicRespirationModel — Method

AutotrophicRespirationModel{FT}(toml_dict::CP.ParamDict,
                               ) where {FT <: AbstractFloat}

Create a AutotrophicRespirationModel using toml_dict of type FT.

ClimaLand.Canopy.AutotrophicRespirationParameters — Type

AutotrophicRespirationParameters{FT<:AbstractFloat}

The required parameters for the autrophic respiration model, which is based off of the JULES model. Clark, D. B., et al. "The Joint UK Land Environment Simulator (JULES), model description–Part 2: carbon fluxes and vegetation dynamics." Geoscientific Model Development 4.3 (2011): 701-722.

ne: Vcmax25 (leaf level) to N factor (mol CO2 m-2 s-1 kg C (kg C)-1)

ηsl: Live stem wood coefficient (kg C m-3)

σl: Specific leaf density (kg C m-2 [leaf])

μr: Ratio root nitrogen to top leaf nitrogen (-), typical value 1.0

μs: Ratio stem nitrogen to top leaf nitrogen (-), typical value 0.1

Rel: Relative contribution or Rgrowth (-)

ClimaLand.Canopy.AutotrophicRespirationParameters — Method

AutotrophicRespirationParameters(toml_dict::CP.ParamDict; kwargs...)

Constructor for the AutotrophicRespirationParameters struct by passing a TOML dictionary. You can manually override any parameter via keyword arguments:

AutotrophicRespirationParameters(toml_dict; ne = 99999)

Methods

ClimaLand.Canopy.nitrogen_content — Function

nitrogen_content(
                 ne::FT, # Mean leaf nitrogen concentration (kg N (kg C)-1)
                 Vcmax25::FT, #
                 LAI::FT, # Leaf area index
                 SAI::FT,
                 RAI::FT,
                 ηsl::FT, # live stem  wood coefficient (kg C m-3)
                 h::FT, # canopy height (m)
                 σl::FT # Specific leaf density (kg C m-2 [leaf])
                 μr::FT, # Ratio root nitrogen to top leaf nitrogen (-), typical value 1.0
                 μs::FT, # Ratio stem nitrogen to top leaf nitrogen (-), typical value 0.1
                ) where {FT}

Computes the nitrogen content of leafs (Nl), roots (Nr) and stems (Ns).

ClimaLand.Canopy.plant_respiration_maintenance — Function

plant_respiration_maintenance(
    Rd::FT, # Dark respiration
    β::FT, # Soil moisture factor
    Nl::FT, # Nitrogen content of leafs
    Nr::FT, # Nitrogen content of roots
    Ns::FT, # Nitrogen content of stems
    ) where {FT}

Computes plant maintenance respiration as a function of dark respiration (Rd), the nitrogen content of leafs (Nl), roots (Nr) and stems (Ns), and the soil moisture factor (β).

ClimaLand.Canopy.plant_respiration_growth — Function

plant_respiration_growth(
    Rel::FT, # Factor of relative contribution
    An::FT, # Net photosynthesis
    Rpm::FT # Plant maintenance respiration
    ) where {FT}

Computes plant growth respiration as a function of net photosynthesis (An), plant maintenance respiration (Rpm), and a relative contribution factor, Rel.