Solar Induced Fluorescence

Models and Parameters

ClimaLand.Canopy.Lee2015SIFModel — Type

Lee2015SIFModel{FT, SP <: SIFParameters{FT}} <: AbstractSIFModel{FT}

Lee2015SIFModel struct containing the parameters needed to compute the SIF parameterization described in: Lee et al, 2015. Global Change Biology 21, 3469-3477, doi:10.1111/gcb.12948.

ClimaLand.Canopy.SIFParameters — Type

SIFParameters{FT<:AbstractFloat}

The required parameters for the SIF parameterisation Lee et al, 2015. Global Change Biology 21, 3469-3477, doi:10.1111/gcb.12948.

kf: The rate coefficient for florescence, unitless

kd_p1: Parameter used to compute the rate coefficient for heat loss in dark-adapted conditions, Tol et al. 2014, unitless

kd_p2: Parameter used to compute the rate coefficient for heat loss in dark-adapted conditions, Tol et al. 2014, unitless

min_kd: Parameter used to compute the rate coefficient for heat loss in dark-adapted conditions, Tol et al. 2014, unitless

kn_p1: Parameter used to compute the rate coefficient for heat loss in light-adapted conditions, Lee et al 2013 (unitless)

kn_p2: Parameter used to compute the rate coefficient for heat loss in light-adapted conditions, Lee et al 2013 (unitless)

kp: Rate coefficient for photochemical quenching

kappa_p1: Slope of line relating leaf-level fluorescence to spectrometer-observed fluorescence as a function of Vcmax 25 (leaf level). Lee et al 2015.

kappa_p2: Intercept of line relating leaf-level fluorescence to spectrometer-observed fluorescence as a function of Vcmax 25 (leaf level). Lee et al 2015.

Methods

ClimaLand.Canopy.update_SIF! — Function

update_SIF!(p, Y, sif_model::Lee2015SIFModel, canopy)

Updates observed SIF at 755 nm in W/m^2. Note that Tc is in Kelvin, and photo synthetic rates are in mol/m^2/s, and APAR is in PPFD. Lee et al, 2015. Global Change Biology 21, 3469-3477, doi:10.1111/gcb.12948

ClimaLand.Canopy.compute_SIF_at_a_point — Function

compute_SIF_at_a_point(
    APAR_canopy_moles::FT,
    Tc::FT,
    Vcmax25_leaf::FT,
    J_over_Jmax::FT,
    T_freeze::FT,
    sif_parameters::SIFParameters{FT},
) where {FT}

Computes the Solar Induced Fluorescence (SIF) at 755 nm in W/m^2 using the Lee et al. 2015 model. This takes as parameters APAR (absorbed photosynthetically active radiation, mol/m^2/s), Tc (canopy temperature, K), Vcmax25 (maximum carboxylation rate at 25 °C, mol/m^2/s), Jmax (electron transport rate, mol/m^2/s), J (electron transport rate, mol/m^2/s), T_freeze (freezing temperature, K), sif_parameters (SIF parameters).