Soil moisture stress

Models and Parameters

ClimaLand.Canopy.TuzetMoistureStressModel — Type

TuzetMoistureStressModel{FT} <: AbstractSoilMoistureStressModel{FT}

An implementation of the Tuzet moisture stress function.

βm = min(1, (1 + exp(sc * pc)) / (1 + exp(sc * (pc - p_leaf)))), where sc and pc are stored in the parameter struct.

TUZET, A., PERRIER, A. and LEUNING, R. (2003), A coupled model of stomatal conductance, photosynthesis and transpiration. Plant, Cell & Environment, 26: 1097-1116. https://doi.org/10.1046/j.1365-3040.2003.01035.x

Duursma, R. A., and Medlyn, B. E. (2012). MAESPA: A model to study interactions between water limitation, environmental drivers and vegetation function at tree and stand levels, with an example application to [CO2] ! drought interactions. Geoscientific Model Development, 5, 919–940.

ClimaLand.Canopy.TuzetMoistureStressModel — Method

TuzetMoistureStressModel{FT}(toml_dict; sc::FT = toml_dict["moisture_stress_sc"], pc::FT = toml_dict["moisture_stress_pc"]) where{FT}

A constructor for TuzetMoistureStressModel which uses the toml_dict values, allowing optional overrides by keyword argument.

ClimaLand.Canopy.PiecewiseMoistureStressModel — Type

PiecewiseMoistureStressModel{FT,  F <: Union{FT, ClimaCore.Fields.Field}}
<: AbstractSoilMoistureStressModel{FT}

An implementation of a piecewise moisture stress model, taking the form

βm(z) = min(1, [(θ(z)-θlow)/(θhigh - θlow)]^c), where θhigh, θ_low, and c are parameters, and θ(z) is the soil moisture at z.

The parameters should fall within the following ranges:

θhigh>θlow should be in (θ_r, ν) where ν is the porosity of the soil

c should be positive

See Egea et al. (2011) for details. Citation: https://doi.org/10.1016/j.agrformet.2011.05.019

ClimaLand.Canopy.PiecewiseMoistureStressModel — Method

PiecewiseMoistureStressModel{FT}(
domain,
toml_dict::CP.ParamDict;
c::FT = toml_dict["moisture_stress_c"],
soil_params

) where {FT <: AbstractFloat}

Helper function to create PiecewiseMoistureStressModel by calculating field capacity (θhigh) and wilting point (θlow) from the soil parameters.

The low and high thresholds for the piecewise soil moisture stress function are given by the residual soil water content and the soil porosity, respectively.

The soil parameters should be a named tuple with keys of ν and θ_r (additional keys may be present but they will not be used). These may be ClimaCore fields or floats, but must be consistently one or the other.

Note that unlike other Canopy components, this model is defined on the subsurface domain.

ClimaLand.Canopy.NoMoistureStressModel — Type

struct NoMoistureStressModel{FT} <: AbstractSoilMoistureStressModel{FT} end

A constructor for NoMoistureStressModel, which sets βm = 1.0 always.

Methods

ClimaLand.Canopy.update_soil_moisture_stress! — Function

update_soil_moisture_stress!(
    p,
    Y,
    model::TuzetMoistureStressModel,
    canopy,
)

Computes and updates in place the soil moisture stress for the Tuzet formulation.

update_soil_moisture_stress!(
    p,
    Y,
    model::PiecewiseMoistureStressModel,
    canopy,
)

This updates the soil moisture stress factor according to the piecewise soil moisture stress model.

ClimaLand.Canopy.update_piecewise_soil_moisture_stress! — Function

update_piecewise_soil_moisture_stress!(ground::PrognosticGroundConditions, p, Y, model, canopy)

Updates the soil moisture stress using the piecewise model for a prognostic soil model, where θ is vertically resolved and prognostic.

update_piecewise_soil_moisture_stress!(ground::PrescribedGroundConditions, p, Y, model, canopy)

Updates the soil moisture stress using the piecewise model for Prescribed GroundConditions (p.drivers.θ prescribed).