ClimaLand Abstract Models and Functions

ClimaLand.AbstractModel — Type

abstract type AbstractModel{FT <: AbstractFloat}

An abstract type for all models.

ClimaLand.AbstractExpModel — Type

AbstractExpModel{FT} <: AbstractModel{FT}

An abstract type for models with only explicitly stepped variables. This inherits all the default function definitions from AbstractModel, as well as a make_imp_tendency default.

ClimaLand.AbstractImExModel — Type

AbstractImExModel{FT} <: AbstractModel{FT}

An abstract type for models which have both implicit and explicitly stepped variables This inherits all the default function definitions from AbstractModel, as well as make_imp_tendency and make_compute_imp_tendency defaults.

ClimaLand.name — Function

name(model::AbstractModel)

Returns a symbol of the model component name, e.g. :soil or :vegetation.

ClimaLand.initialize — Function

initialize(model::AbstractModel)

Creates the prognostic and auxiliary states structures, but with unset values; constructs and returns the coordinates for the model domain. We may need to consider this default more as we add diverse components and Simulations.

ClimaLand.get_earth_param_set — Function

get_earth_param_set(model::AbstractModel)

A helper function which returns the earthparamset of the model.