Bucket

Types

ClimaLand.Bucket.BucketModelParameters — Type

struct BucketModelParameters{
    FT <: AbstractFloat,
    PSE,
}

Container for holding the parameters of the bucket model.

κ_soil: Conductivity of the soil (W/K/m); constant

ρc_soil: Volumetric heat capacity of the soil (J/m^3/K); constant

albedo: Albedo Model

σS_c: Critical σSWE amount (m) where surface transitions from to snow-covered

f_snow: Fraction of critical amount of snow at which sublimation β begins to decay to zero (unitless)

W_f: Capacity of the land bucket (m)

f_bucket: Fraction of bucket capacity at which evaporation β begins to decay to zero (unitless)

p: Exponent used in β decay (unitless)

z_0m: Roughness length for momentum (m)

z_0b: Roughness length for scalars (m)

τc: τc timescale on which snow melts

earth_param_set: Earth Parameter set; physical constants, etc

ClimaLand.Bucket.PrescribedBaregroundAlbedo — Type

PrescribedBaregroundAlbedo{FT, F <: ClimaCore.Fields.Field} <: AbstractBucketAlbedoModel

An albedo model where the static snow-free bareground albedo is prescribed as a function of space or using data from a file, and the land surface albedo is computed each timestep as a linear combination of the snow albedo and the bareground albedo, following the SLIM model (Lague et al 2019).

ClimaLand.Bucket.PrescribedSurfaceAlbedo — Type

PrescribedSurfaceAlbedo{FT, TV <: AbstractTimeVaryingInput}
                   <: AbstractBucketAlbedoModel

An albedo model where the albedo of different surface types is specified. Albedo is specified via a NetCDF file which is a function of time and covers all surface types (soil, vegetation, snow, etc). This albedo type changes over time according to the input file.

Note that this option should only be used with global simulations, i.e. with a ClimaLand.LSMSphericalShellDomain.

ClimaLand.Bucket.BucketModel — Type

struct BucketModel{
     FT,
     PS <: BucketModelParameters{FT},
     ATM <: AbstractAtmosphericDrivers{FT},
     RAD <: AbstractRadiativeDrivers{FT},
     D,
 } <: AbstractBucketModel{FT}

Concrete type for the BucketModel, which store the model domain and parameters, as well as the necessary atmosphere and radiation fields for driving the model.

parameters: Parameters required by the bucket model

atmos: The atmospheric drivers: Prescribed or Coupled

radiation: The radiation drivers: Prescribed or Coupled

domain: The domain of the model

Misc Functions

ClimaLand.surface_albedo — Function

ClimaLand.surface_albedo(
    model::EnergyHydrology{FT},
    Y,
    p,
) where {FT}

Returns the surface albedo field of the EnergyHydrology soil model.

surface_albedo(model::SnowModel, Y, p)

A helper function which computes and returns the snow albedo.

surface_albedo(model::BucketModel, Y, p)

Returns the bulk surface albedo, which gets updated in update_aux via next_albedo.

surface_albedo(model::AbstractModel, Y, p)

A helper function which returns the surface albedo for a given model, needed because different models compute and store α_sfc in different ways and places.

Extending this function for your model is only necessary if you need to compute surface fluxes and radiative fluxes at the surface using the functions in this file.

ClimaLand.Bucket.beta_factor — Function

beta_factor(W::FT, σS::FT, fW_f::FT, fσS_c::FT, p::FT) where {FT}

Computes the beta factor which scales the evaporation/sublimation from the potential rate. The beta factor is given by:

β = (x/xc)^p x < xc 1 otherwise

where x = W and xc = fbucket * Wf for the bucket, and x = σS and xc = fsnow *σSc for snow.

Artifact Path Functions

ClimaLand.Bucket.cesm2_albedo_dataset_path — Function

cesm2_albedo_dataset_path()

Triggers the download of the CESM2 land albedo dataset, if not already downloaded, using Julia Artifacts, and returns the path to this file.

This dataset contains monthly albedo data from 15/01/1850 to 15/12/2014.