Bucket

Types

ClimaLSM.Bucket.BucketModelParameters — Type

struct BucketModelParameters{
    FT <: AbstractFloat,
    PSE,
}

Container for holding the parameters of the bucket model.

κ_soil

: Conductivity of the soil (W/K/m); constant

ρc_soil

: Volumetric heat capacity of the soil (J/m^3/K); constant

albedo

: Albedo Model

σS_c

: Critical σSWE amount (m) where surface transitions from soil to snow

W_f

: Capacity of the land bucket (m)

z_0m

: Roughness length for momentum (m)

z_0b

: Roughness length for scalars (m)

τc

: τc timescale on which snow melts

earth_param_set

: Earth Parameter set; physical constants, etc

ClimaLSM.Bucket.PrescribedAtmosphere — Type

PrescribedAtmosphere{FT, LP, SP, TA, UA, QA, RA} <: AbstractAtmosphericDrivers{FT}

Container for holding prescribed atmospheric drivers and other information needed for computing turbulent surface fluxes when driving the bucket model in standalone mode.

liquid_precip

: Precipitation (m/s) function of time: positive by definition

snow_precip

: Snow precipitation (m/s) function of time: positive by definition

T_atmos

: Prescribed atmospheric temperature (function of time) at the reference height (K)

u_atmos

: Prescribed wind speed (function of time) at the reference height (m/s)

q_atmos

: Prescribed specific humidity (function of time) at the reference height (_)

ρ_atmos

: Prescribed air density (function of time) at the reference height (kg/m^3)

h_atmos

: Reference height, relative to surface elevation(m)

ρ_sfc

: Surface air density (kg/m^3). Note that convergence errors result if ρsfc = ρatmos.

ClimaLSM.Bucket.PrescribedRadiativeFluxes — Type

PrescribedRadiativeFluxes{FT, SW, LW} <: AbstractRadiativeDrivers{FT}

Container for the prescribed radiation functions needed to drive the bucket model in standalone mode.

SW_d

: Downward shortwave radiation function of time (W/m^2): positive indicates towards surface

LW_d

: Downward longwave radiation function of time (W/m^2): positive indicates towards surface

ClimaLSM.Bucket.AbstractAtmosphericDrivers — Type

 AbstractAtmosphericDrivers{FT <: AbstractFloat}

An abstract type of atmospheric drivers of the bucket model.

ClimaLSM.Bucket.AbstractRadiativeDrivers — Type

 AbstractRadiativeDrivers{FT <: AbstractFloat}

An abstract type of radiative drivers of the bucket model.

ClimaLSM.Bucket.BulkAlbedo — Type

BulkAlbedo{FT} <: AbstractLandAlbedoModel

An albedo model where the albedo of different surface types is specified. Snow albedo is treated as constant across snow location and across wavelength. Soil albedo is specified as a function of latitude and longitude, but is also treated as constant across wavelength.

ClimaLSM.Bucket.BucketModel — Type

struct BucketModel{
     FT,
     PS <: BucketModelParameters{FT},
     ATM <: AbstractAtmosphericDrivers{FT},
     RAD <: AbstractRadiativeDrivers{FT},
     D,
 } <: AbstractBucketModel{FT}

Concrete type for the BucketModel, which store the model domain and parameters, as well as the necessary atmosphere and radiation fields for driving the model.

parameters

: Parameters required by the bucket model

atmos

: The atmospheric drivers: Prescribed or Coupled

radiation

: The radiation drivers: Prescribed or Coupled

domain

: The domain of the model

Extendible Methods

ClimaLSM.Bucket.surface_fluxes — Function

surface_fluxes(Y,p,
                t::FT,
                parameters::P,
                atmos::PA,
                ) where {FT <: AbstractFloat, P <: BucketModelParameters{FT},  PA <: PrescribedAtmosphere{FT}}

Computes the turbulent surface flux terms at the ground for a standalone simulation: turbulent energy fluxes as well as the water vapor flux (in units of m^3/m^2/s of water). Positive fluxes indicate flow from the ground to the atmosphere.

It solves for these given atmospheric conditions, stored in atmos, model parameters, and the surface temperature and surface specific humidity.

ClimaLSM.Bucket.net_radiation — Function

net_radiation(
    Y,
    p,
    t::FT,
    parameters::P,
    radiation::PR,
    ) where {FT <: AbstractFloat, P <: BucketModelParameters{FT},  PA <: PrescribedAtmosphere{FT}}

Computes the net radiation at the ground for a standalone simulation. Positive fluxes indicate flow from the ground to the atmosphere.

ClimaLSM.Bucket.surface_air_density — Function

surface_air_density(p, atmos::PrescribedAtmosphere)

Returns the air density (kg/m^3) at the surface.

ClimaLSM.Bucket.liquid_precipitation — Function

liquid_precipitation(p, atmos::PrescribedAtmosphere, t)

Returns the liquid precipitation (m/s) at the surface.

ClimaLSM.Bucket.snow_precipitation — Function

snow_precipitation(p, atmos::PrescribedAtmosphere, t)

Returns the precipitation in snow (m of liquid water/s) at the surface.

ClimaLSM.Bucket.surface_albedo — Function

surface_albedo(albedo::BulkAlbedo{FT}, coords, S::FT, S_c::FT)::FT where {FT <: AbstractFloat}

Returns the bulk surface albedo, linearly interpolating between the albedo of snow and of soil, based on the snow water equivalent S relative to the parameter S_c.

The linear interpolation is taken from Lague et al 2019.

ClimaLSM.Bucket.partition_surface_fluxes — Function

partition_surface_fluxes(
    σS::FT,
    T_sfc::FT,
    τ::FT,
    snow_cover_fraction::FT,
    E::FT,
    F_sfc::FT,
    _LH_f0::FT,
    _T_freeze::FT,
) where{FT}

Partitions the surface fluxes in a flux for melting snow, a flux for sublimating snow, and a ground heat flux.

All fluxes are positive if they are in the direction from land towards the atmosphere.