Bucket

Types

ClimaLSM.Bucket.BucketModelParameters — Type

struct BucketModelParameters{
    FT <: AbstractFloat,
    PSE,
}

Container for holding the parameters of the bucket model.

κ_soil: Conductivity of the soil (W/K/m); constant

ρc_soil: Volumetric heat capacity of the soil (J/m^3/K); constant

albedo: Albedo Model

σS_c: Critical σSWE amount (m) where surface transitions from to snow-covered

W_f: Capacity of the land bucket (m)

z_0m: Roughness length for momentum (m)

z_0b: Roughness length for scalars (m)

τc: τc timescale on which snow melts

earth_param_set: Earth Parameter set; physical constants, etc

ClimaLSM.Bucket.BulkAlbedoMap — Type

BulkAlbedoMap{FT} <: AbstractLandAlbedoModel

An albedo model where the albedo of different surface types is specified. Snow albedo is treated as constant across snow location and across wavelength. Surface albedo is specified via a NetCDF file, which can be a function of time, but is treated as constant across wavelengths; surface is this context refers to soil and vegetation.

Note that this option should only be used with global simulations, i.e. with a ClimaLSM.LSMSphericalShellDomain.

ClimaLSM.Bucket.BulkAlbedoFunction — Type

BulkAlbedoFunction{FT} <: AbstractLandAlbedoModel

An albedo model where the albedo of different surface types is specified. Snow albedo is treated as constant across snow location and across wavelength. Surface albedo (sfc) is specified as a function of latitude and longitude, but is also treated as constant across wavelength; surface is this context refers to soil and vegetation.

ClimaLSM.Bucket.BucketModel — Type

struct BucketModel{
     FT,
     PS <: BucketModelParameters{FT},
     ATM <: AbstractAtmosphericDrivers{FT},
     RAD <: AbstractRadiativeDrivers{FT},
     D,
 } <: AbstractBucketModel{FT}

Concrete type for the BucketModel, which store the model domain and parameters, as well as the necessary atmosphere and radiation fields for driving the model.

parameters: Parameters required by the bucket model

atmos: The atmospheric drivers: Prescribed or Coupled

radiation: The radiation drivers: Prescribed or Coupled

domain: The domain of the model

Misc Functions

ClimaLSM.surface_albedo — Function

ClimaLSM.surface_albedo(
    model::EnergyHydrology{FT},
    Y,
    p,
) where {FT}

Returns the surface albedo field of the EnergyHydrology soil model.

surface_albedo(model::AbstractModel, Y, p)

A helper function which returns the surface albedo for a given model, needed because different models compute and store α_sfc in different ways and places.

Extending this function for your model is only necessary if you need to compute surface fluxes and radiative fluxes at the surface using the functions in this file.

surface_albedo(model::BucketModel, Y, p)

A helper function which computes and returns the bulk surface albedo, linearly interpolating between the albedo of snow and of the surface, based on the snow water equivalent S relative to the parameter S_c.

The linear interpolation is taken from Lague et al 2019.

ClimaLSM.Bucket.beta_factor — Function

beta_factor(W::FT, σS::FT, W_f::FT) where {FT}

Computes the beta factor which scales the evaporation from the potential rate.

Methods for handling parameters read in from NetCDF files

ClimaLSM.Bucket.set_initial_parameter_field! — Function

function set_initial_parameter_field!(
    albedo::BulkAlbedoFunction{FT},
    p,
    surface_coords,
) where {FT}

Updates the spatially-varying but constant in time surface albedo stored in the auxiliary vector p in place, according to the passed function of latitute and longitude stored in albedo.α_sfc.

function set_initial_parameter_field!(
    albedo::BulkAlbedoMap{FT},
    p,
    surface_coords,
) where {FT}

Updates spatially-varying surface albedo stored in the auxiliary vector p in place, according to a NetCDF file. This data file is encapsulated in an object of type ClimaLSM.Regridder.MapInfo in the field albedo.α_sfc.

The NetCDF file is read in, regridded, and projected onto the surface space of the LSM using ClimaCoreTempestRemap. The result is a ClimaCore.Fields.Field of albedo values.

Missing docstring.

Missing docstring for ClimaLSM.Bucket.update_soil_albedo. Check Documenter's build log for details.

Artifact Path Functions

ClimaLSM.Bucket.cesm2_land_albedo_dataset_path — Function

cesm2_land_albedo_dataset_path()

Triggers the download of the CESM2 land albedo dataset, if not already downloaded, using Julia Artifacts, and returns the path to this file.