ClimaLSM

LSM Model Types and methods

ClimaLSM.RootSoilModel — Type

struct RootSoilModel{
    FT,
    SM <: Soil.AbstractSoilModel{FT},
    VM <: Roots.AbstractVegetationModel{FT},
} <: AbstractLandModel{FT}
    soil::SM
    vegetation::VM
end

A concrete type of land model used for simulating systems with a vegetation and a soil component.

soil

The soil model to be used

vegetation

The vegetation model to be used

ClimaLSM.LandHydrology — Type

struct LandHydrology{
    FT,
    SM <: Soil.AbstractSoilModel{FT},
    SW <: Pond.AbstractSurfaceWaterModel{FT},
} <: AbstractLandModel{FT}

A concrete type of land model used for simulating systems with a soil and surface water component.

soil

The soil model

surface_water

The surface water model

ClimaLSM.make_interactions_update_aux — Function

make_interactions_update_aux(land::AbstractLandModel) end

Makes and returns a function which updates the interaction variables, which are a type of auxiliary variable.

The update_aux! function returned is evaluted during the right hand side evaluation.

This is a stub which concrete types of LSMs extend.

make_interactions_update_aux(
    land::RootSoilModel{FT, SM, RM},
) where {FT, SM <: Soil.RichardsModel{FT}, RM <: Roots.RootsModel{FT}}

A method which makes a function; the returned function updates the auxiliary variable p.root_extraction, which is needed for both the boundary condition for the soil model and the source term (runoff) for the surface water model.

This function is called each ode function evaluation.

function make_interactions_update_aux(
    land::LandHydrology{FT, SM, SW},
) where {FT, SM <: Soil.RichardsModel{FT}, SW <: Pond.PondModel{FT}}

A method which makes a function; the returned function updates the auxiliary variable p.soil_infiltration, which is needed for both the boundary condition for the soil model and the source term (runoff) for the surface water model.

This function is called each ode function evaluation.

ClimaLSM.initialize_interactions — Function

initialize_interactions(land::AbstractLandModel) end

Initializes interaction variables, which are a type of auxiliary variable, to empty objects of the correct type for the model.

This function should be called during initialize_auxiliary, and the method will change depending on the type of land model, and potentially due to the type of the component models.

This is a stub which is extended to concrete LSM models.

initialize_interactions(
    land::RootSoilModel{FT, SM, RM},
) where {FT, SM <: Soil.RichardsModel{FT}, RM <: Roots.RootsModel{FT}}

Initializes the interaction auxiliary variable p.root_extraction.

function initialize_interactions(
    land::LandHydrology{FT, SM, SW},
) where {FT, SM <: Soil.RichardsModel{FT}, SW <: Pond.PondModel{FT}}

Initializes the interaction auxiliary variable p.soil_infiltration.

ClimaLSM.land_components — Function

land_components(land::AbstractLandModel)

Returns the component names of the land model, by calling propertynames(land).

Land Hydrology

ClimaLSM.infiltration_capacity — Function

function infiltration_capacity(
    model::Soil.AbstractSoilModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
) where {FT}

Function which computes the infiltration capacity of the soil based on soil characteristics, moisture levels, and pond height.

Defined such that positive means into soil.

ClimaLSM.infiltration_at_point — Function

infiltration_at_point(η::FT, i_c::FT, P::FT)

Returns the infiltration given pond height η, infiltration capacity, and precipitation.

This is defined such that positive means into soil.

ClimaLSM.PrognosticRunoff — Type

PrognosticRunoff{FT} <: Pond.AbstractSurfaceRunoff{FT}

Concrete type of Pond.AbstractSurfaceRunoff for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically.

This is paired with Soil.RunoffBC: both are used at the same time, ensuring the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

ClimaLSM.RunoffBC — Type

RunoffBC{FT} <: Soil.AbstractSoilBoundaryConditions{FT}

Concrete type of Soil.AbstractSoilBoundaryConditions for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically. This infiltration is then used to set an upper boundary condition for the soil.

This is paired with Pond.PrognosticRunoff: both are used at the same time, ensuring that the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

RootSoilModel

ClimaLSM.PrognosticSoilPressure — Type

PrognosticSoilPressure{FT} <: Roots.AbstractRootExtraction{FT}

Concrete type of Roots.AbstractRootExtraction, used for dispatch in an LSM with both soil and plant hydraulic components.

This is paired with the source term Soil.RootExtraction:both are used at the same time, ensuring that the water flow into the roots is extracted correctly from the soil.

ClimaLSM.RootExtraction — Type

RootExtraction{FT} <: Soil.AbstractSoilSource{FT}

Concrete type of Soil.AbstractSoilSource, used for dispatch in an LSM with both soil and plant hydraulic components.

This is paired with the source term Roots.PrognosticSoilPressure:both are used at the same time, ensuring that the water flow into the roots is extracted correctly from the soil.