ClimaLSM

LSM Model Types and methods

Missing docstring.

Missing docstring for ClimaLSM.SoilPlantHydrologyModel. Check Documenter's build log for details.

ClimaLSM.LandSoilBiogeochemistry — Type

struct LandSoilBiogeochemistry{
    FT,
    SEH <: Soil.EnergyHydrology{FT},
    SB <: Soil.Biogeochemistry.SoilCO2Model{FT},
} <: AbstractLandModel{FT}

A concrete type of land model used for simulating systems with a soil energy, hydrology, and biogeochemistry component.

soil: The soil model

soilco2: The biochemistry model

ClimaLSM.LandHydrology — Type

struct LandHydrology{
    FT,
    SM <: Soil.AbstractSoilModel{FT},
    SW <: Pond.AbstractSurfaceWaterModel{FT},
} <: AbstractLandModel{FT}

A concrete type of land model used for simulating systems with a soil and surface water component.

soil: The soil model

surface_water: The surface water model

ClimaLSM.make_interactions_update_aux — Function

make_interactions_update_aux(land::AbstractLandModel) end

Makes and returns a function which updates the interaction variables, which are a type of auxiliary variable.

The update_aux! function returned is evaluted during the right hand side evaluation.

This is a stub which concrete types of LSMs extend.

function make_interactions_update_aux(
    land::LandHydrology{FT, SM, SW},
) where {FT, SM <: Soil.RichardsModel{FT}, SW <: Pond.PondModel{FT}}

A method which makes a function; the returned function updates the auxiliary variable p.soil_infiltration, which is needed for both the boundary condition for the soil model and the source term (runoff) for the surface water model.

This function is called each ode function evaluation.

make_interactions_update_aux(
    land::SoilCanopyModel{FT, MM, SM, RM},
) where {
    FT,
    MM <: Soil.Biogeochemistry.SoilCO2Model{FT},
    SM <: Soil.RichardsModel{FT},
    RM <: Canopy.CanopyModel{FT}
    }

A method which makes a function; the returned function updates the auxiliary variable p.root_extraction, which is needed for a sink term for the soil model and to create the lower water boundary condition for the canopy model. It also updates the soil surface fluxes, which are affected by the presence of a canopy.

This function is called each ode function evaluation.

ClimaLSM.initialize_interactions — Function

initialize_interactions(land::AbstractLandModel) end

Initializes interaction variables, which are a type of auxiliary variable, to empty objects of the correct type for the model.

Interaction variables are specified by interaction_vars, their types by interaction_types, and their spaces by interaction_spaces. This function should be called during initialize_auxiliary step.

ClimaLSM.land_components — Function

land_components(land::AbstractLandModel)

Returns the component names of the land model, by calling propertynames(land).

ClimaLSM.interaction_vars — Function

interaction_vars(m::AbstractLandModel)

Returns the interaction variable symbols for the model in the form of a tuple.

interaction_vars(m::SoilCanopyModel)

The names of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

ClimaLSM.interaction_types — Function

interaction_types(m::AbstractLandModel)

Returns the shared interaction variable types for the model in the form of a tuple.

interaction_types(m::SoilCanopyModel)

The types of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

ClimaLSM.interaction_domain_names — Function

interactiondomainnames(m::AbstractLandModel)

Returns the interaction domain symbols in the form of a tuple e.g. :surface or :subsurface.

This is only required for variables shared between land submodels, and only needed for multi-component models, not standalone components. Component-specific variables should be listed as prognostic or auxiliary variables which do not require this to initialize.

interaction_domain_names(m::SoilCanopyModel)

The domain names of the additional auxiliary variables that are included in the integrated Soil-Canopy model.

Missing docstring.

Missing docstring for ClimaLSM.domain_name. Check Documenter's build log for details.

Land Hydrology

ClimaLSM.infiltration_capacity — Function

function infiltration_capacity(
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
)

Function which computes the infiltration capacity of the soil based on soil characteristics, moisture levels, and pond height.

Defined such that positive means into soil.

ClimaLSM.infiltration_at_point — Function

infiltration_at_point(η::FT, i_c::FT, P::FT)

Returns the infiltration given pond height η, infiltration capacity, and precipitation.

This is defined such that positive means into soil.

ClimaLSM.PrognosticRunoff — Type

PrognosticRunoff{FT} <: Pond.AbstractSurfaceRunoff{FT}

Concrete type of Pond.AbstractSurfaceRunoff for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically.

This is paired with Soil.RunoffBC: both are used at the same time, ensuring the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

ClimaLSM.RunoffBC — Type

RunoffBC <: Soil.AbstractSoilBC

Concrete type of Soil.AbstractSoilBC for use in LSM models, where precipitation is passed in, but infiltration is computed prognostically. This infiltration is then used to set an upper boundary condition for the soil.

This is paired with Pond.PrognosticRunoff: both are used at the same time, ensuring that the infiltration used for the boundary condition of soil is also used to compute the runoff for the surface water.

SoilCanopyModel

ClimaLSM.PrognosticSoil — Type

 PrognosticSoil{FT} <: AbstractSoilDriver{FT}

Concrete type of AbstractSoilDriver used for dispatch in cases where both a canopy model and soil model are run.

α_PAR: Soil albedo for PAR

α_NIR: Soil albedo for NIR

ClimaLSM.RootExtraction — Type

RootExtraction{FT} <: Soil.AbstractSoilSource{FT}

Concrete type of Soil.AbstractSoilSource, used for dispatch in an LSM with both soil and plant hydraulic components.

This is paired with the source term Canopy.PrognosticSoil:both are used at the same time, ensuring that the water flux into the roots is extracted correctly from the soil.

LandSoilBiogeochemistry

Missing docstring.

Missing docstring for ClimaLSM.PrognosticMet. Check Documenter's build log for details.