Shared Utilities

Domains

ClimaLSM.Domains.AbstractDomain — Type

AbstractDomain{FT <:AbstractFloat}

An abstract type for domains.

ClimaLSM.Domains.AbstractLSMDomain — Type

AbstractLSMDomain{FT} <: AbstractDomain{FT}

An abstract type for LSMDomains, which have two components: a surface and a subsurface.

ClimaLSM.Domains.PlantHydraulicsDomain — Type

PlantHydraulicsDomain{FT} <: AbstractVegetationDomain{FT}

Domain for a single bulk plant with roots of varying depths. The user needs to specify the depths of the root tips as wel as the heights of the compartments to be modeled within the plant. The compartment heights are expected to be sorted in ascending order. The number of stem and leaf compartments the plant should be divided into must be specified as well.

ClimaLSM.Domains.AbstractVegetationDomain — Type

AbstractVegetationDomain{FT} <: AbstractDomain{FT}

TBD if this needed longer term.

ClimaLSM.Domains.SphericalShell — Type

struct SphericalShell{FT} <: AbstractDomain{FT}
    radius::FT
    height::FT
    nelements::Tuple{Int, Int}
    npolynomial::Int
end

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space (non-radial directions) x a 1d finite difference space (radial direction), and the resulting coordinate field.

Fields

radius: The radius of the shell

height: The radial extent of the shell

nelements: The number of elements to be used in the non-radial and radial directions

npolynomial: The polynomial order to be used in the non-radial directions

space: The associated ClimaCore Space

ClimaLSM.Domains.SphericalSurface — Type

struct SphericalSurface{FT} <: AbstractDomain{FT}
    radius::FT
    nelements::Tuple{Int, Int}
    npolynomial::Int
end

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space (non-radial directions) and the resulting coordinate field.

Fields

radius: The radius of the surface

nelements: The number of elements to be used in the non-radial directions

npolynomial: The polynomial order to be used in the non-radial directions

space: The associated ClimaCore Space

ClimaLSM.Domains.HybridBox — Type

struct HybridBox{FT} <: AbstractDomain{FT}
    xlim::Tuple{FT, FT}
    ylim::Tuple{FT, FT}
    zlim::Tuple{FT, FT}
    nelements::Tuple{Int, Int, Int}
    npolynomial::Int
    periodic::Tuple{Bool, Bool}
end

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space (horizontal) x a 1d finite difference space (vertical), and the resulting coordinate field. This domain is not periodic along the z-axis. Note that only periodic domains are supported in the horizontal.

Fields

xlim: Domain interval limits along x axis, in meters

ylim: Domain interval limits along y axis, in meters

zlim: Domain interval limits along z axis, in meters

nelements: Number of elements to discretize interval, (nx, ny,nz)

npolynomial: Polynomial order for the horizontal directions

periodic: Flag indicating periodic boundaries in horizontal. only true is supported

space: The associated ClimaCore Space

ClimaLSM.Domains.Column — Type

Column{FT} <: AbstractDomain{FT}

A struct holding the necessary information to construct a domain, a mesh, a center and face space, etc. for use when a finite difference in 1D is suitable, as for a soil column model.

Fields

zlim: Domain interval limits, (zmin, zmax), in meters

nelements: Number of elements used to discretize the interval

boundary_tags: Boundary face identifiers

space: The associated ClimaCore Space

ClimaLSM.Domains.Plane — Type

Plane{FT} <: AbstractDomain{FT}

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space, and the resulting coordinate field. Note that only periodic domains are currently supported.

Fields

xlim: Domain interval limits along x axis, in meters

ylim: Domain interval limits along y axis, in meters

nelements: Number of elements to discretize interval, (nx, ny)

periodic: Flags for periodic boundaries; only true is supported

npolynomial: Polynomial order for both x and y

space: The associated ClimaCore Space

ClimaLSM.Domains.Point — Type

Point{FT} <: AbstractDomain{FT}

A domain for single column surface variables. For models such as ponds, snow, plant hydraulics, etc. Enables consistency in variable initialization across all domains.

Fields

z_sfc: Surface elevation relative to a reference (m)

space: The associated ClimaCore Space

ClimaLSM.Domains.LSMSingleColumnDomain — Type

LSMSingleColumnDomain{FT} <: AbstractLSMDomain{FT}

A mixed domain, consisting of a column domain with z-coordinates at the finite difference cell centers, and a point domain, with a single z coordinate at the top boundary of the column domain. For use in LSM modeling, where a subsurface finite difference space (for modeling soil hydrology and energy) and a surface space are both needed.

Fields

subsurface: The subsurface Column domain

surface: The surface Point domain

ClimaLSM.Domains.LSMMultiColumnDomain — Type

LSMMultiColumnDomain{FT} <: AbstractLSMDomain{FT}

A mixed domain, consisting of a hybdrid box domain with z-coordinates at the finite difference cell centers, and a plane domain, coinciding with the surface of the box.

For use in LSM modeling, where a subsurface finite difference space (for modeling soil hydrology and energy) and a surface space are both needed.

Fields

subsurface: The subsurface box domain

surface: The surface plane domain

ClimaLSM.Domains.LSMSphericalShellDomain — Type

LSMSphericalShellDomain{FT} <: AbstractLSMDomain{FT}

A mixed domain, consisting of a spherical shell domain with z-coordinates at the finite difference cell centers, and a spherical surface domain, coinciding with the surface of the shell.

For use in LSM modeling, where a subsurface finite difference space (for modeling soil hydrology and energy) and a surface space are both needed.

Fields

subsurface: The subsurface shell domain

surface: The surface domain

ClimaLSM.Domains.coordinates — Function

coordinates(domain::AbstractDomain)

Returns the coordinate field for the domain.

coordinates(domain::AbstractLSMDomain{FT}) where {FT}

Returns the coordinates of the AbstractLSMDomain as a named tuple, with keys of subsurface and surface.

ClimaLSM.Domains.obtain_face_space — Function

obtain_face_space(cs::ClimaCore.Spaces.AbstractSpace)

Returns the face space, if applicable, for the center space cs.

obtain_face_space(cs::ClimaCore.Spaces.CenterExtrudedFiniteDifferenceSpace)

Returns the face space for the CenterExtrudedFiniteDifferenceSpace cs.

obtain_face_space(cs::ClimaCore.Spaces.CenterFiniteDifferenceSpace)

Returns the face space corresponding to the CenterFiniteDifferenceSpace cs.

ClimaLSM.Domains.obtain_surface_space — Function

obtain_surface_space(cs::ClimaCore.Spaces.AbstractSpace)

Returns the surface space, if applicable, for the center space cs.

obtain_surface_space(cs::ClimaCore.Spaces.CenterExtrudedFiniteDifferenceSpace)

Returns the horizontal space for the CenterExtrudedFiniteDifferenceSpace cs.

obtain_surface_space(cs::ClimaCore.Spaces.CenterFiniteDifferenceSpace)

Returns the top level of the face space corresponding to the CenterFiniteDifferenceSpace cs.

Models

ClimaLSM.AbstractModel — Type

abstract type AbstractModel{FT <: AbstractFloat}

An abstract type for all models.

ClimaLSM.make_ode_function — Function

make_ode_function(model::AbstractModel)

Returns an ode_function that updates auxiliary variables and updates the prognostic state.

ode_function! should be compatible with OrdinaryDiffEq.jl solvers.

ClimaLSM.make_rhs — Function

make_rhs(model::RichardsModel)

An extension of the function make_rhs, for the Richardson- Richards equation.

This function creates and returns a function which computes the entire right hand side of the PDE for ϑ_l, and updates dY.soil.ϑ_l in place with that value.

This has been written so as to work with Differential Equations.jl.

make_rhs(model::EnergyHydrology)

An extension of the function make_rhs, for the integrated soil energy and heat equations, including phase change.

This function creates and returns a function which computes the entire right hand side of the PDE for Y.soil.ϑ_l, Y.soil.θ_i, Y.soil.ρe_int, and updates dY.soil in place with those values.

This has been written so as to work with Differential Equations.jl.

make_rhs(model::VegetationModel)

A function which creates the rhs! function for the PlantHydraulicsModel. The rhs! function must comply with a rhs function of OrdinaryDiffEq.jl.

make_rhs(model::AbstractModel)

Return a rhs! function that updates state variables.

rhs! should be compatible with OrdinaryDiffEq.jl solvers.

make_rhs(model::SoilCO2Model)

An extension of the function make_rhs, for the soilco2 equation. This function creates and returns a function which computes the entire right hand side of the PDE for C, and updates dY.soil.C in place with that value. This has been written so as to work with Differential Equations.jl.

make_rhs(model::BucketModel{FT}) where {FT}

Creates the rhs! function for the bucket model.

ClimaLSM.make_update_aux — Function

make_update_aux(model::RichardsModel)

An extension of the function make_update_aux, for the Richardson- Richards equation.

This function creates and returns a function which updates the auxiliary variables p.soil.variable in place.

This has been written so as to work with Differential Equations.jl.

make_update_aux(model::EnergyHydrology)

An extension of the function make_update_aux, for the integrated soil hydrology and energy model.

This function creates and returns a function which updates the auxiliary variables p.soil.variable in place.

This has been written so as to work with Differential Equations.jl.

make_update_aux(model::AbstractModel)

Return an update_aux! function that updates auxiliary parameters p.

make_update_aux(model::SoilCO2Model)

An extension of the function make_update_aux, for the soilco2 equation. This function creates and returns a function which updates the auxiliary variables p.soil.variable in place. This has been written so as to work with Differential Equations.jl.

make_update_aux(model::BucketModel{FT}) where {FT}

Creates the update_aux! function for the BucketModel.

ClimaLSM.make_set_initial_aux_state — Function

make_set_initial_aux_state(model::AbstractModel)

Returns the setinitialaux_state! function, which updates the auxiliary state p in place with the initial values corresponding to Y(t=t0) = Y0.

In principle, this function is not needed, because in the very first evaluation of the ode_function, at t=t0, the auxiliary state is updated using the initial conditions for Y=Y0. However, without setting the initial p state prior to running the simulation, the value of p in the saved output at t=t0 will be unset.

Furthermore, specific methods of this function may be useful for models which store time indepedent spatially varying parameter fields in the auxiliary state. In this case, update_aux! does not need to do anything, but they do need to be set with the initial (constant) values before the simulation can be carried out.

ClimaLSM.make_set_initial_aux_state(model::BucketModel{FT}) where{FT}

Returns the setinitialaux_state! function, which updates the auxiliary state p in place with the initial values corresponding to Y(t=t0) = Y0.

In this case, we also use this method to update the initial values for the spatially varying parameter fields, read in from data files.

ClimaLSM.prognostic_vars — Function

prognostic_vars(soil::RichardsModel)

A function which returns the names of the prognostic variables of RichardsModel.

prognostic_vars(soil::EnergyHydrology)

A function which returns the names of the prognostic variables of EnergyHydrology.

prognostic_vars(model::PlantHydraulicsModel)

A function which returns the names of the prognostic variables of the PlantHydraulicsModel.

prognostic_vars(m::AbstractModel)

Returns the prognostic variable symbols for the model in the form of a tuple.

ClimaLSM.prognostic_types — Function

prognostic_types(soil::EnergyHydrology{FT}) where {FT}

A function which returns the types of the prognostic variables of EnergyHydrology.

prognostic_types(m::AbstractModel{FT}) where {FT}

Returns the prognostic variable types for the model in the form of a tuple.

Types provided must have zero(T::DataType) defined. Common examples include

Float64, Float32 for scalar variables (a scalar value at each

coordinate point)

SVector{k,Float64} for a mutable but statically sized array of

length k at each coordinate point.

Here, the coordinate points are those returned by coordinates(model).

ClimaLSM.auxiliary_vars — Function

auxiliary_vars(soil::RichardsModel)

A function which returns the names of the auxiliary variables of RichardsModel.

Note that auxiliary variables are not needed for such a simple model. We could instead compute the conductivity and matric potential within the rhs function explicitly, rather than compute and store them in the auxiliary vector p. We did so in this case as a demonstration.

auxiliary_vars(soil::EnergyHydrology)

A function which returns the names of the auxiliary variables of EnergyHydrology.

auxiliary_vars(m::AbstractModel)

Returns the auxiliary variable symbols for the model in the form of a tuple.

ClimaLSM.auxiliary_types — Function

auxiliary_types(soil::EnergyHydrology{FT}) where {FT}

A function which returns the types of the auxiliary variables of EnergyHydrology.

auxiliary_types(m::AbstractModel{FT}) where {FT}

Returns the auxiliary variable types for the model in the form of a tuple.

Types provided must have zero(T::DataType) defined. Common examples include

Float64, Float32 for scalar variables (a scalar value at each

coordinate point)

SVector{k,Float64} for a mutable but statically sized array of

length k at each coordinate point.

Note that Arrays, MVectors are not isbits and cannot be used.

Here, the coordinate points are those returned by coordinates(model).

ClimaLSM.initialize_prognostic — Function

initialize_prognostic(model::AbstractModel, state::Union{ClimaCore.Fields.Field, Vector{FT}})

Returns a FieldVector of prognostic variables for model with the required structure, with values equal to similar(state). This assumes that all prognostic variables are defined over the entire domain, and that all prognostic variables have the same dimension and type.

If a model has no prognostic variables, the returned FieldVector contains only an empty array.

Adjustments to this - for example because different prognostic variables have different dimensions - require defining a new method.

ClimaLSM.initialize_auxiliary — Function

initialize_auxiliary(model::AbstractModel,state::Union{ClimaCore.Fields.Field, Vector{FT}})

Returns a FieldVector of auxiliary variables for model with the required structure, with values equal to similar(state). This assumes that all auxiliary variables are defined over the entire domain, and that all auxiliary variables have the same dimension and type.

If a model has no auxiliary variables, the returned FieldVector contains only an empty array.

Adjustments to this - for example because different auxiliary variables have different dimensions - require defining a new method.

ClimaLSM.initialize — Function

initialize(model::AbstractModel)

Creates the prognostic and auxiliary states structures, but with unset values; constructs and returns the coordinates for the model domain. We may need to consider this default more as we add diverse components and Simulations.

ClimaLSM.initialize(model::BucketModel{FT}) where {FT}

Initializes the variables for the BucketModel.

Note that the BucketModel has prognostic variables that are defined on different subsets of the domain. Because of that, we have to treat them independently. In LSM models which are combinations of standalone component models, this is not needed, and we can use the default initialize. Here, however, we need to do some hardcoding specific to this model.

ClimaLSM.name — Function

name(model::AbstractModel)

Returns a symbol of the model component name, e.g. :soil or :vegetation.

ClimaLSM.AbstractBC — Type

AbstractBC

An abstract type for types of boundary conditions, which will include prescribed functions of space and time as Dirichlet conditions or Neumann conditions, in addition to other convenient conditions.

ClimaLSM.AbstractSource — Type

AbstractSource{FT <: AbstractFloat}

An abstract type for types of source terms.

ClimaLSM.source! — Function

 source!(dY::ClimaCore.Fields.FieldVector,
         src::PhaseChange{FT},
         Y::ClimaCore.Fields.FieldVector,
         p::ClimaCore.Fields.FieldVector
         )

Computes the source terms for phase change.

 source!(dY::ClimaCore.Fields.FieldVector,
         src::AbstractSource,
         Y::ClimaCore.Fields.FieldVector,
         p::ClimaCore.Fields.FieldVector
         )::ClimaCore.Fields.Field

A stub function, which is extended by ClimaLSM.

ClimaLSM.source!(dY::ClimaCore.Fields.FieldVector,
                      src::RootExtraction{FT},
                      Y::ClimaCore.Fields.FieldVector,
                      p::ClimaCore.Fields.FieldVector
                      params)::ClimaCore.Fields.Field  where {FT}

An extension of the ClimaLSM.source! function, which computes source terms for the soil model; this method returns the water loss or gain due to roots when a plant hydraulic prognostic model is included.

ClimaLSM.source!(dY::ClimaCore.Fields.FieldVector,
                      src::MicrobeProduction,
                      Y::ClimaCore.Fields.FieldVector,
                      p::ClimaCore.Fields.FieldVector,
                      params)

A method which extends the ClimaLSM source! function for the case of microbe production of CO2 in soil.

ClimaLSM.AbstractBoundary — Type

AbstractBoundary{}

An abstract type to indicate which boundary we are doing calculations for. Currently, we support the top boundary (TopBoundary) and bottom boundary (BottomBoundary).

ClimaLSM.TopBoundary — Type

TopBoundary{} <: AbstractBoundary{}

A simple object which should be passed into a function to indicate that we are considering the top boundary of the soil.

ClimaLSM.BottomBoundary — Type

BottomBoundary{} <: AbstractBoundary{}

A simple object which should be passed into a function to indicate that we are considering the bottom boundary of the soil.

ClimaLSM.boundary_flux — Function

ClimaLSM.boundary_flux(bc::FluxBC, _, Δz, _...)::ClimaCore.Fields.Field

A method of boundary fluxes which returns the desired flux.

We add a field of zeros in order to convert the bc (float) into a field.

ClimaLSM.boundary_flux(rre_bc::StateBC, ::ClimaLSM.TopBoundary, Δz, p, t, params)::ClimaCore.Fields.Field

A method of boundary fluxes which converts a state boundary condition on θ_l at the top of the domain into a flux of liquid water.

ClimaLSM.boundary_flux(rre_bc::StateBC, ::ClimaLSM.BottomBoundary, Δz, p, t, params)::ClimaCore.Fields.Field

A method of boundary fluxes which converts a state boundary condition on θ_l at the bottom of the domain into a flux of liquid water.

ClimaLSM.boundary_flux(heat_bc::StateBC, ::ClimaLSM.TopBoundary, Δz, p, t)::ClimaCore.Fields.Field

A method of boundary fluxes which converts a state boundary condition on temperature at the top of the domain into a flux of energy.

ClimaLSM.boundary_flux(heat_bc::StateBC, ::ClimaLSM.BottomBoundary, Δz, p, t)::ClimaCore.Fields.Field

A method of boundary fluxes which converts a state boundary condition on temperature at the bottom of the domain into a flux of energy.

ClimaLSM.boundary_flux(bc::FreeDrainage{FT}, ::ClimaLSM.BottomBoundary, Δz, p, t)::ClimaCore.Fields.Field

A method of boundary fluxes which enforces free drainage at the bottom of the domain.

boundary_flux(bc::AbstractBC, bound_type::AbstractBoundary, Δz, _...)::ClimaCore.Fields.Field

A function which returns the correct boundary flux given any boundary condition (BC).

function ClimaLSM.boundary_flux(
    bc::RunoffBC,
    ::TopBoundary,
    Δz::FT,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
    params,
)::ClimaCore.Fields.Field where {FT}

Extension of the ClimaLSM.boundary_flux function, which returns the water volume boundary flux for the soil. At the top boundary, return the soil infiltration (computed each step and stored in p.soil_infiltration).

ClimaLSM.boundary_flux(
    bc::SoilCO2FluxBC,
    boundary::ClimaLSM.AbstractBoundary,
    Δz::ClimaCore.Fields.Field,
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::ClimaCore.Fields.Field where {FT}

A method of ClimaLSM.boundary_flux which returns the soilco2 flux (kg CO2 /m^2/s) in the case of a prescribed flux BC at either the top or bottom of the domain.

ClimaLSM.boundary_flux(
bc::SoilCO2StateBC,
boundary::ClimaLSM.TopBoundary,
Δz::ClimaCore.Fields.Field,
Y::ClimaCore.Fields.FieldVector,
p::ClimaCore.Fields.FieldVector,
t::FT,
)::ClimaCore.Fields.Field where {FT}

A method of ClimaLSM.boundary_flux which returns the soilco2 flux in the case of a prescribed state BC at top of the domain.

ClimaLSM.boundary_flux(
    bc::SoilCO2StateBC,
    boundary::ClimaLSM.BottomBoundary,
    Δz::ClimaCore.Fields.Field,
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::ClimaCore.Fields.Field where {FT}

A method of ClimaLSM.boundary_flux which returns the soilco2 flux in the case of a prescribed state BC at bottom of the domain.

ClimaLSM.diffusive_flux — Function

diffusive_flux(K, x_2, x_1, Δz)

Calculates the diffusive flux of a quantity x (water content, temp, etc). Here, x2 = x(z + Δz) and x1 = x(z), so x_2 is at a larger z by convention.

ClimaLSM.get_Δz — Function

get_Δz(z::ClimaCore.Fields.Field)

A function to return a tuple containing the distance between the top boundary and its closest center, and the bottom boundary and its closest center, both as Fields.