Shared Utilities

Domains

ClimaLSM.Domains.AbstractDomain — Type

AbstractDomain{FT <:AbstractFloat}

An abstract type for domains.

ClimaLSM.Domains.RootDomain — Type

RootDomain{FT} <: AbstractVegetationDomain{FT}

Domain for a single bulk plant with roots of varying depths. The user needs to specify the depths of the root tips as wel as the heights of the compartments to be modeled within the plant. The compartment heights are expected to be sorted in ascending order.

ClimaLSM.Domains.AbstractVegetationDomain — Type

AbstractVegetationDomain{FT} <: AbstractDomain{FT}

An abstract type for vegetation specific domains.

ClimaLSM.Domains.HybridBox — Type

struct HybridBox{FT} <: AbstractDomain{FT}
    xlim::Tuple{FT, FT}
    ylim::Tuple{FT, FT}
    zlim::Tuple{FT, FT}
    nelements::Tuple{Int, Int, Int}
    npolynomial::Int
    periodic::Tuple{Bool, Bool}
end

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space (horizontal) x a 1d finite difference space (vertical), and the resulting coordinate field.

This domain is not periodic along the z-axis. Note that only periodic domains are supported in the horizontal.

Fields

xlim

Domain interval limits along x axis, in meters

ylim

Domain interval limits along y axis, in meters

zlim

Domain interval limits along z axis, in meters

nelements

Number of elements to discretize interval, (nx, ny,nz)

npolynomial

Polynomial order for the horizontal directions

periodic

Flag indicating periodic boundaries in horizontal. only true is supported

ClimaLSM.Domains.Column — Type

Column{FT} <: AbstractDomain{FT}

A struct holding the necessary information to construct a domain, a mesh, a center and face space, etc. for use when a finite difference in 1D is suitable, as for a soil column model.

Fields

zlim

Domain interval limits, (zmin, zmax), in meters

nelements

Number of elements used to discretize the interval

boundary_tags

Boundary face identifiers

ClimaLSM.Domains.Plane — Type

Plane{FT} <: AbstractDomain{FT}

A struct holding the necessary information to construct a domain, a mesh, a 2d spectral element space, and the resulting coordinate field.

Note that only periodic domains are currently supported.

Fields

xlim

Domain interval limits along x axis, in meters

ylim

Domain interval limits along y axis, in meters

nelements

Number of elements to discretize interval, (nx, ny)

periodic

Flags for periodic boundaries; only true is supported

npolynomial

Polynomial order for both x and y

ClimaLSM.Domains.Point — Type

Point{FT} <: AbstractDomain{FT}

A domain for single column surface variables.

For models such as ponds, snow, roots, etc. Enables consistency in variable initialization across all domains.

Fields

z_sfc

Surface elevation relative to a reference (m)

ClimaLSM.Domains.LSMSingleColumnDomain — Type

LSMSingleColumnDomain{FT} <: AbstractDomain{FT}

A mixed domain, consisting of a column domain with z-coordinates at the finite difference cell centers, and a point domain, with a single z coordinate at the top boundary of the column domain.

For use in LSM modeling, where a subsurface finite difference space (for modeling soil hydrology and energy) and a surface space are both needed.

Fields

subsurface

The subsurface Column domain

surface

The surface Point domain

ClimaLSM.Domains.coordinates — Function

coordinates(model::AbstractSoilModel)

A extension of the coordinates function, which returns the coordinates of a model domain.

The coordinates are stored in the model because they are required in computing the right hand side.

coordinates(domain::AbstractDomain)

Method which returns the coordinates appropriate for a given domain.

The coordinates can be Fields or Vectors.

coordinates(domain::Union{Column{FT}, Plane{FT}, HybridBox{FT}}) where {FT}

Returns the coordinate field for the domain.

coordinates(domain::LSMSingleColumnDomain{FT}) where {FT}

Returns the coordinates of the LSMSingleColumnDomain as a named tuple, with keys of subsurface and surface.

Models

ClimaLSM.AbstractModel — Type

abstract type AbstractModel{FT <: AbstractFloat}

An abstract type for all models.

ClimaLSM.make_ode_function — Function

make_ode_function(model::AbstractModel)

Returns an ode_function that updates auxiliary variables and updates the prognostic state.

ode_function! should be compatible with OrdinaryDiffEq.jl solvers.

ClimaLSM.make_rhs — Function

make_rhs(model::AbstractModel)

Return a rhs! function that updates state variables.

rhs! should be compatible with OrdinaryDiffEq.jl solvers.

make_rhs(model::RootsModel)

A function which creates the rhs! function for the RootsModel. The rhs! function must comply with a rhs function of OrdinaryDiffEq.jl.

make_rhs(model::RichardsModel)

An extension of the function make_rhs, for the Richardson- Richards equation.

This function creates and returns a function which computes the entire right hand side of the PDE for ϑ_l, and updates dY.soil.ϑ_l in place with that value.

This has been written so as to work with Differential Equations.jl.

make_rhs(model::EnergyHydrology)

An extension of the function make_rhs, for the integrated soil energy and heat equations, including phase change.

This function creates and returns a function which computes the entire right hand side of the PDE for Y.soil.ϑ_l, Y.soil.θ_i, Y.soil.ρe_int, and updates dY.soil in place with those values.

This has been written so as to work with Differential Equations.jl.

ClimaLSM.make_update_aux — Function

make_update_aux(model::AbstractModel)

Return an update_aux! function that updates auxiliary parameters p.

make_update_aux(model::RichardsModel)

An extension of the function make_update_aux, for the Richardson- Richards equation.

This function creates and returns a function which updates the auxiliary variables p.soil.variable in place.

This has been written so as to work with Differential Equations.jl.

make_update_aux(model::EnergyHydrology)

An extension of the function make_update_aux, for the integrated soil hydrology and energy model.

This function creates and returns a function which updates the auxiliary variables p.soil.variable in place.

This has been written so as to work with Differential Equations.jl.

ClimaLSM.prognostic_vars — Function

prognostic_vars(m::AbstractModel)

Returns the prognostic variable symbols for the model in the form of a tuple.

prognostic_vars(model::RootsModel)

A function which returns the names of the prognostic variables of the RootsModel.

prognostic_vars(soil::RichardsModel)

A function which returns the names of the prognostic variables of RichardsModel.

prognostic_vars(soil::EnergyHydrology)

A function which returns the names of the prognostic variables of EnergyHydrology.

ClimaLSM.auxiliary_vars — Function

auxiliary_vars(m::AbstractModel)

Returns the auxiliary variable symbols for the model in the form of a tuple.

auxiliary_vars(soil::RichardsModel)

A function which returns the names of the auxiliary variables of RichardsModel.

Note that auxiliary variables are not needed for such a simple model. We could instead compute the conductivity and matric potential within the rhs function explicitly, rather than compute and store them in the auxiliary vector p. We did so in this case as a demonstration.

auxiliary_vars(soil::EnergyHydrology)

A function which returns the names of the auxiliary variables of EnergyHydrology.

ClimaLSM.initialize_prognostic — Function

initialize_prognostic(model::AbstractModel, state::Union{ClimaCore.Fields.Field, Vector{FT}})

Returns a FieldVector of prognostic variables for model with the required structure, with values equal to similar(state). This assumes that all prognostic variables are defined over the entire domain, and that all prognostic variables have the same dimension and type.

If a model has no prognostic variables, the returned FieldVector contains only an empty array.

Adjustments to this - for example because different prognostic variables have different dimensions - require defining a new method.

ClimaLSM.initialize_auxiliary — Function

initialize_auxiliary(model::AbstractModel,state::Union{ClimaCore.Fields.Field, Vector{FT}})

Returns a FieldVector of auxiliary variables for model with the required structure, with values equal to similar(state). This assumes that all auxiliary variables are defined over the entire domain, and that all auxiliary variables have the same dimension and type.

If a model has no auxiliary variables, the returned FieldVector contains only an empty array.

Adjustments to this - for example because different auxiliary variables have different dimensions - require defining a new method.

ClimaLSM.initialize — Function

initialize(model::AbstractModel)

Creates the prognostic and auxiliary states structures, but with unset values; constructs and returns the coordinates for the model domain. We may need to consider this default more as we add diverse components and Simulations.

ClimaLSM.name — Function

name(model::AbstractModel)

Returns a symbol of the model component name, e.g. :soil or :vegetation.