Snow Model

Snow Parameters

ClimaLand.Snow.SnowParameters — Type

SnowParameters{FT <: AbstractFloat, PSE}

A struct for storing parameters of the SnowModel.

Note that in our current implementation of runoff, a physical timescale is required and computed using Ksat and the depth of the snow. For shallow snowpacks, this will fall below the timestep of the model. For that reason, we pass the timestep of the model as a parameter, and take the larger of the timestep and the physical timescale as the value used in the model. Future implementations will revisit this.

ρ_snow: Density of snow (kg/m^3)

z_0m: Roughness length over snow for momentum (m)

z_0b: Roughness length over snow for scalars (m)

α_snow: Albedo of snow (unitless)

ϵ_snow: Emissivity of snow (unitless)

θ_r: Volumetric holding capacity of water in snow (unitless)

Ksat: Hydraulic conductivity of wet snow (m/s)

κ_ice: Thermal conductivity of ice (W/m/K)

Δt: Timestep of the model (s)

ρcD_g: Areal specific heat of ground interacting with snow (J/m^2/K)

earth_param_set: Clima-wide parameters

Snow Functions of State

```@docs ClimaLand.Snow.specificheatcapacity ClimaLand.Snow.snowsurfacetemperature ClimaLand.Snow.snowdepth ClimaLand.Snow.snowthermalconductivity ClimaLand.Snow.snowbulktemperature ClimaLand.Snow.snowliquidmassfraction ClimaLand.Snow.maximumliquidmassfraction ClimaLand.Snow.runofftimescale ClimaLand.Snow.computewaterrunoff ClimaLand.Snow.energyfromqlandswe ClimaLand.Snow.energyfromTand_swe