Soil Models

Soil Models

ClimaLSM.Soil.AbstractSoilModel — Type

AbstractSoilModel{FT} <: AbstractModel{FT}

The abstract type for all soil models.

Currently, we only have plans to support a RichardsModel, simulating the flow of liquid water through soil via the Richardson-Richards equation, and a fully integrated soil heat and water model, with phase change.

ClimaLSM.Soil.RichardsModel — Type

RichardsModel

A model for simulating the flow of water in a porous medium by solving the Richardson-Richards Equation.

parameters

the parameter set

domain

the soil domain, using ClimaCore.Domains

coordinates

the domain coordinates

boundary_conditions

the boundary conditions, of type AbstractSoilBoundaryConditions

sources

A tuple of sources, each of type AbstractSoilSource

ClimaLSM.Soil.EnergyHydrology — Type

EnergyHydrology <: AbstractSoilModel

A model for simulating the flow of water and heat in a porous medium by solving the Richardson-Richards equation and the heat equation, including terms for phase change.

parameters

The parameter sets

domain

the soil domain, using ClimaCore.Domains

coordinates

the domain coordinates

rre_boundary_conditions

the boundary conditions for RRE, of type AbstractSoilBoundaryConditions

heat_boundary_conditions

the boundary conditions for heat equation, of type AbstractSoilBoundaryConditions

sources

A tuple of sources, each of type AbstractSoilSource

Soil Diagnostic Variables

ClimaLSM.Soil.volumetric_liquid_fraction — Function

volumetric_liquid_fraction(ϑ_l::FT, ν_eff::FT) where {FT}

A pointwise function returning the volumetric liquid fraction given the augmented liquid fraction and the effective porosity.

ClimaLSM.Soil.volumetric_internal_energy — Function

volumetric_internal_energy(θ_i::FT, ρc_s::FT, T::FT) where {FT}

A pointwise function for computing the volumetric internal energy of the soil, given the volumetric ice content, volumetric heat capacity, and temperature.

ClimaLSM.Soil.volumetric_internal_energy_liq — Function

volumetric_internal_energy_liq(T::FT) where {FT}

A pointwise function for computing the volumetric internal energy of the liquid water in the soil, given the temperature T.

ClimaLSM.Soil.temperature_from_ρe_int — Function

temperature_from_ρe_int(ρe_int::FT, θ_i::FT, ρc_s::FT) where {FT}

A pointwise function for computing the temperature from the volumetric internal energy, volumetric ice content, and volumetric heat capacity of the soil.

Soil Parameters

ClimaLSM.Soil.RichardsParameters — Type

RichardsParameters{FT <: AbstractFloat}

A struct for storing parameters of the RichardModel.

ν

The porosity of the soil (m^3/m^3)

vg_α

The van Genuchten parameter α (1/m)

vg_n

The van Genuchten parameter n

vg_m

The van Genuchten parameter m

Ksat

The saturated hydraulic conductivity (m/s)

S_s

The specific storativity (1/m)

θ_r

The residual water fraction (m^3/m^3

ClimaLSM.Soil.EnergyHydrologyParameters — Type

EnergyHydrologyParameters{FT <: AbstractFloat}

A parameter structure for the integrated soil water and energy equation system. In this simplest form, we assume the conductivity and volumetric heat capacity of the soil are constant.

κ

The thermal conductivity, W/m/K

ρc_s

The volumetric heat capacity, J/m^3/K

ν

The porosity of the soil (m^3/m^3)

vg_α

The van Genuchten parameter α (1/m)

vg_n

The van Genuchten parameter n

vg_m

The van Genuchten parameter m

Ksat

The saturated hydraulic conductivity (m/s)

S_s

The specific storativity (1/m)

θ_r

The residual water fraction (m^3/m^3

earth_param_set

Physical constants and clima-wide parameters

Soil Methods and Types

ClimaLSM.Soil.boundary_fluxes — Function

function Soil.boundary_fluxes(
    bc::RunoffBC{FT},
    p::ClimaCore.Fields.FieldVector,
    t::FT,
) where {FT}

Extension of the Soil.boundary_fluxes function, which returns the water volume boundary flux for the soil, by returning the soil infiltration (computed each step and stored in p.soil_infiltration) as a top flux, and assuming no flux at the bottom of the soil. This is for use in an LSM.

boundary_fluxes(bc::FluxBC, _...)

A function which returns the correct boundary flux given the boundary condition type FluxBC.

This is a trivial example, but a more complex one would be e.g. Dirichlet conditions on the state, which then must be converted into a flux before being applied as a boundary condition.

ClimaLSM.Soil.FluxBC — Type

FluxBC{FT} <: AbstractSoilBoundaryConditions{FT}

A simple concrete type of boundary condition, which enforces constant fluxes at the top and bottom of the domain.

ClimaLSM.Soil.AbstractSoilSource — Type

AbstractSoilSource{FT <: AbstractFloat}

An abstract type for types of source terms for the soil equations.

In standalone mode, the only supported source type is freezing and thawing. ClimaLSM.jl creates additional sources to include as necessary e.g. root extraction (not available in stand alone mode).

ClimaLSM.Soil.source! — Function

Soil.source!(dY::ClimaCore.Fields.FieldVector,
                      src::RootExtraction{FT},
                      Y::ClimaCore.Fields.FieldVector,
                      p::ClimaCore.Fields.FieldVector)::ClimaCore.Fields.Field  where {FT}

An extension of the Soil.source! function, which computes source terms for the soil model; this method returns the water loss or gain due to roots when a plant hydraulic prognostic model is included.

 source!(dY::ClimaCore.Fields.FieldVector,
         src::AbstractSoilSource,
         Y::ClimaCore.Fields.FieldVector,
         p::ClimaCore.Fields.FieldVector
         )::ClimaCore.Field.Field

A stub function, which is extended by ClimaLSM.

Once we have the freeze thaw source function, we do not need this stub anymore.