Soil Models

Soil Models

ClimaLSM.Soil.AbstractSoilModel — Type

AbstractSoilModel{FT} <: AbstractModel{FT}

The abstract type for all soil models.

Currently, we only have plans to support a RichardsModel, simulating the flow of liquid water through soil via the Richardson-Richards equation, and a fully integrated soil heat and water model, with phase change.

ClimaLSM.Soil.RichardsModel — Type

RichardsModel

A model for simulating the flow of water in a porous medium by solving the Richardson-Richards Equation.

parameters

the parameter set

domain

the soil domain, using ClimaCore.Domains

coordinates

the domain coordinates

boundary_conditions

the boundary conditions, of type AbstractSoilBoundaryConditions

sources

A tuple of sources, each of type AbstractSoilSource

ClimaLSM.Soil.EnergyHydrology — Type

EnergyHydrology <: AbstractSoilModel

A model for simulating the flow of water and heat in a porous medium by solving the Richardson-Richards equation and the heat equation, including terms for phase change.

parameters

The parameter sets

domain

the soil domain, using ClimaCore.Domains

coordinates

the domain coordinates

rre_boundary_conditions

the boundary conditions for RRE, of type AbstractSoilBoundaryConditions

heat_boundary_conditions

the boundary conditions for heat equation, of type AbstractSoilBoundaryConditions

sources

A tuple of sources, each of type AbstractSoilSource

Soil Functions of State

ClimaLSM.Soil.volumetric_liquid_fraction — Function

volumetric_liquid_fraction(ϑ_l::FT, ν_eff::FT) where {FT}

A pointwise function returning the volumetric liquid fraction given the augmented liquid fraction and the effective porosity.

ClimaLSM.Soil.pressure_head — Function

pressure_head(
    α::FT,
    n::FT,
    m::FT,
    θ_r::FT,
    ϑ_l::FT,
    ν_eff::FT,
    S_s::FT,
) where {FT}

A point-wise function returning the pressure head in variably saturated soil, using the van Genuchten formulation for matric potential if the soil is not saturated, and an approximation of the positive pressure in the soil if the soil is saturated.

ClimaLSM.Soil.hydraulic_conductivity — Function

 hydraulic_conductivity(K_sat::FT, m::FT, S::FT) where {FT}

A point-wise function returning the hydraulic conductivity, using the van Genuchten formulation.

ClimaLSM.Soil.impedance_factor — Function

impedance_factor(
    f_i::FT,
    Ω::FT
) where {FT}

Returns the multiplicative factor reducing conductivity when a fraction of ice f_i is present.

Only for use with the EnergyHydrology model.

ClimaLSM.Soil.viscosity_factor — Function

viscosity_factor(
    T::FT,
    γ::FT,
    γT_ref::FT,
) where {FT}

Returns the multiplicative factor which accounts for the temperature dependence of the conductivity.

Only for use with the EnergyHydrology model.

ClimaLSM.Soil.effective_saturation — Function

effective_saturation(porosity::FT, ϑ_l::FT, θr::FT) where {FT}

A point-wise function computing the effective saturation.

ClimaLSM.Soil.matric_potential — Function

 matric_potential(α::FT, n::FT, m::FT, S::FT) where {FT}

A point-wise function returning the matric potential, using the van Genuchten formulation.

ClimaLSM.Soil.volumetric_heat_capacity — Function

volumetric_heat_capacity(
    θ_l::FT,
    θ_i::FT,
    parameters::EnergyHydrologyParameters{FT},
) where {FT}

Compute the expression for volumetric heat capacity.

ClimaLSM.Soil.κ_solid — Function

κ_solid(ν_ss_om::FT,
        ν_ss_quartz::FT,
        κ_om::FT,
        κ_quartz::FT,
        κ_minerals::FT) where {FT}

Computes the thermal conductivity of the solid material in soil. The _ss_ subscript denotes that the volumetric fractions of the soil components are referred to the soil solid components, not including the pore space.

ClimaLSM.Soil.κ_sat_frozen — Function

function κ_sat_frozen(
    κ_solid::FT,
    ν::FT,
    κ_ice::FT
) where {FT}

Computes the thermal conductivity for saturated frozen soil.

ClimaLSM.Soil.κ_sat_unfrozen — Function

function κ_sat_unfrozen(
    κ_solid::FT,
    ν::FT,
    κ_l::FT
) where {FT}

Computes the thermal conductivity for saturated unfrozen soil.

ClimaLSM.Soil.κ_sat — Function

κ_sat(
    θ_l::FT,
    θ_i::FT,
    κ_sat_unfrozen::FT,
    κ_sat_frozen::FT
) where {FT}

Compute the expression for saturated thermal conductivity of soil matrix.

ClimaLSM.Soil.κ_dry — Function

function κ_dry(ρp::FT,
               ν::FT,
               κ_solid::FT,
               κ_air::FT;
               a::FT = 0.053) where {FT}

Computes the thermal conductivity of dry soil according to the model of Balland and Arp.

ClimaLSM.Soil.kersten_number — Function

kersten_number(
    θ_i::FT,
    S_r::FT,
    parameters::EnergyHydrologyParameters{FT},
) where {FT}

Compute the expression for the Kersten number, using the Balland and Arp model.

ClimaLSM.Soil.relative_saturation — Function

relative_saturation(
        θ_l::FT,
        θ_i::FT,
        ν::FT
) where {FT}

Compute the expression for relative saturation. This is referred to as θ_sat in Balland and Arp's paper.

ClimaLSM.Soil.volumetric_internal_energy — Function

volumetric_internal_energy(θ_i::FT, ρc_s::FT, T::FT,
                             parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the volumetric internal energy of the soil, given the volumetric ice content, volumetric heat capacity, and temperature.

ClimaLSM.Soil.volumetric_internal_energy_liq — Function

volumetric_internal_energy_liq(T::FT, parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the volumetric internal energy of the liquid water in the soil, given the temperature T.

ClimaLSM.Soil.temperature_from_ρe_int — Function

temperature_from_ρe_int(ρe_int::FT, θ_i::FT, ρc_s::FT
                        parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the temperature from the volumetric internal energy, volumetric ice content, and volumetric heat capacity of the soil.

ClimaLSM.Soil.thermal_conductivity — Function

thermal_conductivity(
    κ_dry::FT,
    K_e::FT,
    κ_sat::FT
) where {FT}

Compute the expression for thermal conductivity of soil matrix.

Soil Parameters

ClimaLSM.Soil.RichardsParameters — Type

RichardsParameters{FT <: AbstractFloat}

A struct for storing parameters of the RichardModel.

ν

The porosity of the soil (m^3/m^3)

vg_α

The van Genuchten parameter α (1/m)

vg_n

The van Genuchten parameter n

vg_m

The van Genuchten parameter m

K_sat

The saturated hydraulic conductivity (m/s)

S_s

The specific storativity (1/m)

θ_r

The residual water fraction (m^3/m^3

ClimaLSM.Soil.EnergyHydrologyParameters — Type

EnergyHydrologyParameters{FT <: AbstractFloat}

A parameter structure for the integrated soil water and energy equation system. In this simplest form, we assume the conductivity and volumetric heat capacity of the soil are constant.

κ_dry

The dry soil thermal conductivity, W/m/K

κ_sat_frozen

The saturated thermal conductivity of frozen soil, W/m/K

κ_sat_unfrozen

The saturated thermal conductivity of unfrozen soil, W/m/K

ρc_ds

The volumetric heat capacity of dry soil, J/m^3/K

ν

The porosity of the soil (m^3/m^3)

ν_ss_om

The volumetric fraction of the soil solids in organic matter (m^3/m^3)

ν_ss_quartz

The volumetric fraction of the soil solids in quartz (m^3/m^3)

ν_ss_gravel

The volumetric fraction of the soil solids in gravel (m^3/m^3)

α

The parameter α used in computing Kersten number, unitless

β

The parameter β used in computing Kersten number, unitless

vg_α

The van Genuchten parameter α (1/m)

vg_n

The van Genuchten parameter n

vg_m

The van Genuchten parameter m

K_sat

The saturated hydraulic conductivity (m/s)

S_s

The specific storativity (1/m)

θ_r

The residual water fraction (m^3/m^3

Ω

Ice impedance factor for the hydraulic conductivity

γ

Coefficient of viscosity factor for the hydraulic conductivity

γT_ref

Reference temperature for the viscosity factor

earth_param_set

Physical constants and clima-wide parameters

Soil Methods and Types

ClimaLSM.Soil.boundary_fluxes — Function

boundary_fluxes(bc::FluxBC, _...)

A function which returns the correct boundary flux given the boundary condition type FluxBC.

This is a trivial example, but a more complex one would be e.g. Dirichlet conditions on the state, which then must be converted into a flux before being applied as a boundary condition.

function Soil.boundary_fluxes(
    bc::RunoffBC{FT},
    p::ClimaCore.Fields.FieldVector,
    t::FT,
) where {FT}

Extension of the Soil.boundary_fluxes function, which returns the water volume boundary flux for the soil, by returning the soil infiltration (computed each step and stored in p.soil_infiltration) as a top flux, and assuming no flux at the bottom of the soil. This is for use in an LSM.

ClimaLSM.Soil.FluxBC — Type

FluxBC{FT} <: AbstractSoilBoundaryConditions{FT}

A simple concrete type of boundary condition, which enforces constant normal fluxes at the top and bottom of the domain.

ClimaLSM.Soil.AbstractSoilSource — Type

AbstractSoilSource{FT <: AbstractFloat}

An abstract type for types of source terms for the soil equations.

In standalone mode, the only supported source type is freezing and thawing. ClimaLSM.jl creates additional sources to include as necessary e.g. root extraction (not available in stand alone mode).

ClimaLSM.Soil.source! — Function

 source!(dY::ClimaCore.Fields.FieldVector,
         src::AbstractSoilSource,
         Y::ClimaCore.Fields.FieldVector,
         p::ClimaCore.Fields.FieldVector
         )::ClimaCore.Field.Field

A stub function, which is extended by ClimaLSM.

Once we have the freeze thaw source function, we do not need this stub anymore.

Soil.source!(dY::ClimaCore.Fields.FieldVector,
                      src::RootExtraction{FT},
                      Y::ClimaCore.Fields.FieldVector,
                      p::ClimaCore.Fields.FieldVector)::ClimaCore.Fields.Field  where {FT}

An extension of the Soil.source! function, which computes source terms for the soil model; this method returns the water loss or gain due to roots when a plant hydraulic prognostic model is included.