Soil Models

Soil Models

ClimaLSM.Soil.AbstractSoilModel — Type

AbstractSoilModel{FT} <: ClimaLSM.AbstractImExModel{FT}

The abstract type for all soil models.

Currently, we only have plans to support a RichardsModel, simulating the flow of liquid water through soil via the Richardson-Richards equation, and a fully integrated soil heat and water model, with phase change.

ClimaLSM.Soil.RichardsModel — Type

RichardsModel

A model for simulating the flow of water in a porous medium by solving the Richardson-Richards Equation.

parameters: the parameter set

domain: the soil domain, using ClimaCore.Domains

boundary_conditions: the boundary conditions, of type AbstractSoilBoundaryConditions

sources: A tuple of sources, each of type AbstractSoilSource

ClimaLSM.Soil.EnergyHydrology — Type

EnergyHydrology <: AbstractSoilModel

A model for simulating the flow of water and heat in a porous medium by solving the Richardson-Richards equation and the heat equation, including terms for phase change.

parameters: The parameter sets

domain: the soil domain, using ClimaCore.Domains

boundary_conditions: the boundary conditions for RRE and heat, of type AbstractSoilBoundaryConditions

sources: A tuple of sources, each of type AbstractSoilSource

Soil Functions of State

ClimaLSM.Soil.volumetric_liquid_fraction — Function

volumetric_liquid_fraction(ϑ_l::FT, ν_eff::FT, θ_r::FT) where {FT}

A pointwise function returning the volumetric liquid fraction given the augmented liquid fraction and the effective porosity.

ClimaLSM.Soil.pressure_head — Function

pressure_head(
    cm::vanGenuchten{FT},
    θ_r::FT,
    ϑ_l::FT,
    ν_eff::FT,
    S_s::FT,
) where {FT}

A point-wise function returning the pressure head in variably saturated soil, using the van Genuchten matric potential if the soil is not saturated, and an approximation of the positive pressure in the soil if the soil is saturated.

pressure_head(
    cm::BrooksCorey{FT},
    θ_r::FT,
    ϑ_l::FT,
    ν_eff::FT,
    S_s::FT,
) where {FT}

A point-wise function returning the pressure head in variably saturated soil, using the Brooks and Corey matric potential if the soil is not saturated, and an approximation of the positive pressure in the soil if the soil is saturated.

ClimaLSM.Soil.hydraulic_conductivity — Function

 hydraulic_conductivity(cm::vanGenuchten{FT}, K_sat::FT, S::FT) where {FT}

A point-wise function returning the hydraulic conductivity, using the van Genuchten formulation.

 hydraulic_conductivity(cm::BrooksCorey{FT}, K_sat::FT, S::FT) where {FT}

A point-wise function returning the hydraulic conductivity, using the Brooks and Corey formulation.

ClimaLSM.Soil.impedance_factor — Function

impedance_factor(
    f_i::FT,
    Ω::FT
) where {FT}

Returns the multiplicative factor reducing conductivity when a fraction of ice f_i is present.

Only for use with the EnergyHydrology model.

ClimaLSM.Soil.viscosity_factor — Function

viscosity_factor(
    T::FT,
    γ::FT,
    γT_ref::FT,
) where {FT}

Returns the multiplicative factor which accounts for the temperature dependence of the conductivity.

Only for use with the EnergyHydrology model.

ClimaLSM.Soil.effective_saturation — Function

effective_saturation(porosity::FT, ϑ_l::FT, θr::FT) where {FT}

A point-wise function computing the effective saturation.

ClimaLSM.Soil.matric_potential — Function

 matric_potential(cm::vanGenuchten{FT}, S::FT) where {FT}

A point-wise function returning the matric potential, using the van Genuchten formulation.

 matric_potential(cm::BrooksCorey{FT}, S::FT) where {FT}

A point-wise function returning the matric potential, using the Brooks and Corey formulation.

ClimaLSM.Soil.inverse_matric_potential — Function

 inverse_matric_potential(cm::vanGenuchten{FT}, ψ::FT) where {FT}

A point-wise function returning the effective saturation, given the matric potential, using the van Genuchten formulation.

 inverse_matric_potential(cm::BrooksCorey{FT}, ψ::FT) where {FT}

A point-wise function returning the effective saturation, given the matric potential, using the Brooks and Corey formulation.

ClimaLSM.Soil.volumetric_heat_capacity — Function

volumetric_heat_capacity(
    θ_l::FT,
    θ_i::FT,
    parameters::EnergyHydrologyParameters{FT},
) where {FT}

Compute the expression for volumetric heat capacity.

ClimaLSM.Soil.κ_solid — Function

κ_solid(ν_ss_om::FT,
        ν_ss_quartz::FT,
        κ_om::FT,
        κ_quartz::FT,
        κ_minerals::FT) where {FT}

Computes the thermal conductivity of the solid material in soil. The _ss_ subscript denotes that the volumetric fractions of the soil components are referred to the soil solid components, not including the pore space.

ClimaLSM.Soil.κ_sat_frozen — Function

function κ_sat_frozen(
    κ_solid::FT,
    ν::FT,
    κ_ice::FT
) where {FT}

Computes the thermal conductivity for saturated frozen soil.

ClimaLSM.Soil.κ_sat_unfrozen — Function

function κ_sat_unfrozen(
    κ_solid::FT,
    ν::FT,
    κ_l::FT
) where {FT}

Computes the thermal conductivity for saturated unfrozen soil.

ClimaLSM.Soil.κ_sat — Function

κ_sat(
    θ_l::FT,
    θ_i::FT,
    κ_sat_unfrozen::FT,
    κ_sat_frozen::FT
) where {FT}

Compute the expression for saturated thermal conductivity of soil matrix.

ClimaLSM.Soil.κ_dry — Function

function κ_dry(ρp::FT,
               ν::FT,
               κ_solid::FT,
               κ_air::FT;
               a = FT(0.053)) where {FT}

Computes the thermal conductivity of dry soil according to the model of Balland and Arp.

ClimaLSM.Soil.kersten_number — Function

kersten_number(
    θ_i::FT,
    S_r::FT,
    parameters::EnergyHydrologyParameters{FT},
) where {FT}

Compute the expression for the Kersten number, using the Balland and Arp model.

ClimaLSM.Soil.relative_saturation — Function

relative_saturation(
        θ_l::FT,
        θ_i::FT,
        ν::FT
) where {FT}

Compute the expression for relative saturation. This is referred to as θ_sat in Balland and Arp's paper.

ClimaLSM.Soil.volumetric_internal_energy — Function

volumetric_internal_energy(θ_i::FT, ρc_s::FT, T::FT,
                             parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the volumetric internal energy of the soil, given the volumetric ice content, volumetric heat capacity, and temperature.

ClimaLSM.Soil.volumetric_internal_energy_liq — Function

volumetric_internal_energy_liq(T::FT, parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the volumetric internal energy of the liquid water in the soil, given the temperature T.

ClimaLSM.Soil.temperature_from_ρe_int — Function

temperature_from_ρe_int(ρe_int::FT, θ_i::FT, ρc_s::FT
                        parameters::EnergyHydrologyParameters{FT}) where {FT}

A pointwise function for computing the temperature from the volumetric internal energy, volumetric ice content, and volumetric heat capacity of the soil.

ClimaLSM.Soil.thermal_conductivity — Function

thermal_conductivity(
    κ_dry::FT,
    K_e::FT,
    κ_sat::FT
) where {FT}

Compute the expression for thermal conductivity of soil matrix.

ClimaLSM.Soil.phase_change_source — Function

phase_change_source(
    θ_l::FT,
    θ_i::FT,
    T::FT,
    τ::FT,
    params::EnergyHydrologyParameters{FT},
) where {FT}

Returns the source term (1/s) used for converting liquid water and ice into each other during phase changes. Note that there are unitless prefactors multiplying this term in the equations.

Note that these equations match what is in Dall'Amico (for θstar, ψ(T), ψw0). We should double check them in the case where we have ϑl > θl, but they should be very close to the form we want regardless.

ClimaLSM.Soil.thermal_time — Function

thermal_time(ρc::FT, Δz::FT, κ::FT) where {FT}

Returns the thermal timescale for temperature differences across a typical thickness Δz to equilibrate.

ClimaLSM.Soil.dψdϑ — Function

dψdϑ(cm::vanGenuchten{FT}, ϑ, ν, θr, Ss)

Computes and returns the derivative of the pressure head with respect to ϑ for the van Genuchten formulation.

dψdϑ(cm::BrooksCorey{FT}, ϑ, ν, θr, Ss)

Computes and returns the derivative of the pressure head with respect to ϑ for the Brooks and Corey formulation.

Missing docstring.

Missing docstring for ClimaLSM.soil.soil_resistance. Check Documenter's build log for details.

ClimaLSM.Soil.dry_soil_layer_thickness — Function

dry_soil_layer_thickness(S_l_sfc::FT, S_c::FT, d_ds::FT) where {FT}

Returns the maximum dry soil layer thickness that can develop under evaporation; this is used when computing the soil resistance to evaporation according to Swenson et al (2012).

ClimaLSM.Soil.soil_tortuosity — Function

soil_tortuosity(θ_l::FT, θ_i::FT, ν::FT) where {FT}

Computes the tortuosity of water vapor in a porous medium, as a function of porosity ν and the volumetric liquid water and ice contents, θ_l and θ_i.

See Equation (1) of : Shokri, N., P. Lehmann, and D. Or (2008), Effects of hydrophobic layers on evaporation from porous media, Geophys. Res. Lett., 35, L19407, doi:10.1029/ 2008GL035230.

Soil Parameters

ClimaLSM.Soil.RichardsParameters — Type

RichardsParameters{FT <: AbstractFloat}

A struct for storing parameters of the RichardModel.

ν: The porosity of the soil (m^3/m^3)

hydrology_cm: The hydrology closure model: vanGenuchten or BrooksCorey

K_sat: The saturated hydraulic conductivity (m/s)

S_s: The specific storativity (1/m)

θ_r: The residual water fraction (m^3/m^3

ClimaLSM.Soil.EnergyHydrologyParameters — Type

EnergyHydrologyParameters{FT <: AbstractFloat}

A parameter structure for the integrated soil water and energy equation system. In this simplest form, we assume the conductivity and volumetric heat capacity of the soil are constant.

κ_dry: The dry soil thermal conductivity, W/m/K

κ_sat_frozen: The saturated thermal conductivity of frozen soil, W/m/K

κ_sat_unfrozen: The saturated thermal conductivity of unfrozen soil, W/m/K

ρc_ds: The volumetric heat capacity of dry soil, J/m^3/K

ν: The porosity of the soil (m^3/m^3)

ν_ss_om: The volumetric fraction of the soil solids in organic matter (m^3/m^3)

ν_ss_quartz: The volumetric fraction of the soil solids in quartz (m^3/m^3)

ν_ss_gravel: The volumetric fraction of the soil solids in gravel (m^3/m^3)

α: The parameter α used in computing Kersten number, unitless

β: The parameter β used in computing Kersten number, unitless

hydrology_cm: The soil hydrology closure model: van Genucthen or Brooks and Corey

K_sat: The saturated hydraulic conductivity (m/s)

S_s: The specific storativity (1/m)

θ_r: The residual water fraction (m^3/m^3

Ω: Ice impedance factor for the hydraulic conductivity

γ: Coefficient of viscosity factor for the hydraulic conductivity

γT_ref: Reference temperature for the viscosity factor

albedo: Soil Albedo

emissivity: Soil Emissivity

z_0m: Roughness length for momentum

z_0b: Roughness length for scalars

d_ds: Maximum dry soil layer thickness under evaporation (m)

earth_param_set: Physical constants and clima-wide parameters

Soil BC and Source Methods and Types

ClimaLSM.Soil.AbstractSoilBC — Type

AbstractSoilBC <: ClimaLSM. AbstractBC

An abstract type for soil-specific types of boundary conditions, like free drainage.

ClimaLSM.Soil.MoistureStateBC — Type

MoistureStateBC <: AbstractSoilBC

A simple concrete type of boundary condition, which enforces a state boundary condition ϑ_l = f(p,t) at either the top or bottom of the domain.

ClimaLSM.Soil.FluxBC — Type

FluxBC <: AbstractSoilBC

A simple concrete type of boundary condition, which enforces a normal flux boundary condition f(p,t) at either the top or bottom of the domain.

ClimaLSM.Soil.TemperatureStateBC — Type

TemperatureStateBC <: AbstractSoilBC

A simple concrete type of boundary condition, which enforces a state boundary condition T = f(p,t) at either the top or bottom of the domain.

ClimaLSM.Soil.FreeDrainage — Type

FreeDrainage <: AbstractSoilBC

A concrete type of soil boundary condition, for use at the BottomBoundary only, where the flux is set to be F = -K∇h = -K.

ClimaLSM.Soil.AtmosDrivenFluxBC — Type

AtmosDrivenFluxBC{
    A <: AbstractAtmosphericDrivers,
    B <: AbstractRadiativeDrivers,
} <: AbstractSoilBC

A concrete type of soil boundary condition for use at the top of the domain. This holds the conditions for the atmosphere AbstractAtmosphericDrivers and for the radiation state AbstractRadiativeDrivers.

This choice indicates the Monin-Obukhov Surface Theory will be used to compute the sensible and latent heat fluxes, as well as evaporation, and that the net radiation and precipitation will also be computed. The net energy and water fluxes are used as boundary conditions.

ClimaLSM.Soil.AbstractSoilSource — Type

AbstractSoilSource{FT} <:  ClimaLSM.AbstractSource{FT}

An abstract type for types of source terms for the soil equations.

In standalone mode, the only supported source type is freezing and thawing. ClimaLSM.jl creates additional sources to include as necessary e.g. root extraction (not available in stand alone mode).

ClimaLSM.Soil.PhaseChange — Type

PhaseChange{FT} <: AbstractSoilSource{FT}

PhaseChange source type.

ClimaLSM.Soil.AbstractSoilHydrologyClosure — Type

AbstractSoilHydrologyClosure{FT <: AbstractFloat}

The abstract type of soil hydrology closure, of which vanGenuchten{FT} and BrooksCorey{FT} are the two supported concrete types.

To add a new parameterization, methods are required for:

matric_potential,

inversematricpotential,

pressure_head,

dψdϑ,

hydraulic_conductivity.

ClimaLSM.Soil.vanGenuchten — Type

vanGenuchten{FT} <: AbstractSoilHydrologyClosure{FT}

The van Genuchten soil hydrology closure, chosen when the hydraulic conductivity and matric potential are modeled using the van Genuchten parameterization (van Genuchten 1980; see also Table 8.2 of G. Bonan 2019).

α: The inverse of the air entry potential (1/m)

n: The van Genuchten pore-size distribution index (unitless)

m: The van Genuchten parameter m = 1 - 1/n (unitless)

S_c: A derived parameter: the critical saturation at which capillary flow no longer replenishes the surface

ClimaLSM.Soil.BrooksCorey — Type

BrooksCorey{FT} <: AbstractSoilHydrologyClosure{FT}

The Brooks and Corey soil hydrology closure, chosen when the hydraulic conductivity and matric potential are modeled using the Brooks and Corey parameterization (Brooks and Corey, 1964, 1966; see also Table 8.2 of G. Bonan 2019).

c: The pore-size distribution index (unitless)

ψb: The air entry matric potential, when S=1 (m)

S_c: A derived parameter: the critical saturation at which capillary flow no longer replenishes the surface

Soil Jacobian Structures

ClimaLSM.Soil.RichardsTridiagonalW — Type

RichardsTridiagonalW{R, J, W, T} <: ClimaLSM.AbstractTridiagonalW

A struct containing the necessary information for constructing a tridiagonal Jacobian matrix (W) for solving Richards equation, treating only the vertical diffusion term implicitly.

Note that the diagonal, upper diagonal, and lower diagonal entry values are stored in this struct and updated in place.

dtγ_ref: Reference to dtγ, which is specified by the ODE solver

∂ϑₜ∂ϑ: Diagonal entries of the Jacobian stored as a ClimaCore.Fields.Field

W_column_arrays: Array of tridiagonal matrices containing W for each column

temp1: An allocated cache used to evaluate ldiv!

temp2: An allocated cache used to evaluate ldiv!

transform: A flag indicating whether this struct is used to compute Wfact_t or Wfact

ones_face_space: A pre-allocated cache storing ones on the face space