Soil Biogeochemistry

Functions of State

ClimaLSM.Soil.Biogeochemistry.volumetric_air_content — Function

volumetric_air_content(θ_w::FT, params::DETECTModelParameters{FT}) where {FT}

Computes the volumetric air content (θ_a) in the soil, which is related to the total soil porosity (ν) and volumetric soil water content (θ_w = θ_l+θ_i).

ClimaLSM.Soil.Biogeochemistry.co2_diffusivity — Function

co2_diffusivity(
                T_soil::FT,
                P_sfc::FT,
                θ_a::FT,
                params::DETECTModelParameters{FT}
                ) where {FT}

Computes the diffusivity of CO₂ within the soil (D).

First, D0 is computed using the temperature within the soil (T_soil in K) and pressure at the surface of the soil (P_sfc in Pa), using reference values of T_ref and P_ref (273 K and 101325 Pa). Here, θ_a is the Volumetric air content and θ_a100 is the volumetric air content at a soil water potential of 100cm, and b is the pore size distribution of the soil.

ClimaLSM.Soil.Biogeochemistry.root_source_moisture_coeff — Function

root_source_moisture_coeff(
                           θ_l::FT,
                           θ_ant_roots::FT,
                           params::DETECTModelParameters{FT}
                           ) where {FT}

Returns the scaling coefficient/adjustment factor for the root co2 source term, accounting for antecedent "ant" conditions of soil moisture θ_l.

ClimaLSM.Soil.Biogeochemistry.source_temperature_coeff — Function

source_temperature_coeff(
    T_soil::FT,
    T_ant_soil::FT,
    params::DETECTModelParameters{FT},
) where {FT}

Computes the temperature scaling factor function (unitless), which is motivated by Lloyd and Taylor (1994).

ClimaLSM.Soil.Biogeochemistry.energy_activation — Function

energy_activation(T_ant_soil::FT, params::DETECTModelParameters{FT}) where {FT}

Computes the energy activation temperature based on the antecendent soil temperature T_ant_soil.

This corresponds to an energy of activation term that governs the apparent temperature sensitivity of the root and microbe source term.

ClimaLSM.Soil.Biogeochemistry.root_source — Function

root_source(
            T_soil::FT,
            T_ant_soil::FT,
            θ_l::FT,
            θ_ant_roots::FT,
            Cr::FT,
            params::DETECTModelParameters{FT}
            ) where {FT}

Computes the CO₂ production in the soil by roots, in depth and time (kg C / m^3/s).

ClimaLSM.Soil.Biogeochemistry.microbe_source_moisture_coeff — Function

microbe_source_moisture_coeff(
                              θ_l::FT,
                              θ_ant_microbe::FT,
                              params::DETECTModelParameters{FT}
                              ) where {FT}

Returns the scaling coefficient/adjustment factor for the microbe co2 source term, accounting for antecedent "ant" conditions of soil moisture θ_l.

ClimaLSM.Soil.Biogeochemistry.decomposition_potential — Function

decomposition_potential(
                        T_soil::FT,
                        T_ant_soil::FT,
                        θ_l::FT,
                        θ_ant_microbe::FT,
                        params::DETECTModelParameters{FT}
                        ) where {FT}

Computes the maximum potential decomposition rate (kg C/m^3/s).

ClimaLSM.Soil.Biogeochemistry.soluble_soil_carbon — Function

soluble_soil_carbon(
                    θ_l::FT,
                    Csom::FT,
                    params::DETECTModelParameters{FT}
                    ) where {FT}

Computes the soluble soil-C pool (kg C/m^3).

ClimaLSM.Soil.Biogeochemistry.microbe_source — Function

microbe_source(
               T_soil::FT,
               T_ant_soil::FT,
               θ_l::FT,
               θ_ant_microbe::FT,
               Csom::FT,
               Cmic::FT,
               params::DETECTModelParameters{FT},
               ) where {FT}

Computes the CO₂ production in the soil by microbes, in depth and time (kg C / m^3/s).

Parameter Structures

ClimaLSM.Soil.Biogeochemistry.DETECTModelParameters — Type

DETECTParameters{FT <: AbstractFloat}

A struct for storing parameters of the DETECTModel.