Soil Biogeochemistry

Model Structure

ClimaLSM.Soil.Biogeochemistry.SoilCO2Model — Type

SoilCO2Model

A model for simulating the production and transport of CO₂ in the soil with dynamic source and diffusion terms.

parameters

: the parameter set

domain

: the soil domain, using ClimaCore.Domains

boundary_conditions

: the boundary conditions, of type AbstractBoundaryConditions

sources

: A tuple of sources, each of type AbstractSource

driver

: Drivers

Parameter Structure

ClimaLSM.Soil.Biogeochemistry.SoilCO2ModelParameters — Type

SoilCO2ModelParameters{FT <: AbstractFloat}

A struct for storing parameters of the SoilCO2Model.

P_sfc

: Pressure at the surface of the soil (Pa)

Rb

: Root mass-base respiration rate at 10°C and mean environmental conditions (kg C m⁻³ s⁻¹)

α1r

: The effect of soil water content (θ) on root respiration (unitless)

α2r

: The effect of antecedent θ on root respiration (unitless)

α3r

: The interactive effect of θ and antecedent θ on root respiration (unitless)

Vb

: Value of Vₘₐₓ at 10°C and mean environmental conditions (kg C m⁻³ s⁻¹)

α1m

: The effect of soil water content (θ) on microbial respiration (unitless)

α2m

: The effect of antecedent θ on microbial respiration (unitless)

α3m

: The interactive effect of θ and antecedent θ on microbial respiration (unitless)

Km

: Michaelis-Menten half-saturation constant (kg C m⁻³ s⁻¹)

CUE

: Microbial carbon-use efficiency (kg C kg⁻¹ C⁻¹)

soluble_fraction

: Fraction of soil organic C that is soluble (-)

D_liq

: Diffusivity of soil C substrate in liquid (unitless)

Estar

: Temperature sensitivity parameter, somewhat analogous to an energy activation (Kelvin)

α4

: The effect of antecedent soil temperature on root and microbial respiration (unitless)

T_ref_soil

: Reference temperature (Kelvin)

ν

: Soil porosity (m³ m⁻³)

θ_a100

: Air-filled porosity at soil water potential of -100 cm H₂O (~ 10 Pa)

D_ref

: Diffusion coefficient for CO₂ in air at standard temperature and pressure (m² s⁻¹)

b

: Absolute value of the slope of the line relating log(ψ) versus log(θ) (unitless)

earth_param_set

: Physical constants used Clima-wide

Model-specific Types

ClimaLSM.Soil.Biogeochemistry.RootProduction — Type

RootProduction{FT} <: AbstractCarbonSource{FT}

Struct for the root production of CO2, which appears as a source term in the differential equation.

ClimaLSM.Soil.Biogeochemistry.MicrobeProduction — Type

MicrobeProduction{FT} <: AbstractCarbonSource{FT}

Struct for the microbe production of CO2, appearing as a source term in the differential equation.

ClimaLSM.Soil.Biogeochemistry.SoilCO2FluxBC — Type

SoilCO2FluxBC <: AbstractSoilCO2BC

A container holding the CO2 flux boundary condition, which is a function f(p,t), where p is the auxiliary state vector.

ClimaLSM.Soil.Biogeochemistry.SoilCO2StateBC — Type

SoilCO2StateBC <: AbstractSoilCO2BC

A container holding the CO2 state boundary condition (kg CO2 m−3), which is a function f(p,t), where p is the auxiliary state vector.

ClimaLSM.Soil.Biogeochemistry.AbstractSoilDriver — Type

AbstractSoilDriver

An abstract type for drivers of soil CO2 production. These are soil temperature, soil moisture, root carbon, soil organic matter and microbe carbon. All varying in space (horizontally and vertically) and time.

ClimaLSM.Soil.Biogeochemistry.PrescribedSoil — Type

PrescribedSoil <: AbstractSoilDriver

A container which holds the prescribed functions for soil temperature and moisture.

This is meant for use when running the biogeochemistry model in standalone mode, without a prognostic soil model.

temperature

: The temperature of the soil, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

volumetric_liquid_fraction

: Soil moisture, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

antecedent_temperature

: Antecedant soil temperature, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

antecedent_volumetric_liquid_fraction_m

: Antecedant soil moisture acting on microbe, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

antecedent_volumetric_liquid_fraction_r

: Antecedant soil moisture acting on roots, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

root_carbon

: Carbon content of root in soil, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

soil_organic_carbon

: Carbon content of soil organic matter, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

microbe_carbon

: Carbon content of microbes in soil, of the form f(z::FT,t::FT) where {FT <: AbstractFloat}

Functions of State

ClimaLSM.Soil.Biogeochemistry.volumetric_air_content — Function

volumetric_air_content(θ_w::FT,
                       params::SoilCO2ModelParameters{FT}
                       ) where {FT}

Computes the volumetric air content (θ_a) in the soil, which is related to the total soil porosity (ν) and volumetric soil water content (θ_w = θ_l+θ_i).

ClimaLSM.Soil.Biogeochemistry.co2_diffusivity — Function

co2_diffusivity(
                T_soil::FT,
                θ_w::FT,
                params::SoilCO2ModelParameters{FT}
                ) where {FT}

Computes the diffusivity of CO₂ within the soil (D).

First, D0 is computed using the temperature within the soil (T_soil in K) and pressure at the surface of the soil (P_sfc in Pa), using reference values of T_ref and P_ref (273 K and 101325 Pa). Here, θ_a is the volumetric air content and θ_a100 is the volumetric air content at a soil water potential of 100cm, and b is the pore size distribution of the soil.

ClimaLSM.Soil.Biogeochemistry.root_source_moisture_coeff — Function

root_source_moisture_coeff(
                           θ_l::FT,
                           θ_ant_roots::FT,
                           params::SoilCO2ModelParameters{FT}
                           ) where {FT}

Returns the scaling coefficient/adjustment factor for the root co2 source term, accounting for antecedent "ant" conditions of soil moisture θ_l.

ClimaLSM.Soil.Biogeochemistry.source_temperature_coeff — Function

source_temperature_coeff(
    T_soil::FT,
    T_ant_soil::FT,
    params::SoilCO2ModelParameters{FT},
) where {FT}

Computes the temperature scaling factor function (unitless), which is motivated by Lloyd and Taylor (1994).

ClimaLSM.Soil.Biogeochemistry.energy_activation — Function

energy_activation(T_ant_soil::FT, params::SoilCO2ModelParameters{FT}) where {FT}

Computes the energy activation temperature based on the antecendent soil temperature T_ant_soil.

This corresponds to an energy of activation term that governs the apparent temperature sensitivity of the root and microbe source term.

ClimaLSM.Soil.Biogeochemistry.root_source — Function

root_source(
            T_soil::FT,
            T_ant_soil::FT,
            θ_l::FT,
            θ_ant_roots::FT,
            Cr::FT,
            params::SoilCO2ModelParameters{FT}
            ) where {FT}

Computes the CO₂ production in the soil by roots, in depth and time (kg C / m^3/s).

ClimaLSM.Soil.Biogeochemistry.microbe_source_moisture_coeff — Function

microbe_source_moisture_coeff(
                              θ_l::FT,
                              θ_ant_microbe::FT,
                              params::SoilCO2ModelParameters{FT}
                              ) where {FT}

Returns the scaling coefficient/adjustment factor for the microbe co2 source term, accounting for antecedent "ant" conditions of soil moisture θ_l.

ClimaLSM.Soil.Biogeochemistry.decomposition_potential — Function

decomposition_potential(
                        T_soil::FT,
                        T_ant_soil::FT,
                        θ_l::FT,
                        θ_ant_microbe::FT,
                        params::SoilCO2ModelParameters{FT}
                        ) where {FT}

Computes the maximum potential decomposition rate (kg C/m^3/s).

ClimaLSM.Soil.Biogeochemistry.soluble_soil_carbon — Function

soluble_soil_carbon(
                    θ_l::FT,
                    Csom::FT,
                    params::SoilCO2ModelParameters{FT}
                    ) where {FT}

Computes the soluble soil-C pool (kg C/m^3).

ClimaLSM.Soil.Biogeochemistry.microbe_source — Function

microbe_source(
               T_soil::FT,
               T_ant_soil::FT,
               θ_l::FT,
               θ_ant_microbe::FT,
               Csom::FT,
               Cmic::FT,
               params::SoilCO2ModelParameters{FT},
               ) where {FT}

Computes the CO₂ production in the soil by microbes, in depth and time (kg C / m^3/s).