Vegetation

Models

ClimaLSM.PlantHydraulics.AbstractVegetationModel — Type

AbstractVegetationModel{FT} <: AbstractModel{FT}

An abstract type for vegetation models. Concrete types include a plant hydraulics model, but future types will include multi-layer canopy models and possibly a big leaf model.

ClimaLSM.PlantHydraulics.PlantHydraulicsModel — Type

PlantHydraulicsModel{FT, PS, D, RE, T, B} <: AbstractVegetationModel{FT}

Defines, and constructs instances of, the PlantHydraulicsModel type, which is used for simulation flux of water to/from soil, along roots of different depths, along a stem, to a leaf, and ultimately being lost from the system by transpiration. This model can be used in standalone mode by prescribing the transpiration rate and soil pressure at the root tips or flux in the roots, or with a dynamic soil model using ClimaLSM.

parameters

: Parameters required by the Plant Hydraulics model

domain

: The plant hydraulics model domain, of type AbstractVegetationDomain

root_extraction

: The root extraction model, of type AbstractRootExtraction

transpiration

: The transpiration model, of type AbstractTranspiration

Plan Hydraulics Diagnostic Variables

ClimaLSM.PlantHydraulics.effective_saturation — Function

effective_saturation(
    ν::FT, 
    ϑ_l::FT) where {FT}

Computes the effective saturation (volumetric water content relative to porosity).

ClimaLSM.PlantHydraulics.augmented_liquid_fraction — Function

augmented_liquid_fraction(
    ν::FT, 
    S_l::FT) where {FT}

Computes the augmented liquid fraction from porosity and effective saturation. Augmented liquid fraction allows for oversaturation: an expansion of the volume of space available for storage in a plant compartment. It is analogous to the augmented liquid fraction state variable in the soil model.

ClimaLSM.PlantHydraulics.water_retention_curve — Function

water_retention_curve(
    vg_α::FT,
    vg_n::FT,
    vg_m::FT,
    S_l::FT,
    ν::FT,
    S_s::FT) where {FT}

Converts augmented liquid fraction (ϑl) to effective saturation (Sl), and then effective saturation to absolute pressure (p), for both the unsaturated and saturated regimes. Note pressure is in SI units [m].

ClimaLSM.PlantHydraulics.inverse_water_retention_curve — Function

inverse_water_retention_curve(
    vg_α::FT,
    vg_n::FT,
    vg_m::FT,
    p::FT,
    ν::FT,
    S_s::FT) where {FT}

Converts absolute pressure (p) to effective saturation (S_l).

ClimaLSM.PlantHydraulics.flux_out_roots — Function

PlantHydraulics.flux_out_roots(
    re::PrognosticSoilPressure{FT},
    model::PlantHydraulics.PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

An extension of the PlantHydraulics.flux_out_roots function, which returns the net flux of water between the roots and the soil, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water between roots and soil at each soil layer.

PlantHydraulics.flux_out_roots(
    re::PrognosticRootFlux{FT},
    t::FT,
)::FT where {FT}

An extension of the PlantHydraulics.flux_out_roots function, which returns the net flux of water between the roots and the soil, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water between roots and soil at each soil layer.

flux_out_roots(
    re::PrescribedSoilPressure{FT},
    model::PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

A method which computes the flux between the soil and the stem, via the roots, in the case of a standalone plant hydraulics model with prescribed soil pressure (in m) at the root tips. This assumes that the stem compartment is the first element of Y.vegetation.ϑ_l.

Plan Hydraulics Parameters

ClimaLSM.PlantHydraulics.PlantHydraulicsParameters — Type

PlantHydraulicsParameters{FT <: AbstractFloat}

A struct for holding parameters of the PlantHydraulics Model.

area_index

: RAI, SAI, LAI in [m2 m-2]

K_sat

: water conductivity in the different plant compartments (m/s) when pressure is zero, a maximum

vg_α

: van Genuchten parameter

vg_n

: van Genuchten parameter

vg_m

: van Genuchten parameter

ν

: porosity

S_s

: storativity

earth_param_set

: Physical Constants and other Clima-wide parameters

Plan Hydraulics Methods and Types

ClimaLSM.PlantHydraulics.PrescribedSoilPressure — Type

PrescribedSoilPressure{FT} <: AbstractRootExtraction{FT}

A concrete type used for dispatch when computing the flux_out_roots, in the case where the soil pressure at each root layer is prescribed.

Missing docstring.

Missing docstring for ClimaLSM.PlantHydraulics.PrescribedRootFlux. Check Documenter's build log for details.

ClimaLSM.PlantHydraulics.PrescribedTranspiration — Type

PrescribedTranspiration{FT} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is prescribed.

ClimaLSM.PlantHydraulics.AbstractRootExtraction — Type

AbstractRootExtraction{FT <: AbstractFloat}

An abstract type for types representing different models of water exchange between soil and plants. Currently, prescribed soil pressure and prescribed flux models are supported for standalone plant hydraulics. Use within an LSM requires types defined within ClimaLSM, and include a prognostic soil pressure for models with both soil and roots.