Vegetation

Models

ClimaLSM.PlantHydraulics.AbstractVegetationModel — Type

AbstractVegetationModel{FT} <: AbstractModel{FT}

An abstract type for vegetation models. Concrete types include a plant hydraulics model, but future types will include multi-layer canopy models and possibly a big leaf model.

ClimaLSM.PlantHydraulics.PlantHydraulicsModel — Type

PlantHydraulicsModel{FT, PS, D, RE, T, B} <: AbstractVegetationModel{FT}

Defines, and constructs instances of, the PlantHydraulicsModel type, which is used for simulation flux of water to/from soil, along roots of different depths, along a stem, to a leaf, and ultimately being lost from the system by transpiration. This model can be used in standalone mode by prescribing the transpiration rate and soil matric potential at the root tips or flux in the roots, or with a dynamic soil model using ClimaLSM.

parameters: Parameters required by the Plant Hydraulics model

domain: The plant hydraulics model domain, of type AbstractVegetationDomain

root_extraction: The root extraction model, of type AbstractRootExtraction

transpiration: The transpiration model, of type AbstractTranspiration

Plan Hydraulics Diagnostic Variables

ClimaLSM.PlantHydraulics.effective_saturation — Function

effective_saturation(
    ν::FT, 
    ϑ_l::FT) where {FT}

Computes the effective saturation (volumetric water content relative to porosity).

ClimaLSM.PlantHydraulics.augmented_liquid_fraction — Function

augmented_liquid_fraction(
    ν::FT, 
    S_l::FT) where {FT}

Computes the augmented liquid fraction from porosity and effective saturation. Augmented liquid fraction allows for oversaturation: an expansion of the volume of space available for storage in a plant compartment. It is analogous to the augmented liquid fraction state variable in the soil model.

ClimaLSM.PlantHydraulics.water_retention_curve — Function

water_retention_curve(
    vg_α::FT,
    vg_n::FT,
    vg_m::FT,
    S_l::FT,
    ν::FT,
    S_s::FT) where {FT}

Converts augmented liquid fraction (ϑl) to effective saturation (Sl), and then effective saturation to absolute pressure, represented by the height (ψ) of the water column that would give rise to this pressure. Pressure for both the unsaturated and saturated regimes are calculated. Units are in meters.

ClimaLSM.PlantHydraulics.inverse_water_retention_curve — Function

inverse_water_retention_curve(
    vg_α::FT,
    vg_n::FT,
    vg_m::FT,
    ψ::FT,
    ν::FT,
    S_s::FT) where {FT}

Converts absolute pressure, represented by the height (ψ) of the water column that would give rise to this pressure, to effective saturation (S_l).

ClimaLSM.PlantHydraulics.flux_out_roots — Function

flux_out_roots(
    re::PrescribedSoilPressure{FT},
    model::PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

A method which computes the flux between the soil and the stem, via the roots, in the case of a standalone plant hydraulics model with prescribed soil pressure (in m) at the root tips. This assumes that the stem compartment is the first element of Y.vegetation.ϑ_l.

PlantHydraulics.flux_out_roots(
    re::PrognosticSoilPressure{FT},
    model::PlantHydraulics.PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::ClimaCore.Fields.FieldVector,
    t::FT,
)::FT where {FT}

An extension of the PlantHydraulics.flux_out_roots function, which returns the net flux of water between the roots and the soil, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water between roots and soil at each soil layer.

PlantHydraulics.flux_out_roots(
    re::PrognosticRootFlux{FT},
    t::FT,
)::FT where {FT}

An extension of the PlantHydraulics.flux_out_roots function, which returns the net flux of water between the roots and the soil, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water between roots and soil at each soil layer.

ClimaLSM.PlantHydraulics.flux — Function

flux(
    z1,
    z2,
    ψ1,
    ψ2,
    vg_α,
    vg_n,
    vg_m,
    ν,
    S_s,
    K_sat1,
    K_sat2,

) where {FT} Computes the water flux given 1) the absolute pressure (represented by corresponding water column height) at two points located at z1 and z2, corresponding here to the middle of each compartment (for eg. a stem or leaf compartment), and 2) the maximum conductivity along the flow path between these two points, assuming an arithmetic mean for mean Ksat between the two points (Bonan, 2019; Zhu, 2008) to take into account the change in Ksat halfway between z1 and z2. The expression is given in full in the Clima docs in section 6.4 "Bulk plant hydraulics model".

Plan Hydraulics Parameters

ClimaLSM.PlantHydraulics.PlantHydraulicsParameters — Type

PlantHydraulicsParameters{FT <: AbstractFloat}

A struct for holding parameters of the PlantHydraulics Model.

area_index: RAI, SAI, LAI in [m2 m-2]

K_sat: water conductivity in the above-ground plant compartments (m/s) when pressure is zero, a maximum

vg_α: van Genuchten parameter (1/m)

vg_n: van Genuchten parameter (unitless)

vg_m: van Genuchten parameter (unitless)

ν: porosity (m3/m3)

S_s: storativity (m3/m3)

root_distribution: Root distribution function P(z)

earth_param_set: Physical Constants and other Clima-wide parameters

Plan Hydraulics Methods and Types

ClimaLSM.PlantHydraulics.PrescribedSoilPressure — Type

PrescribedSoilPressure{FT} <: AbstractRootExtraction{FT}

A concrete type used for dispatch when computing the flux_out_roots, in the case where the soil matric potential at each root layer is prescribed.

ClimaLSM.PlantHydraulics.PrescribedTranspiration — Type

PrescribedTranspiration{FT} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is prescribed.

ClimaLSM.PlantHydraulics.AbstractRootExtraction — Type

AbstractRootExtraction{FT <: AbstractFloat}

An abstract type for types representing different models of water exchange between soil and plants. Currently, prescribed soil matric potential and prescribed flux models are supported for standalone plant hydraulics. Use within an LSM requires types defined within ClimaLSM, and include a prognostic soil pressure for models with both soil and roots.