Autotrophic Respiration

Parameters

ClimaLSM.Canopy.AutotrophicRespirationParameters — Type

AutotrophicRespirationParameters{FT<:AbstractFloat}

The required parameters for the autrophic respiration model.

ne: Vcmax25 to N factor (mol CO2 m-2 s-1 kg C (kg C)-1)

ηsl: Live stem wood coefficient (kg C m-3)

σl: Specific leaf density (kg C m-2 [leaf])

μr: Ratio root nitrogen to top leaf nitrogen (-), typical value 1.0

μs: Ratio stem nitrogen to top leaf nitrogen (-), typical value 0.1

f1: Factor to convert from mol CO2 to kg C

f2: Factor of relative contribution or Rgrowth (-)

Methods

ClimaLSM.Canopy.nitrogen_content — Function

nitrogen_content(
                 ne::FT, # Mean leaf nitrogen concentration (kg N (kg C)-1)
                 Vcmax25::FT, #
                 LAI::FT, # Leaf area index
                 RAI::FT,
                 ηsl::FT, # live stem  wood coefficient (kg C m-3) 
                 h::FT, # canopy height (m)
                 σl::FT # Specific leaf density (kg C m-2 [leaf])
                 μr::FT, # Ratio root nitrogen to top leaf nitrogen (-), typical value 1.0
                 μs::FT, # Ratio stem nitrogen to top leaf nitrogen (-), typical value 0.1 
                ) where {FT}

Computes the nitrogen content of leafs (Nl), roots (Nr) and stems (Ns) as a function of leaf area index (LAI), specific leaf density (σl), the carbon content of roots (Rc), the carbon content of stems (Rs), and mean leaf nitrogen concentration (nm).

ClimaLSM.Canopy.plant_respiration_maintenance — Function

plant_respiration_maintenance(
    Rd::FT, # Dark respiration
    β::FT, # Soil moisture factor
    Nl::FT, # Nitrogen content of leafs
    Nr::FT, # Nitrogen content of roots
    Ns::FT, # Nitrogen content of stems
    f::FT # Factor to convert from mol CO2 to kg C
    ) where {FT}

Computes plant maintenance respiration as a function of dark respiration (Rd), the nitrogen content of leafs (Nl), roots (Nr) and stems (Ns), and the soil moisture factor (β).

ClimaLSM.Canopy.plant_respiration_growth — Function

plant_respiration_growth(
    f::FT, # Factor of relative contribution
    GPP::FT, # Gross primary productivity
    Rpm::FT # Plant maintenance respiration
    ) where {FT}

Computes plant growth respiration as a function of gross primary productivity (GPP), plant maintenance respiration (Rpm), and a relative contribution factor, f.