Canopy

Canopy Model Structs

ClimaLand.Canopy.CanopyModel — Type

 CanopyModel{FT, AR, RM, PM, SM, PHM, EM, SM, A, R, S, PS, D} <: ClimaLand.AbstractImExModel{FT}

The model struct for the canopy, which contains

the canopy model domain (a point for site-level simulations, or

an extended surface (plane/spherical surface) for regional or global simulations.

subcomponent model type for radiative transfer. This is of type

AbstractRadiationModel.

subcomponent model type for photosynthesis. This is of type

AbstractPhotosynthesisModel, and currently only the FarquharModel is supported.

subcomponent model type for stomatal conductance. This is of type

AbstractStomatalConductanceModel and currently only the MedlynModel is supported

subcomponent model type for plant hydraulics. This is of type

AbstractPlantHydraulicsModel and currently only a version which prognostically solves Richards equation in the plant is available.

subcomponent model type for canopy energy. This is of type

AbstractCanopyEnergyModel and currently we support a version where the canopy temperature is prescribed, and one where it is solved for prognostically.

subcomponent model type for canopy SIF. prognostically.

canopy model parameters, which include parameters that are shared

between canopy model components or those needed to compute boundary fluxes.

The atmospheric conditions, which are either prescribed

(of type PrescribedAtmosphere) or computed via a coupled simulation (of type CoupledAtmosphere).

The radiative flux conditions, which are either prescribed

(of type PrescribedRadiativeFluxes) or computed via a coupled simulation (of type CoupledRadiativeFluxes).

The soil conditions, which are either prescribed (of type PrecribedSoil, for

running the canopy model in standalone mode), or prognostic (of type PrognosticSoil, for running integrated soil+canopy models)

Note that the canopy height is specified as part of the PlantHydraulicsModel, along with the area indices of the leaves, roots, and stems. Eventually, when plant biomass becomes a prognostic variable (by integrating with a carbon model), some parameters specified here will be treated differently.

autotrophic_respiration: Autotrophic respiration model, a canopy component model

radiative_transfer: Radiative transfer model, a canopy component model

photosynthesis: Photosynthesis model, a canopy component model

conductance: Stomatal conductance model, a canopy component model

hydraulics: Plant hydraulics model, a canopy component model

energy: Energy balance model, a canopy component model

sif: SIF model, a canopy component model

atmos: Atmospheric forcing: prescribed or coupled

radiation: Radiative forcing: prescribed or coupled

soil_driver: Soil pressure: prescribed or prognostic

parameters: Shared canopy parameters between component models

domain: Canopy model domain

ClimaLand.Canopy.SharedCanopyParameters — Type

SharedCanopyParameters{FT <: AbstractFloat, PSE}

A place to store shared parameters that are required by multiple canopy components.

z_0m: Roughness length for momentum (m)

z_0b: Roughness length for scalars (m)

earth_param_set: Earth param set

Canopy Model Fluxes

ClimaLand.Canopy.PlantHydraulics.DiagnosticTranspiration — Type

DiagnosticTranspiration{FT} <: AbstractTranspiration{FT}

A concrete type used for dispatch in the case where transpiration is computed diagnostically, as a function of prognostic variables and parameters, and stored in p during the update_aux! step.

Canopy Model Soil Drivers

ClimaLand.Canopy.AbstractSoilDriver — Type

An abstract type of soil drivers of the canopy model.

ClimaLand.Canopy.PrescribedSoil — Type

 PrescribedSoil <: AbstractSoilDriver

A container for holding prescribed soil parameters needed by the canopy model when running the canopy in standalone mode, including the soil pressure, surface temperature, and albedo.

root_depths: The depth of the root tips, in meters

ψ: Prescribed soil potential (m) in the root zone a function of time

T: Prescribed soil surface temperature (K) as a function of time

α_PAR: Soil albedo for PAR

α_NIR: Soil albedo for NIR

ϵ: Soil emissivity