Canopy

Canopy Model Structs

ClimaLSM.Canopy.CanopyModel — Type

 CanopyModel{FT, RM, PM, SM, PHM, A, R, PS, D} <: AbstractModel{FT}

The model struct for the canopy, which contains

the canopy model domain (a point for site-level simulations, or

an extended surface (plane/spherical surface) for regional or global simulations.

subcomponent model type for radiative transfer. This is of type

AbstractRadiationModel and currently only the BeerLambertModel is supported.

subcomponent model type for photosynthesis. This is of type

AbstractPhotosynthesisModel, and currently only the FarquharModel is supported.

subcomponent model type for stomatal conductance. This is of type

AbstractStomatalConductanceModel and currently only the MedlynModel is supported

subcomponent model type for plant hydraulics. This is of type

AbstractPlantHydraulicsModel and currently only a version which prognostically solves Richards equation in the plant is available.

canopy model parameters, which include parameters that are shared

between canopy model components or those needed to compute boundary fluxes.

The atmospheric conditions, which are either prescribed

(of type PrescribedAtmosphere) or computed via a coupled simulation (of type CoupledAtmosphere).

The radiative flux conditions, which are either prescribed

(of type PrescribedRadiativeFluxes) or computed via a coupled simulation (of type CoupledRadiativeFluxes).

radiative_transfer: Radiative transfer model, a canopy component model

photosynthesis: Photosynthesis model, a canopy component model

conductance: Stomatal conductance model, a canopy component model

hydraulics: Plant hydraulics model, a canopy component model

atmos: Atmospheric forcing: prescribed or coupled

radiation: Radiative forcing: prescribed or coupled

parameters: Shared canopy parameters between component models

domain: Canopy model domain

ClimaLSM.Canopy.SharedCanopyParameters — Type

SharedCanopyParameters{FT <: AbstractFloat, PSE}

A place to store shared parameters that are required by all canopy components.

LAI: Leaf Area Index (m2/m2)

h_c: Canopy height (m)

z_0m: Roughness length for momentum (m)

z_0b: Roughness length for scalars (m)

earth_param_set: Earth param set

Canopy Model Fluxes

ClimaLSM.Canopy.PlantHydraulics.DiagnosticTranspiration — Type

DiagnosticTranspiration{FT} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is computed diagnostically, as a function of prognostic variables and parameters, and stored in p during the update_aux! step.

ClimaLSM.Canopy.canopy_surface_fluxes — Function

canopy_surface_fluxes(atmos::PrescribedAtmosphere{FT},
                      model::CanopyModel,
                      Y,
                      p,
                      t::FT) where {FT}

Computes canopy transpiration using Monin-Obukhov Surface Theory, the prescribed atmospheric conditions, and the canopy conductance.

Please note that in the future the SurfaceFluxes.jl code will compute fluxes taking into account the canopy conductance, so that what is returned by surface_fluxes is correct. At present, it does not, so we are adjusting for it after the fact here in both ET and LHF.