Photosynthesis

Models and Parameters

ClimaLand.Canopy.PModel — Type

PModel{FT,
            OPFT <: PModelParameters{FT},
            OPCT <: PModelConstants{FT}
            } <: AbstractPhotosynthesisModel{FT}

An implementation of the optimality photosynthesis model "P-model v1.0" of Stocker et al. (2020).

Stocker, B. D., Wang, H., Smith, N. G., Harrison, S. P., Keenan, T. F., Sandoval, D., Davis, T.,
    and Prentice, I. C.: P-model v1.0: an optimality-based light use efficiency model for simulating
    ecosystem gross primary production, Geosci. Model Dev., 13, 1545–1581,
    https://doi.org/10.5194/gmd-13-1545-2020, 2020.

ClimaLand.Canopy.PModelParameters — Type

PModelParameters{FT<:AbstractFloat}

The required parameters for P-model (Stocker et al. 2020). Parameters are typically tunable with considerable uncertainty.

cstar: Constant describing cost of maintaining electron transport (unitless) Typical value = 0.41

β: Ratio of unit costs of transpiration and carboxylation (unitless) Typical value = 146

ϕc: Scaling parameter for temp-dependent intrinsic quantum yield (unitless) Typical value = 0.087

ϕ0: Temp-independent intrinsic quantum yield. If provided, overrides ϕc. (unitless) Typical value = 0.05

ϕa0: Constant term in temp-dependent intrinsic quantum yield (unitless).

ϕa1: First order term in temp-dependent intrinsic quantum yield (K^-1).

ϕa2: Second order term in temp-dependent intrinsic quantum yield (K^-2).

α: Timescale parameter used in EMA for acclimation of optimal photosynthetic capacities (unitless). Setting this to 0 represents no incorporation of past values. Since we update the EMA equation once per day, α = 1 - 1 day/τ where τ is the acclimation timescale in days.

sc: Sensitivity to low water pressure, in the moisture stress factor, (Pa^{-1}) [Tuzet et al. (2003)]

pc: Reference water pressure for the moisture stress factor (Pa) [Tuzet et al. (2003)]

ClimaLand.Canopy.PModelConstants — Type

PModelConstants{FT<:AbstractFloat}

The required constants for P-model (Stocker et al. 2020). These are physical or biochemical constants that are not expected to change with time or space.

R: Gas constant (J mol^-1 K^-1)

Kc25: Michaelis-Menten parameter for carboxylation at 25°C (μmol mol^-1)

Ko25: Michaelis-Menten parameter for oxygenation at 25°C (μmol mol^-1)

To: Reference temperature equal to 25˚C (K)

ΔHkc: Energy of activation for Kc (J mol^-1)

ΔHko: Energy of activation for Ko (J mol^-1)

Drel: Relative diffusivity of CO2 in the stomatal pores, equal to 1.6.

ΔHΓstar: Effective energy of activation for Γstar (J mol^-1)

Γstar25: Γstar at 25 °C (Pa)

Ha_Vcmax: Effective energy of activation for Vcmax (J mol^-1)

Hd_Vcmax: Effective energy of deactivation for Vcmax (J mol^-1)

aS_Vcmax: Intercept term for dS in Vcmax deactivation factor (J K^-1 mol^-1)

bS_Vcmax: Slope term for dS in Vcmax deactivation factor (J K^-2 mol^-1)

Ha_Jmax: Effective energy of activation for Jmax (J mol^-1)

Hd_Jmax: Effective energy of deactivation for Jmax (J mol^-1)

aS_Jmax: Intercept term for dS in Jmax deactivation factor (J K^-1 mol^-1)

bS_Jmax: Slope term for dS in Jmax deactivation factor (J K^-2 mol^-1)

Mc: Molar mass of carbon (kg mol^-1)

oi: Intercellular O2 mixing ratio (unitless)

aRd: First order coefficient for temp-dependent Rd (K^-1)

bRd: Second order coefficient for temp-dependent Rd (K^-2)

fC3: Constant factor appearing the dark respiration term for C3 plants (unitless)

planck_h: Planck constant (J s)

lightspeed: Speed of light (m s^-1)

N_a: Avogadro constant (mol^-1)

ρ_water: Density of water (kg m^-3)

ClimaLand.Canopy.FarquharModel — Type

FarquharModel{FT}(
    domain;
    photosynthesis_parameters = clm_photosynthesis_parameters(
        domain.space.surface,
    ),
    sc::FT = LP.get_default_parameter(FT, :low_water_pressure_sensitivity),
    pc::FT = LP.get_default_parameter(FT, :moisture_stress_ref_water_pressure),
) where {
    FT <: AbstractFloat,
    MECH <: Union{FT, ClimaCore.Fields.Field},
    VC <: Union{FT, ClimaCore.Fields.Field},
}

Creates a FarquharModel using default parameters of type FT.

The photosynthesis_parameters argument is a NamedTuple that contains

is_c3: a Float or Field indicating if plants are C3 (1) or C4 (0) (unitless)

Vcmax25: a Float or Field representing the maximum carboxylation rate at 25C (mol m^-2 s^-1)

By default, these parameters are set by the clm_photosynthesis_parameters function, which reads in CLM data onto the surface space as ClimaUtilities SpaceVaryingInputs.

The following additional default parameters are used:

sc = 5e-6 (Pa^{-1}) - sensitivity to low water pressure in the moisture stress factor [Tuzet et al. (2003)]

pc = -2e6 (Pa) - reference water pressure for the moisture stress factor [Tuzet et al. (2003)]

ClimaLand.Canopy.FarquharParameters — Type

FarquharParameters{
    FT<:AbstractFloat,
    MECH <: Union{FT, ClimaCore.Fields.Field},
    VC <: Union{FT, ClimaCore.Fields.Field},
}

The required parameters for the Farquhar photosynthesis model.

Vcmax25: Vcmax at 25 °C (mol CO2/m^2/s)

Γstar25: Γstar at 25 °C (mol/mol)

Kc25: Michaelis-Menten parameter for CO2 at 25 °C (mol/mol)

Ko25: Michaelis-Menten parameter for O2 at 25 °C (mol/mol)

ΔHkc: Energy of activation for CO2 (J/mol)

ΔHko: Energy of activation for oxygen (J/mol)

ΔHVcmax: Energy of activation for Vcmax (J/mol)

ΔHΓstar: Energy of activation for Γstar (J/mol)

ΔHJmax: Energy of activation for Jmax (J/mol)

ΔHRd: Energy of activation for Rd (J/mol)

To: Reference temperature equal to 25 degrees Celsius (K)

oi: Intercelluar O2 concentration (mol/mol); taken to be constant

ϕ: Quantum yield of photosystem II (Bernacchi, 2003; unitless)

θj: Curvature parameter, a fitting constant to compute J, unitless

fC3: Constant factor appearing the dark respiration term for C3, equal to 0.015.

fC4: Constant factor appearing the dark respiration term for C4, equal to 0.025.

sc: Sensitivity to low water pressure, in the moisture stress factor, (Pa^{-1}) [Tuzet et al. (2003)]

pc: Reference water pressure for the moisture stress factor (Pa) [Tuzet et al. (2003)]

Q10: Q10 temperature parameter for Vcmax and Rd for C4 photosynthesis; unitless

s1: Parameter appearing in temperature dependence of C4 Vcmax; K^{-1}

s2: Parameter appearing in temperature dependence of C4 Vcmax; K

s3: Parameter appearing in temperature dependence of C4 Vcmax; K^{-1}

s4: Parameter appearing in temperature dependence of C4 Vcmax; K

s5: Parameter appearing in temperature dependence of C4 Rd; K^{-1}

s6: Parameter appearing in temperature dependence of C4 Rd; K

E: Quantum yield for C4 photosynthesis; mol/mol

is_c3: Photosynthesis mechanism: 1.0 indicates C3, 0.0 indicates C4

ClimaLand.Canopy.OptimalityFarquharModel — Type

OptimalityFarquharModel{FT,
                        OPFT <: OptimalityFarquharParameters{FT}
                        } <: AbstractPhotosynthesisModel{FT}

Optimality model of Smith et al. (2019) for estimating Vcmax, based on the assumption that Aj = Ac. Smith et al. (2019). Global photosynthetic capacity is optimized to the environment. Ecology Letters, 22(3), 506–517. https://doi.org/10.1111/ele.13210

ClimaLand.Canopy.OptimalityFarquharParameters — Type

OptimalityFarquharParameters{FT<:AbstractFloat}

The required parameters for the optimality Farquhar photosynthesis model. Currently, only C3 photosynthesis is supported.

is_c3: Photosynthesis mechanism: C3 only

Γstar25: Γstar at 25 °C (mol/mol)

Kc25: Michaelis-Menten parameter for CO2 at 25 °C (mol/mol)

Ko25: Michaelis-Menten parameter for O2 at 25 °C (mol/mol)

ΔHkc: Energy of activation for CO2 (J/mol)

ΔHko: Energy of activation for oxygen (J/mol)

ΔHVcmax: Energy of activation for Vcmax (J/mol)

ΔHΓstar: Energy of activation for Γstar (J/mol)

ΔHJmax: Energy of activation for Jmax (J/mol)

ΔHRd: Energy of activation for Rd (J/mol)

To: Reference temperature equal to 25 degrees Celsius (K)

oi: Intercellular O2 concentration (mol/mol); taken to be constant

ϕ: Quantum yield of photosystem II (Bernacchi, 2003; unitless)

θj: Curvature parameter, a fitting constant to compute J, unitless

fC3: Constant factor appearing the dark respiration term for C3 plants, equal to 0.015.

sc: Fitting constant to compute the moisture stress factor (Pa^{-1})

pc: Fitting constant to compute the moisture stress factor (Pa)

c: Constant describing cost of maintaining electron transport (unitless)

FarquharModel Methods

ClimaLand.Canopy.arrhenius_function — Function

arrhenius_function(T::FT, To::FT, R::FT, ΔH::FT)

Computes the Arrhenius function at temperature T given the reference temperature To=298.15K, the universal gas constant R, and the energy activation ΔH.

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.intercellular_co2 — Function

intercellular_co2(ca::FT, Γstar::FT, medlyn_factor::FT) where{FT}

Computes the intercellular CO2 concentration (mol/mol) given the atmospheric concentration (ca, mol/mol), the CO2 compensation (Γstar, mol/mol), and the Medlyn factor (unitless).

ClimaLand.Canopy.co2_compensation — Function

co2_compensation(Γstar25::FT,
                 ΔHΓstar::FT,
                 T::FT,
                 To::FT,
                 R::FT) where {FT}

Computes the CO2 compensation point (Γstar), in units of mol/mol, as a function of its value at 25 °C (Γstar25), a constant energy of activation (ΔHΓstar), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.rubisco_assimilation — Function

rubisco_assimilation(is_c3::AbstractFloat, args...)

Calls the correct rubisco assimilation function based on the is_c3.

A is_c3 value of 1.0 corresponds to C3 photosynthesis and calls c3_rubisco_assimilation, while 0.0 corresponds to C4 photsynthesis and calls c4_rubisco_assimilation.

ClimaLand.Canopy.light_assimilation — Function

light_assimilation(is_c3::AbstractFloat, args...)

Calls the correct light assimilation function based on the is_c3.

A is_c3 value of 1.0 corresponds to C3 photosynthesis and calls c3_light_assimilation, while 0.0 corresponds to C4 photsynthesis and calls c4_light_assimilation.

ClimaLand.Canopy.max_electron_transport — Function

max_electron_transport(Vcmax::FT) where {FT}

Computes the maximum potential rate of electron transport (Jmax), in units of mol/m^2/s, as a function of Vcmax at 25 °C (Vcmax25), a constant (ΔHJmax), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.electron_transport — Function

electron_transport(APAR::FT,
                   Jmax::FT,
                   θj::FT,
                   ϕ::FT) where {FT}

Computes the rate of electron transport (J), in units of mol/m^2/s, as a function of the maximum potential rate of electron transport (Jmax), absorbed photosynthetically active radiation (APAR), an empirical "curvature parameter" (θj; Bonan Eqn 11.21) and the quantum yield of photosystem II (ϕ).

See Ch 11, G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.net_photosynthesis — Function

net_photosynthesis(Ac::FT,
                   Aj::FT,
                   Rd::FT,
                   β::FT) where {FT}

Computes the total net carbon assimilation (An), in units of mol CO2/m^2/s, as a function of the Rubisco limiting factor (Ac), the electron transport limiting rate (Aj), dark respiration (Rd), and the moisture stress factor (β).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.optimality_max_photosynthetic_rates — Function

optimalitymaxphotosynthetic_rates(APAR::FT, θj::FT, ϕ::FT, oi::FT, ci::FT, Γstar::FT, Kc::FT, Ko::FT)

Computes the photosynthesis rates Vcmax and Jmax in mol/m^2/s given absorbed photosynthetically active radiation (APAR), an empirical "curvature parameter" (θj; Bonan Eqn 11.21) the quantum yield of photosystem II (ϕ), the intercellular o2 content (oi), the intercellular CO2 concentration (ci), Γstar, and Kc and Ko.

See Smith et al. 2019.

ClimaLand.Canopy.moisture_stress — Function

moisture_stress(pl::FT,
                sc::FT,
                pc::FT) where {FT}

Computes the moisture stress factor (β), which is unitless, as a function of a constant (sc, 1/Pa), a reference pressure (pc, Pa), and the leaf water pressure (pl, Pa) .

See Eqn 12.57 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.dark_respiration — Function

dark_respiration(is_c3::AbstractFloat, args...)

Calls the correct dark respiration function based on is_c3.

A is_c3 value of 1.0 corresponds to C3 photosynthesis and calls c3_dark_respiration, while 0.0 corresponds to C4 photsynthesis and calls c4_dark_respiration.

ClimaLand.Canopy.compute_GPP — Function

compute_GPP(An::FT,
         K::FT,
         LAI::FT,
         Ω::FT) where {FT}

Computes the total canopy photosynthesis (GPP) as a function of the total net carbon assimilation (An), the extinction coefficient (K), leaf area index (LAI) and the clumping index (Ω).

ClimaLand.Canopy.MM_Kc — Function

MM_Kc(Kc25::FT,
      ΔHkc::FT,
      T::FT,
      To::FT,
      R::FT) where {FT}

Computes the Michaelis-Menten coefficient for CO2 (Kc), in units of mol/mol or Pa (depending on the units of Kc25), as a function of its value at 25 °C (Kc25), a constant (ΔHkc), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.MM_Ko — Function

MM_Ko(Ko25::FT,
      ΔHko::FT,
      T::FT,
      To::FT,
      R::FT) where {FT}

Computes the Michaelis-Menten coefficient for O2 (Ko), in units of mol/mol or Pa (depending on the units of Ko25), as a function of its value at 25 °C (Ko25), a constant (ΔHko), a standard temperature (To), the universal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLand.Canopy.compute_Vcmax — Function

compute_Vcmax(is_c3::AbstractFloat, args...)

Calls the correct Vcmax function based on is_c3.

A is_c3 value of 1.0 corresponds to C3 photosynthesis and calls c3_compute_Vcmax, while 0.0 corresponds to C4 photsynthesis and calls c4_compute_Vcmax.

PModel Methods

ClimaLand.Canopy.computefullpmodeloutputs ClimaLand.Canopy.sethistoricalcache! ClimaLand.Canopy.updateoptimalEMA ClimaLand.Canopy.makePModelcallback ClimaLand.Canopy.intrinsicquantumyield ClimaLand.Canopy.computeviscosityratio ClimaLand.Canopy.computeKmm ClimaLand.Canopy.optimalco2ratioc3 ClimaLand.Canopy.intercellularco2pmodel ClimaLand.Canopy.gsco2pmodel ClimaLand.Canopy.gsh2opmodel ClimaLand.Canopy.vcmaxpmodel ClimaLand.Canopy.computeLUE ClimaLand.Canopy.computemjwithjmaxlimitation ClimaLand.Canopy.electrontransportpmodel ClimaLand.Canopy.co2compensationp ClimaLand.Canopy.quadraticsoilmoisturestress ClimaLand.Canopy.compute_APAR