Photosynthesis

Parameters

ClimaLSM.Canopy.FarquharParameters — Type

FarquharParameters{FT<:AbstractFloat}

The required parameters for the Farquhar photosynthesis model.

mechanism: Photosynthesis mechanism: C3 or C4

Vcmax25: Vcmax at 25 °C (mol CO2/m^2/s)

Γstar25: Γstar at 25 °C (mol/mol)

Kc25: Michaelis-Menten parameter for CO2 at 25 °C (mol/mol)

Ko25: Michaelis-Menten parameter for O2 at 25 °C (mol/mol)

ΔHkc: Energy of activation for CO2 (J/mol)

ΔHko: Energy of activation for oxygen (J/mol)

ΔHVcmax: Energy of activation for Vcmax (J/mol)

ΔHΓstar: Energy of activation for Γstar (J/mol)

ΔHJmax: Energy of activation for Jmax (J/mol)

ΔHRd: Energy of activation for Rd (J/mol)

To: Reference temperature equal to 25 degrees Celsius (K)

oi: Intercelluar O2 concentration (mol/mol); taken to be constant

ϕ: Quantum yield of photosystem II (Bernacchi, 2003; unitless)

θj: Curvature parameter, a fitting constant to compute J, unitless

f: Constant factor appearing the dark respiration term, equal to 0.015.

sc: Fitting constant to compute the moisture stress factor (Pa^{-1})

ψc: Fitting constant to compute the moisture stress factor (Pa)

Methods

ClimaLSM.Canopy.arrhenius_function — Function

arrhenius_function(T::FT, To::FT, R::FT, ΔH::FT)

Computes the Arrhenius function at temperature T given the reference temperature To=298.15K, the universal gas constant R, and the energy activation ΔH.

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.intercellular_co2 — Function

intercellular_co2(ca::FT, Γstar::FT, medlyn_factor::FT) where{FT}

Computes the intercellular CO2 concentration (mol/mol) given the atmospheric concentration (ca, mol/mol), the CO2 compensation (Γstar, mol/mol), and the Medlyn factor (unitless).

ClimaLSM.Canopy.co2_compensation — Function

co2_compensation(Γstar25::FT,
                 ΔHΓstar::FT,
                 T::FT,
                 To::FT,
                 R::FT) where {FT}

Computes the CO2 compensation point (Γstar), in units of mol/mol, as a function of its value at 25 °C (Γstar25), a constant energy of activation (ΔHΓstar), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.rubisco_assimilation — Function

rubisco_assimilation(::C3,
                     Vcmax::FT,
                     ci::FT,
                     Γstar::FT,
                     Kc::FT,
                     Ko::FT,
                     oi::FT) where {FT}

Computes the Rubisco limiting rate of photosynthesis for C3 plants (Ac), in units of moles CO2/m^2/s, as a function of the maximum rate of carboxylation of Rubisco (Vcmax), the leaf internal carbon dioxide partial pressure (ci), the CO2 compensation point (Γstar), and Michaelis-Menten parameters for CO2 and O2, respectively, (Kc) and (Ko).

The empirical parameter oi is equal to 0.209 (mol/mol). See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

rubisco_assimilation(::C4, Vcmax::FT,_...) where {FT}

Computes the Rubisco limiting rate of photosynthesis for C4 plants (Ac) in units of moles CO2/m^2/s, as equal to the maximum rate of carboxylation of Rubisco (Vcmax).

ClimaLSM.Canopy.light_assimilation — Function

light_assimilation(::C3,
                   J::FT,
                   ci::FT,
                   Γstar::FT) where {FT}

Computes the electron transport limiting rate (Aj), in units of moles CO2/m^2/s, for C3 plants as a function of the rate of electron transport (J), the leaf internal carbon dioxide partial pressure (ci), and the CO2 compensation point (Γstar).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

light_assimilation(::C4, J::FT, _...) where {FT}

Computes the electron transport limiting rate (Aj), in units of moles CO2/m^2/s, for C4 plants, as equal to the rate of electron transport (J).

ClimaLSM.Canopy.C3 — Type

C3 <: AbstractPhotosynthesisMechanism

Helper struct for dispatching between C3 and C4 photosynthesis.

ClimaLSM.Canopy.C4 — Type

C4 <: AbstractPhotosynthesisMechanism

Helper struct for dispatching between C3 and C4 photosynthesis.

ClimaLSM.Canopy.max_electron_transport — Function

max_electron_transport(Vcmax::FT) where {FT}

Computes the maximum potential rate of electron transport (Jmax), in units of mol/m^2/s, as a function of Vcmax at 25 °C (Vcmax25), a constant (ΔHJmax), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.electron_transport — Function

electron_transport(APAR::FT,
                   Jmax::FT,
                   θj::FT,
                   ϕ::FT) where {FT}

Computes the rate of electron transport (J), in units of mol/m^2/s, as a function of the maximum potential rate of electron transport (Jmax), absorbed photosynthetically active radiation (APAR), an empirical "curvature parameter" (θj; Bonan Eqn 11.21) and the quantum yield of photosystem II (ϕ).

See Ch 11, G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.net_photosynthesis — Function

net_photosynthesis(Ac::FT,
                   Aj::FT,
                   Rd::FT,
                   β::FT) where {FT}

Computes the total net carbon assimilation (An), in units of mol CO2/m^2/s, as a function of the Rubisco limiting factor (Ac), the electron transport limiting rate (Aj), dark respiration (Rd), and the moisture stress factor (β).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.moisture_stress — Function

moisture_stress(ψl::FT,
                sc::FT,
                ψc::FT) where {FT}

Computes the moisture stress factor (β), which is unitless, as a function of a constant (sc), a constant (ψc), and the leaf water potential (ψl).

See Eqn 12.57 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.dark_respiration — Function

dark_respiration(Vcmax25::FT,
                 β::FT,
                 f::FT,
                 ΔHkc::FT,
                 T::FT,
                 To::FT,
                 R::FT) where {FT}

Computes dark respiration (Rd), in units of mol CO2/m^2/s, as a function of the maximum rate of carboxylation of Rubisco (Vcmax25), and the moisture stress factor (β), an empirical factor f is equal to 0.015, a constant (ΔHRd), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.compute_GPP — Function

compute_GPP(An::FT,
         K::FT,
         LAI::FT,
         Ω::FT) where {FT}

Computes the total canopy photosynthesis (GPP) as a function of the total net carbon assimilation (An), the extinction coefficient (K), leaf area index (LAI) and the clumping index (Ω).

ClimaLSM.Canopy.MM_Kc — Function

MM_Kc(Kc25::FT,
      ΔHkc::FT,
      T::FT,
      To::FT,
      R::FT) where {FT}

Computes the Michaelis-Menten coefficient for CO2 (Kc), in units of mol/mol, as a function of its value at 25 °C (Kc25), a constant (ΔHkc), a standard temperature (To), the unversal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.MM_Ko — Function

MM_Ko(Ko25::FT,
      ΔHko::FT,
      T::FT,
      To::FT,
      R::FT) where {FT}

Computes the Michaelis-Menten coefficient for O2 (Ko), in units of mol/mol, as a function of its value at 25 °C (Ko25), a constant (ΔHko), a standard temperature (To), the universal gas constant (R), and the temperature (T).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).

ClimaLSM.Canopy.compute_Vcmax — Function

compute_Vcmax(Vcmax25::FT,
       T::FT,
       To::FT,
       R::FT,
       ep5::FT) where {FT}

Computes the maximum rate of carboxylation of Rubisco (Vcmax), in units of mol/m^2/s, as a function of temperature (T), Vcmax at the reference temperature 25 °C (Vcmax25), the universal gas constant (R), and the reference temperature (To).

See Table 11.5 of G. Bonan's textbook, Climate Change and Terrestrial Ecosystem Modeling (2019).