PlantHydraulics

Models

ClimaLand.Canopy.PlantHydraulics.PlantHydraulicsModel — Type

PlantHydraulicsModel{FT, PS, T, AA} <: AbstractPlantHydraulicsModel{FT}

Defines, and constructs instances of, the PlantHydraulicsModel type, which is used for simulation flux of water to/from soil, along roots of different depths, along a stem, to a leaf, and ultimately being lost from the system by transpiration. Note that the canopy height is specified as part of the PlantHydraulicsModel, along with the area indices of the leaves, roots, and stems.

This model can also be combined with the soil model using ClimaLand, in which case the prognostic soil water content is used to determine root extraction, and the transpiration is also computed diagnostically. In global run with patches of bare soil, you can "turn off" the canopy model (to get zero root extraction, zero absorption and emission, zero transpiration and sensible heat flux from the canopy), by setting:

n_leaf = 1

n_stem = 0

LAI = SAI = RAI = 0.

A plant model can have leaves but no stem, but not vice versa. If n_stem = 0, SAI must be zero.

Finally, the model can be used in Canopy standalone mode by prescribing the soil matric potential at the root tips or flux in the roots. There is also the option (intendend only for debugging) to use a prescribed transpiration rate.

n_stem: The number of stem compartments for the plant; can be zero

n_leaf: The number of leaf compartments for the plant; must be >=1

compartment_midpoints: The height of the center of each leaf compartment/stem compartment, in meters

compartment_surfaces: The height of the compartments' top faces, in meters. The canopy height is the last element of the vector.

compartment_labels: The label (:stem or :leaf) of each compartment

parameters: Parameters required by the Plant Hydraulics model

transpiration: The transpiration model, of type AbstractTranspiration

Plant Hydraulics Diagnostic Variables

ClimaLand.Canopy.PlantHydraulics.effective_saturation — Function

effective_saturation(
    ν::FT,
    ϑ_l::FT) where {FT}

Computes the effective saturation given the augmented liquid fraction.

ClimaLand.Canopy.PlantHydraulics.augmented_liquid_fraction — Function

augmented_liquid_fraction(
    ν::FT,
    S_l::FT) where {FT}

Computes the augmented liquid fraction from porosity and effective saturation.

Augmented liquid fraction allows for oversaturation: an expansion of the volume of space available for storage in a plant compartment.

ClimaLand.Canopy.PlantHydraulics.water_retention_curve — Function

water_retention_curve(
    S_l::FT,
    b::FT,
    ν::FT,
    S_s::FT) where {FT}

Returns the potential ψ given the effective saturation S at a point, according to a linear model for the retention curve with parameters specified by retention_params.

ClimaLand.Canopy.PlantHydraulics.inverse_water_retention_curve — Function

inverse_water_retention_curve(
    ψ::FT,
    b::FT,
    ν::FT,
    S_s::FT) where {FT}

Returns the effective saturation given the potential at a point, according to the linear retention curve model.

ClimaLand.Canopy.PlantHydraulics.root_water_flux_per_ground_area! — Function

PlantHydraulics.root_water_flux_per_ground_area!(
    fa::ClimaCore.Fields.Field,
    s::PrognosticSoilConditions,
    model::Canopy.PlantHydraulics.PlantHydraulicsModel,
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
    t,
)

An extension of the PlantHydraulics.root_water_flux_per_ground_area! function, which returns the net flux of water between the roots and the soil, per unit ground area, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water per ground area between roots and soil at each soil layer.

root_water_flux_per_ground_area!(
    fa::ClimaCore.Fields.Field,
    ground::PrescribedGroundConditions,
    model::PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
    t,
) where {FT}

A method which computes the water flux between the soil and the stem, via the roots, and multiplied by the RAI, in the case of a model running without a prognostic soil model:

Flux = -Keff x [(ψstem - ψsoil)/(zstem - zsoil) + 1], where Keff = Ksoil Kstem /(Kstem + Ksoil)

Note that in PrescribedSoil mode, we compute the flux using Ksoil = Kplant(ψsoil) and Kstem = Kplant(ψstem). In PrognosticSoil mode, we compute the flux using Ksoil = Ksoil(ψsoil) and Kstem = Kplant(ψstem). The latter is a better model, but our PrescribedSoil struct does not store Ksoil, only ψsoil.

The returned flux is per unit ground area. This assumes that the stem compartment is the first element of Y.canopy.hydraulics.ϑ_l.

ClimaLand.Canopy.PlantHydraulics.water_flux — Function

water_flux(
    z1,
    z2,
    ψ1,
    ψ2,
    K1,
    K2,
) where {FT}

Computes the water flux given the absolute potential ψ (pressure/(ρg)) and the conductivity K (m/s) at the center of the two layers with midpoints z1 and z2.

We currently assuming a harmonic mean for effective conducticity between the two layers (see CLM Technical Documentation).

To account for different path lengths in the two compartments Δz1 and Δz2, we would require the following conductance k (1/s) keff = K1/Δz1*K2/Δz2/(K1/Δz1+K2/Δz2) and a water flux of F = -keff * (ψ1 +z1 - ψ2 - z2) (m/s).

This currently assumes the path lengths are equal.

ClimaLand.Canopy.PlantHydraulics.hydraulic_conductivity — Function

hydraulic_conductivity(conductivity_params::Weibull{FT}, ψ::FT) where {FT}

Computes the hydraulic conductivity at a point, using the Weibull formulation, given the potential ψ.

Plant Hydraulics Parameters

ClimaLand.Canopy.PlantHydraulics.PlantHydraulicsParameters — Type

PlantHydraulicsParameters

A struct for holding parameters of the PlantHydraulics Model.

ai_parameterization: The area index model for LAI, SAI, RAI

ν: porosity (m3/m3)

S_s: storativity (m3/m3)

conductivity_model: Conductivity model and parameters

retention_model: Water retention model and parameters

rooting_depth: Rooting depth parameter (m) - a characteristic depth below which 1/e of the root mass lies

ClimaLand.Canopy.PlantHydraulics.Weibull — Type

Weibull{FT} <: AbstractConductivityModel{FT}

A concrete type specifying that a Weibull conductivity model is to be used; the struct contains the require parameters for this model.

Fields

K_sat: Maximum Water conductivity in the above-ground plant compartments (m/s) at saturation

ψ63: The absolute water potential in xylem (or xylem water potential) at which ∼63% of maximum xylem conductance is lost (Liu, 2020).

c: Weibull parameter c, which controls shape the shape of the conductance curve (Sperry, 2016).

ClimaLand.Canopy.PlantHydraulics.LinearRetentionCurve — Type

LinearRetentionCurve{FT} <: AbstractRetentionModel{FT}

A concrete type specifying that a linear water retention model is to be used; the struct contains the require parameters for this model.

When ψ = 0, the effective saturation is one, so the intercept is not a free parameter, and only the slope must be specified.

Fields

a: Bulk modulus of elasticity and slope of potential to volume curve. See also Corcuera, 2002, and Christoffersen, 2016.

Plant Hydraulics Methods and Types

Missing docstring.

Missing docstring for ClimaLand.PlantHydraulics.PrescribedSoilPressure. Check Documenter's build log for details.

ClimaLand.Canopy.PlantHydraulics.PrescribedTranspiration — Type

PrescribedTranspiration{FT, F <: Function} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is prescribed.

Missing docstring.

Missing docstring for ClimaLand.PlantHydraulics.AbstractRootExtraction. Check Documenter's build log for details.