PlantHydraulics

Models

ClimaLSM.Canopy.PlantHydraulics.PlantHydraulicsModel — Type

PlantHydraulicsModel{FT, PS, T, AA} <: AbstractPlantHydraulicsModel{FT}

Defines, and constructs instances of, the PlantHydraulicsModel type, which is used for simulation flux of water to/from soil, along roots of different depths, along a stem, to a leaf, and ultimately being lost from the system by transpiration. Note that the canopy height is specified as part of the PlantHydraulicsModel, along with the area indices of the leaves, roots, and stems.

This model can also be combined with the soil model using ClimaLSM, in which case the prognostic soil water content is used to determine root extraction, and the transpiration is also computed diagnostically. In global run with patches of bare soil, you can "turn off" the canopy model (to get zero root extraction, zero absorption and emission, zero transpiration and sensible heat flux from the canopy), by setting:

n_leaf = 1

n_stem = 0

LAI = SAI = RAI = 0.

A plant model can have leaves but no stem, but not vice versa. If n_stem = 0, SAI must be zero.

Finally, the model can be used in Canopy standalone mode by prescribing the soil matric potential at the root tips or flux in the roots. There is also the option (intendend only for debugging) to use a prescribed transpiration rate.

n_stem: The number of stem compartments for the plant; can be zero

n_leaf: The number of leaf compartments for the plant; must be >=1

compartment_midpoints: The height of the center of each leaf compartment/stem compartment, in meters

compartment_surfaces: The height of the compartments' top faces, in meters. The canopy height is the last element of the vector.

compartment_labels: The label (:stem or :leaf) of each compartment

parameters: Parameters required by the Plant Hydraulics model

transpiration: The transpiration model, of type AbstractTranspiration

Plant Hydraulics Diagnostic Variables

ClimaLSM.Canopy.PlantHydraulics.effective_saturation — Function

effective_saturation(
    ν::FT,
    ϑ_l::FT) where {FT}

Computes the effective saturation given the augmented liquid fraction.

ClimaLSM.Canopy.PlantHydraulics.augmented_liquid_fraction — Function

augmented_liquid_fraction(
    ν::FT,
    S_l::FT) where {FT}

Computes the augmented liquid fraction from porosity and effective saturation.

Augmented liquid fraction allows for oversaturation: an expansion of the volume of space available for storage in a plant compartment.

ClimaLSM.Canopy.PlantHydraulics.water_retention_curve — Function

water_retention_curve(
    S_l::FT,
    b::FT,
    ν::FT,
    S_s::FT) where {FT}

Returns the potential ψ given the effective saturation S at a point, according to a linear model for the retention curve with parameters specified by retention_params.

ClimaLSM.Canopy.PlantHydraulics.inverse_water_retention_curve — Function

inverse_water_retention_curve(
    ψ::FT,
    b::FT,
    ν::FT,
    S_s::FT) where {FT}

Returns the effective saturation given the potential at a point, according to the linear retention curve model.

ClimaLSM.Canopy.PlantHydraulics.root_water_flux_per_ground_area! — Function

root_water_flux_per_ground_area!(
    fa::ClimaCore.Fields.Field,
    s::PrescribedSoil,
    model::PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
    t,
) where {FT}

A method which computes the water flux between the soil and the stem, via the roots, and multiplied by the RAI, in the case of a model running without an integrated soil model.

The returned flux is per unit ground area. This assumes that the stem compartment is the first element of Y.canopy.hydraulics.ϑ_l.

PlantHydraulics.root_water_flux_per_ground_area!(
    fa::ClimaCore.Fields.Field,
    s::PrognosticSoil,
    model::Canopy.PlantHydraulics.PlantHydraulicsModel{FT},
    Y::ClimaCore.Fields.FieldVector,
    p::NamedTuple,
    t,
)

An extension of the PlantHydraulics.root_water_flux_per_ground_area! function, which returns the net flux of water between the roots and the soil, per unit ground area, when both soil and plant hydraulics are modeled prognostically. This is for use in an LSM.

It is computed by summing the flux of water per ground area between roots and soil at each soil layer.

ClimaLSM.Canopy.PlantHydraulics.flux — Function

flux(
    z1,
    z2,
    ψ1,
    ψ2,
    K1,
    K2,
) where {FT}

Computes the water flux given the absolute potential (pressure/(ρg)) at the center of the two compartments z1 and z2, and the conductivity along the flow path between these two points.

We currently assuming an arithmetic mean for mean Ksat between the two points (Bonan, 2019; Zhu, 2008) to take into account the change in Ksat halfway between z1 and z2; this is incorrect for compartments of differing sizes.

ClimaLSM.Canopy.PlantHydraulics.hydraulic_conductivity — Function

hydraulic_conductivity(conductivity_params::Weibull{FT}, ψ::FT) where {FT}

Computes the hydraulic conductivity at a point, using the Weibull formulation, given the potential ψ.

Plant Hydraulics Parameters

ClimaLSM.Canopy.PlantHydraulics.PlantHydraulicsParameters — Type

PlantHydraulicsParameters

A struct for holding parameters of the PlantHydraulics Model.

ai_parameterization: The area index model for LAI, SAI, RAI

ν: porosity (m3/m3)

S_s: storativity (m3/m3)

conductivity_model: Conductivity model and parameters

retention_model: Water retention model and parameters

root_distribution: Root distribution function P(z)

ClimaLSM.Canopy.PlantHydraulics.Weibull — Type

Weibull{FT} <: AbstractConductivityModel{FT}

A concrete type specifying that a Weibull conductivity model is to be used; the struct contains the require parameters for this model.

Fields

K_sat: Maximum Water conductivity in the above-ground plant compartments (m/s) at saturation

ψ63: The absolute water potential in xylem (or xylem water potential) at which ∼63% of maximum xylem conductance is lost (Liu, 2020).

c: Weibull parameter c, which controls shape the shape of the conductance curve (Sperry, 2016).

ClimaLSM.Canopy.PlantHydraulics.LinearRetentionCurve — Type

LinearRetentionCurve{FT} <: AbstractRetentionModel{FT}

A concrete type specifying that a linear water retention model is to be used; the struct contains the require parameters for this model.

When ψ = 0, the effective saturation is one, so the intercept is not a free parameter, and only the slope must be specified.

Fields

a: Bulk modulus of elasticity and slope of potential to volume curve. See also Corcuera, 2002, and Christoffersen, 2016.

Plant Hydraulics Methods and Types

Missing docstring.

Missing docstring for ClimaLSM.PlantHydraulics.PrescribedSoilPressure. Check Documenter's build log for details.

ClimaLSM.Canopy.PlantHydraulics.PrescribedTranspiration — Type

PrescribedTranspiration{FT, F <: Function} <: AbstractTranspiration{FT}

A concrete type used for dispatch when computing the transpiration from the leaves, in the case where transpiration is prescribed.

Missing docstring.

Missing docstring for ClimaLSM.PlantHydraulics.AbstractRootExtraction. Check Documenter's build log for details.