Radiative Transfer

Parameters

ClimaLand.Canopy.BeerLambertParameters — Type

BeerLambertParameters{FT <: AbstractFloat}

The required parameters for the Beer-Lambert radiative transfer model.

α_PAR_leaf: PAR leaf reflectance (unitless)

α_NIR_leaf: NIR leaf reflectance

ϵ_canopy: Emissivity of the canopy

Ω: Clumping index following Braghiere (2021) (unitless)

λ_γ_PAR: Typical wavelength per PAR photon (m)

λ_γ_NIR: Typical wavelength per NIR photon (m)

G_Function: Leaf angle distribution function

Methods

Missing docstring.

Missing docstring for ClimaLand.Canopy.compute_absorbances. Check Documenter's build log for details.

ClimaLand.Canopy.plant_absorbed_pfd — Function

plant_absorbed_pfd(
    RT::BeerLambertModel{FT},
    SW_IN:FT,
    α_leaf::FT,
    LAI::FT,
    K::FT,
    α_soil::FT
)

Computes the absorbed, reflected, and transmitted photon flux density in terms of mol photons per m^2 per second for a radiation band.

This applies the Beer-Lambert law, which is a function of incident radiation (SW_IN; moles of photons/m^2/), leaf reflectance (α_leaf), the extinction coefficient (K), leaf area index (LAI), and the albedo of the soil (α_soil).

Returns a tuple of reflected, absorbed, and transmitted radiation in mol photons/m^2/s.

plant_absorbed_pfd(
    RT::TwoStreamModel{FT},
    α_leaf,
    SW_IN::FT,
    LAI::FT,
    K::FT,
    τ_leaf,
    θs::FT,
    α_soil::FT,
)

Computes the absorbed, transmitted, and reflected photon flux density in terms of mol photons per m^2 per second for a radiation band.

This applies the two-stream radiative transfer solution which takes into account the impacts of scattering within the canopy. The function takes in all parameters from the parameter struct of a TwoStreamModel, along with the incident radiation, LAI, extinction coefficient K, soil albedo from the canopy soil_driver, solar zenith angle, and τ.

Returns a tuple of reflected, absorbed, and transmitted radiation in mol photons/m^2/s.

ClimaLand.Canopy.extinction_coeff — Function

extinction_coeff(ld::FT,
                 θs::FT) where {FT}

Computes the vegetation extinction coefficient (K), as a function of the sun zenith angle (θs), and the leaf angle distribution (ld).

Missing docstring.

Missing docstring for ClimaLand.Canopy.extinction_coeff. Check Documenter's build log for details.

ClimaLand.Canopy.canopy_radiant_energy_fluxes! — Function

Canopy.canopy_radiant_energy_fluxes!(p::NamedTuple,
                                     s::PrognosticSoil{F},
                                     canopy,
                                     radiation::PrescribedRadiativeFluxes,
                                     earth_param_set::PSE,
                                     Y::ClimaCore.Fields.FieldVector,
                                     t,
                                    ) where {FT, PSE}

In standalone mode, this function computes and stores the net long and short wave radition, in W/m^2, absorbed by the canopy.

In integrated mode, we have already computed those quantities in lsm_radiant_energy_fluxes!, so this method does nothing additional.

LW and SW net radiation are stored in p.canopy.radiative_transfer.LW_n and p.canopy.radiative_transfer.SW_n.

canopy_radiant_energy_fluxes!(p::NamedTuple,
                              s::PrescribedSoil,
                              canopy,
                              radiation::PrescribedRadiativeFluxes,
                              earth_param_set::PSE,
                              Y::ClimaCore.Fields.FieldVector,
                              t,
                             ) where {PSE}

Computes and stores the net long and short wave radition, in W/m^2, absorbed by the canopy when the canopy is run in standalone mode, with a PrescribedSoil conditions.

LW and SW net radiation are stored in p.canopy.radiative_transfer.LW_n and p.canopy.radiative_transfer.SW_n.