Radiative Transfer

Parameters

ClimaLSM.Canopy.BeerLambertParameters — Type

BeerLambertParameters{FT <: AbstractFloat}

The required parameters for the Beer-Lambert radiative transfer model.

ld: Leaf angle distribution function (unitless)

ρ_leaf: PAR canopy reflectance (unitless)

Ω: Clumping index following Braghiere (2021) (unitless)

λ_γ: Typical wavelength per PAR photon (m)

Methods

ClimaLSM.Canopy.plant_absorbed_ppfd — Function

plant_absorbed_ppfd(PAR::FT,
                   ρ_leaf::FT,
                   K::FT,
                   LAI::FT,
                   Ω::FT) where {FT}

Computes the absorbed photosynthetically active radiation in terms of mol photons per m^2 per second (APAR).

This assumes the Beer-Lambert law, which is a function of photosynthetically active radiation (PAR; moles of photons/m^2/), PAR canopy reflectance (ρ_leaf), the extinction coefficient (K), leaf area index (LAI) and the clumping index (Ω).

ClimaLSM.Canopy.extinction_coeff — Function

extinction_coeff(ld::FT,
                 θs::FT) where {FT}

Computes the vegetation extinction coefficient (K), as a function of the sun zenith angle (θs), and the leaf angle distribution (ld).