Stomatal Conductance

Models and Parameters

ClimaLand.Canopy.MedlynConductanceModel — Type

MedlynConductanceModel{FT}(
    domain,
    toml_dict::CP.ParamDict;
    g0::FT = toml_dict["min_stomatal_conductance"],
    g1 = clm_medlyn_g1(domain.space.surface),
) where {FT <: AbstractFloat}

Creates a MedlynConductanceModel using default parameters of type FT.

The conductance_parameters argument is a NamedTuple that contains

g1: a Float or ClimaCore Field representing the slope parameter (PA^{1/2})

By default, this parameter is set by the clm_medlyn_g1 function, which reads in CLM data onto the surface space as a ClimaUtilities SpaceVaryingInput.

The following default parameter is used:

g0 = FT(1e-4) (mol m^-2 s^-1) - minimum stomatal conductance

ClimaLand.Canopy.MedlynConductanceModel — Method

MedlynConductanceModel{FT}(
    domain,
    toml_dict::CP.ParamDict;
    g0::FT = toml_dict["min_stomatal_conductance"],
    g1 = clm_medlyn_g1(domain.space.surface),
) where {FT <: AbstractFloat}

Creates a MedlynConductanceModel using default parameters of type FT.

The conductance_parameters argument is a NamedTuple that contains

g1: a Float or ClimaCore Field representing the slope parameter (PA^{1/2})

By default, this parameter is set by the clm_medlyn_g1 function, which reads in CLM data onto the surface space as a ClimaUtilities SpaceVaryingInput.

The following default parameter is used:

g0 = FT(1e-4) (mol m^-2 s^-1) - minimum stomatal conductance

ClimaLand.Canopy.MedlynConductanceParameters — Type

MedlynConductanceParameters{FT <: AbstractFloat}

The required parameters for the Medlyn stomatal conductance model.

Drel: Relative diffusivity of water vapor (unitless)

g0: Minimum stomatal conductance mol/m^2/s

g1: Slope parameter, inversely proportional to the square root of marginal water use efficiency (Pa^{1/2})

ClimaLand.Canopy.MedlynConductanceParameters — Method

function MedlynConductanceParameters(
    toml_dict::CP.ParamDict;
    g1,
    g0 = toml_dict["min_stomatal_conductance"],
)

TOML dict based constructor supplying default values for the MedlynConductanceParameters struct.

ClimaLand.Canopy.PModelConductance — Type

PModelConductance{FT}(toml_dict;
                      Drel = toml_dict["Drel"],
                    ) where {FT <: AbstractFloat}

Creates a PModelConductance using default parameters of type FT.

The following default parameter is used:

Drel = FT(1.6) (unitless) - relative diffusivity of H2O to CO2 (Bonan Table A.3)

ClimaLand.Canopy.PModelConductance — Method

PModelConductance{FT}(toml_dict;
                      Drel = toml_dict["Drel"],
                    ) where {FT <: AbstractFloat}

Creates a PModelConductance using default parameters of type FT.

The following default parameter is used:

Drel = FT(1.6) (unitless) - relative diffusivity of H2O to CO2 (Bonan Table A.3)

ClimaLand.Canopy.PModelConductanceParameters — Type

PModelConductanceParameters{FT <: AbstractFloat}

The required parameters for the P-Model stomatal conductance model.

Drel: Relative diffusivity of water vapor (unitless)

Methods

ClimaLand.Canopy.medlyn_term — Function

medlyn_term(g1::FT, T_air::FT, P_air::FT, q_air::FT, thermo_params) where {FT}

Computes the Medlyn term, equal to 1+g1/sqrt(VPD), by first computing the VPD, where VPD is the vapor pressure deficit in the atmosphere (Pa), and g_1 is a constant with units of sqrt(Pa).

thermo_params is the Thermodynamics.jl parameter set.

Note that in supersaturated conditions the vapor pressure deficit will be negative, which leads to an imaginary Medlyn term m. Clipping to zero solves this, but this leads to division by zero, so we regularize the division by adding a small quantity.

An alternative to consider in the future is to compute the inverse of this quantity and stomatal resistance instead of conductance.

ClimaLand.Canopy.medlyn_conductance — Function

medlyn_conductance(g0::FT,
                   Drel::FT,
                   medlyn_term::FT,
                   An::FT,
                   ca::FT) where {FT}

Computes the stomatal conductance according to Medlyn, as a function of the minimum stomatal conductance (g0), the relative diffusivity of water vapor with respect to CO2 (Drel), the Medlyn term (unitless), the biochemical demand for CO2 (An), and the atmospheric concentration of CO2 (ca).

This returns the conductance in units of mol/m^2/s. It must be converted to m/s using the molar density of water prior to use in SurfaceFluxes.jl.

ClimaLand.Canopy.penman_monteith — Function

penman_monteith(
    Δ::FT, # Rate of change of saturation vapor pressure with air temperature. (Pa K−1)
    Rn::FT, # Net irradiance (W m−2)
    G::FT, # Ground heat flux (W m−2)
    ρa::FT, # Dry air density (kg m−3)
    cp::FT, # Specific heat capacity of air (J kg−1 K−1)
    VPD::FT, # vapor pressure deficit (Pa)
    ga::FT, # atmospheric conductance (m s−1)
    γ::FT, # Psychrometric constant (γ ≈ 66 Pa K−1)
    gs::FT, # surface or stomatal conductance (m s−1)
    Lv::FT, # Volumetric latent heat of vaporization (J m-3)
    ) where {FT}

Computes the evapotranspiration in m/s using the Penman-Monteith equation.