Stomatal Conductance

Parameters

ClimaLSM.Canopy.MedlynConductanceParameters — Type

MedlynConductanceParameters{FT <: AbstractFloat}

The required parameters for the Medlyn stomatal conductance model.

Drel: Relative diffusivity of water vapor (unitless)

g0: Minimum stomatal conductance mol/m^2/s

g1: Slope parameter, inversely proportional to the square root of marginal water use efficiency (Pa^{1/2})

Methods

ClimaLSM.Canopy.medlyn_term — Function

medlyn_term(g1::FT, T_air::FT, P_air::FT, q_air::FT, thermo_params) where {FT}

Computes the Medlyn term, equal to 1+g1/sqrt(VPD), by first computing the VPD, where VPD is the vapor pressure deficit in the atmosphere (Pa), and g_1 is a constant with units of sqrt(Pa).

thermo_params is the Thermodynamics.jl parameter set.

ClimaLSM.Canopy.medlyn_conductance — Function

medlyn_conductance(g0::FT,
                   Drel::FT,
                   medlyn_term::FT,
                   An::FT,
                   ca::FT) where {FT}

Computes the stomatal conductance according to Medlyn, as a function of the minimum stomatal conductance (g0), the relative diffusivity of water vapor with respect to CO2 (Drel), the Medlyn term (unitless), the biochemical demand for CO2 (An), and the atmospheric concentration of CO2 (ca).

This returns the conductance in units of mol/m^2/s. It must be converted to m/s using the molar density of water prior to use in SurfaceFluxes.jl.

ClimaLSM.Canopy.upscale_leaf_conductance — Function

upscale_leaf_conductance(gs::FT, LAI::FT, T::FT, R::FT, P::FT) where {FT}

This currently takes a leaf conductance (moles per leaf area per second) and (1) converts it to m/s, (2) upscales to the entire canopy, by assuming the leaves in the canopy are in parallel and hence multiplying by LAI.

TODO: Check what CLM does, and check if we can use the same function for GPP from An, and make more general.

ClimaLSM.Canopy.penman_monteith — Function

penman_monteith(
    Δ::FT, # Rate of change of saturation vapor pressure with air temperature. (Pa K−1)  
    Rn::FT, # Net irradiance (W m−2)
    G::FT, # Ground heat flux (W m−2)
    ρa::FT, # Dry air density (kg m−3)
    cp::FT, # Specific heat capacity of air (J kg−1 K−1) 
    VPD::FT, # vapor pressure deficit (Pa)
    ga::FT, # atmospheric conductance (m s−1)
    γ::FT, # Psychrometric constant (γ ≈ 66 Pa K−1)
    gs::FT, # surface or stomatal conductance (m s−1)
    Lv::FT, # Volumetric latent heat of vaporization (J m-3)
    ) where {FT}

Computes the evapotranspiration in m/s using the Penman-Monteith equation.